Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0073766780  0.0166667671  0.0296739784 -0.0374249878 -0.0071655224 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.1503466551  0.0341448221 -0.0198348773  0.0883548545  0.0317313026 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0291518078  0.0656022226 -0.0564117935  0.0610069136  0.0995026245 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0586877582  0.0326508367 -0.0733183446 -0.0048545631 -0.1268619283 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0179102101  0.0893035892  0.1113587652  0.0790647874 -0.0005090520 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0864868288  0.0139389684 -0.0017653791 -0.0193566329  0.0404885780 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.1706106132  0.0604089527 -0.0420817423  0.0276759700  0.0569918107 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0234826044  0.0068101410  0.0623357630 -0.0531853343 -0.0213300780 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0375744020 -0.0232059676 -0.0022190987  0.1944862913 -0.0183211429 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0292591571 -0.0496416407 -0.0786326236 -0.0495475629  0.0878522435 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.1263340776 -0.0101391845 -0.1527377608  0.0055867736 -0.0564736296 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0148914661  0.0097999143 -0.0340250080  0.0195805852 -0.0428698847 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0857365431 -0.1007648683  0.0056564155 -0.0089547164  0.0679009873 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0803455844  0.0073082666  0.1393312691 -0.0215686148 -0.0008356671 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0155315696 -0.1095893776 -0.0752970118  0.0134276901 -0.1010621011 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0668065185  0.0799266462 -0.0984827788 -0.0085102826 -0.1472894058 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0352035710  0.0846008656  0.1543816611 -0.0087102590  0.0292418892 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.1394293667  0.0390259308  0.0527810896  0.0023321443  0.0058228392 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0219600820 -0.0367983893  0.0035620838  0.0840498533 -0.0192664846 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0051418719  0.0559292807  0.0422949660 -0.0527925729  0.0673458660 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -0.0116305018  0.0423328333 -0.0449679387  0.0758308930 -0.0142968245 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0215255981 -0.0062204559 -0.0206496493 -0.0689904391  0.0556993370 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0557497945  0.0180995496 -0.0109819666  0.0207034706 -0.0276635345 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0119865590  0.0176589855  0.0136747119  0.1124342594 -0.1420792027 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.1380792098 -0.0810706858  0.1309519173  0.1298282483  0.0176610160 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0804219880 -0.1212129456 -0.0898622597  0.0245820409 -0.0431405634 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0912245890 -0.0221695820 -0.0313208564 -0.0417551104  0.0505544606 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0089690977  0.0969458922  0.0488541043 -0.0150067428 -0.1454785347 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0046152524 -0.1674637999 -0.0391932750  0.0708653273  0.1398828584 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0118269562  0.0405001820 -0.1052448725 -0.1895476488  0.0494589390 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0086882091  0.0267110496 -0.0715909852 -0.1675625923  0.0508162167 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0576444778 -0.0239937984  0.0116121767  0.0422988753 -0.0061519055 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0725948230 -0.0770936283  0.0025260979 -0.0906022436 -0.0485686612 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0863238138 -0.0779388801  0.0060676735  0.0650546837 -0.0045722233 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -0.0028506837  0.1005072390 -0.0833230514 -0.0162124508  0.0202722496 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0080644701  0.0136313881 -0.0775020517  0.0986138430  0.0706645132 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0120442539  0.0565705438  0.1342916286 -0.0436812278 -0.0734126309 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0173361590  0.0335665228 -0.0736530172  0.0952248147  0.0336055984 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0795988064  0.0901849424 -0.0069263647  0.0029922438 -0.0634264293 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0502994231 -0.0712737000  0.0746671846  0.0534098660  0.0129830910 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0419576242  0.0696285993 -0.0657879350 -0.0298398316  0.0345756794 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0510099922 -0.1458462572 -0.1202448511 -0.0688205548  0.0351263143 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0343561270  0.0067847324 -0.1191733726 -0.0382342805  0.0111703350 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0163529401  0.0693623658 -0.0481839577 -0.0214752568  0.0984439103 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0377753211  0.0980202455  0.0894835016  0.0547390868 -0.1329630162 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0087504340  0.0836841708 -0.0857970752 -0.0001741230 -0.1278307871 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0182838545  0.0540215004 -0.0773547774  0.1372455586 -0.0154776254 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.0610430508  0.0376460993 -0.0315001670 -0.0051713155  0.0125908851 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.1132090112 -0.0803755882  0.0293419613  0.0173130583 -0.0716320969 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0266328397  0.0819414815 -0.0176921657  0.0510067166 -0.0186062155 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0457357316 -0.0597351629 -0.0174256313  0.0176566473  0.0520014576 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.1124673116  0.1029316379 -0.0797576996  0.0185874988  0.0566024858 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.1063007670  0.0390968720 -0.1370413626  0.0545332978 -0.1095834247 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0302754625  0.0569057661  0.0869361632  0.0574556052  0.0391794389 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0634844724 -0.0393391514  0.0717165063 -0.0360582253  0.0509979584 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.0473274454 -0.0712472526 -0.0040970936 -0.0338514205 -0.0439325611 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0193948810  0.0801728373  0.0697565141 -0.0772250740 -0.0675037722 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0744731014  0.1589140851 -0.1345443203 -0.0748614702 -0.0701616594 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0711383695 -0.1459453411 -0.0302629111 -0.0831048606 -0.0486067325 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.0708623855 -0.0325429011 -0.0331677390 -0.0517115756 -0.1088067345 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0053787165 -0.0134145410  0.0874593146 -0.0823955670  0.0488075617 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.1246738770  0.0883529105 -0.0218712462 -0.0846027882 -0.0312786724 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0149335287  0.0949743029  0.0248181925 -0.0450106060  0.1032739046 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0037609145  0.0001263643 -0.0036506378 -0.1075544153 -0.0070349989 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0001949194 -0.0843528050  0.0028262559  0.0378738955  0.0338145886 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0663083195  0.0228996677 -0.0079569486  0.0333818005  0.0208575236 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0215420013 -0.0444508680 -0.0017205279 -0.0287048070  0.0762418017 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0388114920 -0.1057116733 -0.0883703503  0.0918674771  0.0085828033 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.1250158012 -0.0504863362 -0.0190919868 -0.1441251715 -0.0734800165 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0347386009  0.0087287756  0.0548972140  0.0806594450 -0.0223154600 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  0.0240669091  0.0049562010  0.0710819499 -0.0758498371 -0.1252104540 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.0517200928  0.0025189573 -0.0289451906  0.0779598317  0.0006173708 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0881719199 -0.0108060211 -0.0500880548  0.1104668565  0.0108735992 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0492820051  0.0131859479  0.0780113489  0.0421519705  0.0113081200 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0362702888 -0.0407811533  0.0338695015 -0.0298492861 -0.0318657575 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0355499151 -0.0640317330 -0.1119802402 -0.0323902724  0.0513550021 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.0610688628 -0.0881471096  0.0891582498 -0.0177160801  0.0521438868 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0245273845 -0.0595958587  0.0409375128 -0.1474626982 -0.0438458901 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0258014469  0.0219579550  0.0312724057 -0.0747996832 -0.0589227376 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0184868079 -0.0227890026 -0.1780635540 -0.0697729265 -0.0834509065 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0774619110 -0.0251559074  0.0659042668  0.0381950910  0.1413421834 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0621221569 -0.0465517413  0.1374912289 -0.0596631372 -0.0007590721 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0094829068 -0.0114621810 -0.0294466617 -0.0027200573  0.0207156898 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.1314900831  0.1195604226 -0.0805697005  0.0114795679  0.0898693925 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.1262315364 -0.0368149667  0.0803536887  0.0064672688 -0.0429626535 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.1057614696 -0.0055065093 -0.0630886184  0.0297866542  0.0065997275 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0213307866  0.1710363023 -0.0203455093  0.0277584995 -0.1444634928 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0166180849  0.0190016445 -0.0427200374  0.0314573098  0.0386079497 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0568352482 -0.0185323468 -0.0832429849  0.0407496538 -0.0235568094 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0169197555  0.0724701398  0.0089399477