Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  5.293152e-02 -1.007140e-01 -7.373763e-02  1.925496e-01  2.780755e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  3.594844e-03  5.640898e-02  8.563741e-02  6.370642e-02  6.736392e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  6.425884e-02  4.936128e-02  2.158951e-03 -2.730390e-02  3.913168e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -3.152090e-02  7.517521e-02  8.521708e-03 -1.699415e-02 -2.555214e-03 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  6.812658e-02  3.876940e-02  9.022675e-02  3.099927e-02  1.200724e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  5.500008e-02  5.510612e-02  3.018345e-02  2.919129e-02  7.430580e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  8.088925e-02  5.057701e-02 -5.354178e-03 -9.781491e-03 -6.686791e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  7.634706e-02  4.302579e-03  5.416116e-02 -1.133766e-02  1.069459e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -8.159892e-02 -4.090775e-02  1.177821e-01  1.580313e-01  4.288311e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  1.354758e-01 -9.682795e-02 -1.525149e-02 -1.676248e-02 -6.217715e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  1.016550e-01 -3.244863e-03 -1.123301e-02 -6.680694e-02  7.963448e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -3.226898e-02  8.082864e-02 -8.514978e-03 -1.154622e-01 -9.691280e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  1.145644e-02  3.804481e-02 -7.400140e-02  1.538730e-02 -8.494303e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -7.367774e-02 -1.108480e-01  4.499885e-02  1.313583e-01  1.136248e-01 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  6.010973e-03 -1.939851e-01 -2.366136e-02 -4.326966e-02 -6.717955e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.111148e-01 -7.694586e-02  7.514996e-02  4.314423e-02 -1.170248e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  2.520586e-01  2.371378e-02 -4.256736e-02  1.457239e-02  3.331486e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -5.041710e-02  3.201426e-02 -1.003213e-01 -8.635614e-02 -5.536776e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -1.003017e-02  1.165720e-01  2.007962e-02  2.046303e-02 -2.921957e-03 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -1.705597e-02  4.950164e-03  6.867373e-02 -7.016392e-03 -1.501647e-03 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -3.803154e-02  3.084567e-02 -3.723517e-02 -3.653677e-02  1.834655e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -2.177319e-02  2.360134e-01 -2.389672e-03  1.974188e-02 -2.284919e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  9.478013e-02  1.974226e-02 -6.167147e-02 -6.224278e-02 -1.343994e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -4.509267e-03 -7.237165e-04 -3.004557e-02 -4.895003e-02  1.103543e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -4.588658e-03  1.249661e-02  1.037352e-01 -2.668701e-02 -5.246215e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  1.025284e-01 -1.774322e-01 -9.583889e-02  3.823383e-02 -6.397257e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  5.222715e-02 -7.369808e-02  1.035017e-02  5.980654e-02  1.378340e-01 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  5.448537e-02  3.572843e-02 -2.965814e-02  9.590014e-02 -9.893023e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  4.220767e-02  2.029418e-02  2.937669e-02 -9.722502e-03  7.489467e-04 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -3.129165e-02  5.188966e-02  1.620347e-02  7.704189e-02 -4.695055e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -3.869522e-02 -6.305201e-02  5.034270e-02  3.782558e-02  2.943655e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -1.244098e-01 -2.521447e-02 -6.100335e-02 -4.525333e-02 -6.824234e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -6.322779e-03 -1.462611e-01  2.842927e-02 -1.467317e-01  9.068544e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -4.101014e-02 -8.002617e-03 -2.433517e-02 -4.259500e-02 -5.462849e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  7.874800e-02 -6.190764e-03  3.575948e-02 -6.957703e-02 -3.338114e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  6.546424e-02 -1.431824e-01  7.218866e-03  6.397400e-02  4.089099e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -6.828066e-02 -3.964774e-03  8.557931e-02  6.928888e-02  1.042587e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  5.346024e-02  1.577607e-02 -1.935422e-02 -2.712643e-02 -6.722657e-04 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  3.630982e-02 -9.203606e-02 -7.571260e-02  9.211612e-02 -1.144627e-01 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -1.427971e-01 -5.700811e-02  2.047535e-02 -4.414485e-04  6.155555e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  8.708104e-02 -1.107893e-02  5.411087e-02  3.980313e-02  4.412164e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -1.147085e-01  1.301008e-01 -2.012156e-02 -5.992284e-02  4.329579e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  4.961529e-02  4.755408e-02 -3.773573e-02  2.576294e-02 -1.402098e-01 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -6.159432e-02 -1.151450e-01 -1.710344e-01  1.388902e-02  6.262978e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  3.556433e-02  6.266388e-02  5.901605e-02  1.024170e-01  7.309903e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -1.269093e-01 -3.543288e-02 -1.118173e-01 -5.749902e-02 -9.261621e-03 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -2.470726e-01  8.362188e-02 -2.527111e-02 -1.676274e-03  3.956232e-03 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -7.250334e-02  1.135315e-01 -6.833607e-02 -3.628343e-02  9.050334e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  7.419763e-03 -3.250076e-02  8.679200e-03  2.088418e-02  7.614083e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  5.690183e-02  7.215351e-02 -5.138075e-02  4.662054e-02 -1.063456e-01 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -6.587950e-02  3.415886e-02 -3.281536e-02 -2.051456e-02 -2.004258e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -3.937246e-02 -1.076527e-02 -1.681923e-02  9.331010e-02 -4.258649e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -3.961577e-02  6.731017e-02 -4.858191e-02  1.123590e-01  8.079156e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  1.210108e-01 -9.236393e-02 -1.370526e-02 -8.102699e-02 -6.149342e-03 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -1.848791e-01 -5.726963e-02  1.953408e-02  3.060694e-02  4.126223e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -8.124300e-02  1.924148e-02  8.594445e-03  1.762682e-02  3.357066e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -1.279475e-01 -3.781979e-02  1.244588e-01 -1.165421e-01 -5.062894e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -8.732728e-04  3.847452e-02 -4.326576e-02  1.909146e-01 -1.611496e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -2.490936e-02  1.224061e-01  3.883369e-02  3.250088e-02  2.050903e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  1.710755e-02 -2.606751e-02  5.824371e-02 -8.878230e-03  8.603460e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  2.998311e-02 -3.427460e-02 -4.541880e-02  1.606721e-02 -5.849547e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -1.516801e-02  6.691174e-02  5.667506e-02  3.784155e-02 -8.930499e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  4.862301e-02  2.287145e-02 -8.465471e-02 -1.570541e-02  1.459862e-01 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -3.176990e-02 -1.025134e-01  8.769045e-02 -1.012598e-01  9.833126e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  3.571841e-02  1.485467e-01  3.274487e-02  3.464388e-02 -1.048525e-01 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -8.372452e-02 -6.192054e-02 -1.284198e-01  1.337475e-01 -2.211667e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -1.800419e-02  9.323975e-03  3.195669e-02  7.967519e-02 -3.218321e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -4.514357e-02 -2.412102e-02 -6.292713e-02  8.729005e-02 -3.632121e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  2.903864e-02 -2.135675e-01 -5.439470e-02 -6.237558e-02  1.317406e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -2.840726e-02 -6.171049e-03 -1.097855e-01 -2.301475e-02  1.354965e-01 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -6.597462e-02 -7.349196e-03  8.513838e-02  3.650440e-03 -2.866025e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  3.326387e-02 -4.769037e-03 -1.153138e-01 -2.464802e-02  1.486683e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -4.566203e-02 -4.757916e-02 -1.273307e-01 -5.967044e-02  1.110520e-01 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -3.254194e-02  3.806780e-02  1.250787e-02 -2.435901e-02 -2.819813e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -3.680328e-02  1.818132e-02 -2.215856e-03  3.203736e-02  1.305220e-01 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  4.148222e-02 -9.750940e-02  7.957449e-04 -1.047389e-01 -2.158216e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  7.041521e-02 -2.095404e-01 -2.777559e-02 -2.972298e-02 -5.407054e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -8.880517e-02  4.434060e-02 -1.544844e-02 -4.749441e-03  1.393969e-01 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -2.668421e-03  9.950197e-02 -8.121611e-02 -5.142616e-02 -3.059185e-03 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -5.584137e-02 -1.438950e-01 -1.758067e-01 -4.864996e-02 -2.563742e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  1.003177e-01 -7.236503e-02  8.191548e-02  1.630761e-02  8.121500e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  3.445012e-02 -3.316618e-02 -2.681421e-02  9.742421e-02  5.008549e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -8.936964e-02  8.580754e-02  5.480244e-02  4.701383e-02  2.991988e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  1.016207e-01  5.374694e-02 -2.677011e-02 -1.083897e-01  1.943282e-03 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -3.594004e-03  1.946789e-02  1.056455e-01 -9.534541e-02  1.009354e-01 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  4.881441e-02  1.512220e-01 -3.852206e-03 -5.167826e-02 -3.041734e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -4.220870e-02  2.226667e-02  1.231343e-01 -7.441744e-02  9.374627e-03 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  1.058552e-01 -9.504403e-02  8.507766e-05  2.668197e-03  3.419453e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  3.851041e-02 -4.796416e-02  4.113091e-02 -1.396852e-02 -4.117121e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  4.758756e-02  1.216944e-01  9.177079e-02