Amino Acid Composition Descriptor

This function calculates the Amino Acid Composition descriptor (dim: 20).

extractAAC(x)

Arguments

x
A character vector, as the input protein sequence.

Value

A length 20 named vector

References

M. Bhasin, G. P. S. Raghava. Classification of Nuclear Receptors Based on Amino Acid Composition and Dipeptide Composition. Journal of Biological Chemistry, 2004, 279, 23262.

See also

Examples

x = readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractAAC(x)

#>          A          R          N          D          C          E          Q 
#> 0.06405694 0.07117438 0.03914591 0.05160142 0.06761566 0.04804270 0.04804270 
#>          G          H          I          L          K          M          F 
#> 0.08185053 0.03024911 0.03558719 0.07651246 0.03914591 0.01245552 0.03202847 
#>          P          S          T          W          Y          V 
#> 0.05338078 0.08896797 0.04448399 0.02313167 0.04270463 0.04982206