Amphiphilic Pseudo Amino Acid Composition (APseAAC) Descriptor

This function calculates the Amphiphilic Pseudo Amino Acid Composition (APseAAC, or APAAC) descriptor (dim: 20 + (n * lambda), n is the number of properties selected, default is 80).

extractAPAAC(x, props = c("Hydrophobicity", "Hydrophilicity"),
  lambda = 30, w = 0.05, customprops = NULL)

Arguments

x
A character vector, as the input protein sequence.

props
A character vector, specifying the properties used. 2 properties are used by default, as listed below:

'Hydrophobicity'

Hydrophobicity value of the 20 amino acids

'Hydrophilicity'

Hydrophilicity value of the 20 amino acids

lambda
The lambda parameter for the APAAC descriptors, default is 30.

w
The weighting factor, default is 0.05.

customprops
A n x 21 named data frame contains n customized property. Each row contains one property. The column order for different amino acid types is 'AccNo', 'A', 'R', 'N', 'D', 'C', 'E', 'Q', 'G', 'H', 'I', 'L', 'K', 'M', 'F', 'P', 'S', 'T', 'W', 'Y', 'V', and the columns should also be exactly named like this. The AccNo column contains the properties' names. Then users should explicitly specify these properties with these names in the argument props. See the examples below for a demonstration. The default value for customprops is NULL.

Value

A length 20 + n * lambda named vector, n is the number of properties selected.

Note

Note the default 20 * 2 prop values have already been independently given in the function. Users can also specify other (up to 544) properties with the Accession Number in the AAindex data, with or without the default three properties, which means users should explicitly specify the properties to use. For this descriptor type, users need to intelligently evaluate the underlying details of the descriptors provided, instead of using this function with their data blindly. It would be wise to use some negative and positive control comparisons where relevant to help guide interpretation of the results.

References

Kuo-Chen Chou. Prediction of Protein Cellular Attributes Using Pseudo-Amino Acid Composition. PROTEINS: Structure, Function, and Genetics, 2001, 43: 246-255.

Kuo-Chen Chou. Using Amphiphilic Pseudo Amino Acid Composition to Predict Enzyme Subfamily Classes. Bioinformatics, 2005, 21, 10-19.

JACS, 1962, 84: 4240-4246. (C. Tanford). (The hydrophobicity data)

PNAS, 1981, 78:3824-3828 (T.P.Hopp & K.R.Woods). (The hydrophilicity data)

See also

Examples

x = readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractAPAAC(x)

#>                 Pc1.A                 Pc1.R                 Pc1.N 
#>          3.537412e+01          3.930458e+01          2.161752e+01 
#>                 Pc1.D                 Pc1.C                 Pc1.E 
#>          2.849582e+01          3.733935e+01          2.653059e+01 
#>                 Pc1.Q                 Pc1.G                 Pc1.H 
#>          2.653059e+01          4.520027e+01          1.670445e+01 
#>                 Pc1.I                 Pc1.L                 Pc1.K 
#>          1.965229e+01          4.225242e+01          2.161752e+01 
#>                 Pc1.M                 Pc1.F                 Pc1.P 
#>          6.878302e+00          1.768706e+01          2.947844e+01 
#>                 Pc1.S                 Pc1.T                 Pc1.W 
#>          4.913073e+01          2.456536e+01          1.277399e+01 
#>                 Pc1.Y                 Pc1.V  Pc2.Hydrophobicity.1 
#>          2.358275e+01          2.751321e+01          2.196320e-04 
#>  Pc2.Hydrophilicity.1  Pc2.Hydrophobicity.2  Pc2.Hydrophilicity.2 
#>          1.025766e-03         -3.088876e-04         -1.834385e-04 
#>  Pc2.Hydrophobicity.3  Pc2.Hydrophilicity.3  Pc2.Hydrophobicity.4 
#>          1.174146e-03          7.400156e-04         -1.105715e-03 
#>  Pc2.Hydrophilicity.4  Pc2.Hydrophobicity.5  Pc2.Hydrophilicity.5 
#>         -4.493680e-04          1.766358e-03          1.471212e-03 
#>  Pc2.Hydrophobicity.6  Pc2.Hydrophilicity.6  Pc2.Hydrophobicity.7 
#>         -1.441572e-03         -4.913600e-03         -1.678053e-05 
#>  Pc2.Hydrophilicity.7  Pc2.Hydrophobicity.8  Pc2.Hydrophilicity.8 
#>          7.312356e-04         -1.885399e-03         -1.928708e-03 
#>  Pc2.Hydrophobicity.9  Pc2.Hydrophilicity.9 Pc2.Hydrophobicity.10 
#>         -2.931177e-03         -1.555660e-03          2.916597e-03 
#> Pc2.Hydrophilicity.10 Pc2.Hydrophobicity.11 Pc2.Hydrophilicity.11 
#>          3.602591e-03          1.055082e-04          8.697920e-04 
#> Pc2.Hydrophobicity.12 Pc2.Hydrophilicity.12 Pc2.Hydrophobicity.13 
#>         -9.276413e-04         -2.001384e-03          1.705044e-03 
#> Pc2.Hydrophilicity.13 Pc2.Hydrophobicity.14 Pc2.Hydrophilicity.14 
#>          4.364007e-03          7.883453e-04         -9.441693e-04 
#> Pc2.Hydrophobicity.15 Pc2.Hydrophilicity.15 Pc2.Hydrophobicity.16 
#>         -3.133437e-04         -3.599332e-03          3.689079e-05 
#> Pc2.Hydrophilicity.16 Pc2.Hydrophobicity.17 Pc2.Hydrophilicity.17 
#>          2.483867e-03          4.832798e-04          2.465788e-03 
#> Pc2.Hydrophobicity.18 Pc2.Hydrophilicity.18 Pc2.Hydrophobicity.19 
#>         -3.142728e-04          2.021961e-03          6.421283e-05 
#> Pc2.Hydrophilicity.19 Pc2.Hydrophobicity.20 Pc2.Hydrophilicity.20 
#>         -8.896690e-04         -2.986886e-04          9.304039e-04 
#> Pc2.Hydrophobicity.21 Pc2.Hydrophilicity.21 Pc2.Hydrophobicity.22 
#>         -6.777458e-04          1.646818e-03          3.193506e-03 
#> Pc2.Hydrophilicity.22 Pc2.Hydrophobicity.23 Pc2.Hydrophilicity.23 
#>          3.270656e-03          2.533569e-03          2.478252e-03 
#> Pc2.Hydrophobicity.24 Pc2.Hydrophilicity.24 Pc2.Hydrophobicity.25 
#>         -2.489106e-03         -1.031008e-03         -3.992322e-03 
#> Pc2.Hydrophilicity.25 Pc2.Hydrophobicity.26 Pc2.Hydrophilicity.26 
#>         -2.596060e-03          8.690771e-04         -1.221378e-03 
#> Pc2.Hydrophobicity.27 Pc2.Hydrophilicity.27 Pc2.Hydrophobicity.28 
#>          5.208649e-03          4.617400e-03         -1.088584e-03 
#> Pc2.Hydrophilicity.28 Pc2.Hydrophobicity.29 Pc2.Hydrophilicity.29 
#>         -2.512263e-03          1.387641e-03          2.060890e-03 
#> Pc2.Hydrophobicity.30 Pc2.Hydrophilicity.30 
#>          3.177340e-04          1.451909e-03 

myprops = data.frame(
  AccNo = c("MyProp1", "MyProp2", "MyProp3"),
  A = c(0.62,  -0.5, 15),  R = c(-2.53,   3, 101),
  N = c(-0.78,  0.2, 58),  D = c(-0.9,    3, 59),
  C = c(0.29,    -1, 47),  E = c(-0.74,   3, 73),
  Q = c(-0.85,  0.2, 72),  G = c(0.48,    0, 1),
  H = c(-0.4,  -0.5, 82),  I = c(1.38, -1.8, 57),
  L = c(1.06,  -1.8, 57),  K = c(-1.5,    3, 73),
  M = c(0.64,  -1.3, 75),  F = c(1.19, -2.5, 91),
  P = c(0.12,     0, 42),  S = c(-0.18, 0.3, 31),
  T = c(-0.05, -0.4, 45),  W = c(0.81, -3.4, 130),
  Y = c(0.26,  -2.3, 107), V = c(1.08, -1.5, 43))

# use 2 default properties, 4 properties from the
# AAindex database, and 3 cutomized properties
extractAPAAC(
  x, customprops = myprops,
  props = c(
    "Hydrophobicity", "Hydrophilicity",
    "CIDH920105", "BHAR880101",
    "CHAM820101", "CHAM820102",
    "MyProp1", "MyProp2", "MyProp3")
)

#>                 Pc1.A                 Pc1.R                 Pc1.N 
#>          2.726537e+01          3.029486e+01          1.666217e+01 
#>                 Pc1.D                 Pc1.C                 Pc1.E 
#>          2.196377e+01          2.878011e+01          2.044903e+01 
#>                 Pc1.Q                 Pc1.G                 Pc1.H 
#>          2.044903e+01          3.483909e+01          1.287531e+01 
#>                 Pc1.I                 Pc1.L                 Pc1.K 
#>          1.514743e+01          3.256697e+01          1.666217e+01 
#>                 Pc1.M                 Pc1.F                 Pc1.P 
#>          5.301600e+00          1.363269e+01          2.272114e+01 
#>                 Pc1.S                 Pc1.T                 Pc1.W 
#>          3.786857e+01          1.893429e+01          9.845828e+00 
#>                 Pc1.Y                 Pc1.V  Pc2.Hydrophobicity.1 
#>          1.817691e+01          2.120640e+01          1.692861e-04 
#>  Pc2.Hydrophilicity.1      Pc2.CIDH920105.1      Pc2.BHAR880101.1 
#>          7.906313e-04          3.089061e-03          1.997195e-03 
#>      Pc2.CHAM820101.1      Pc2.CHAM820102.1         Pc2.MyProp1.1 
#>          1.525781e-03          2.118788e-03          1.692861e-04 
#>         Pc2.MyProp2.1         Pc2.MyProp3.1  Pc2.Hydrophobicity.2 
#>          7.906313e-04          1.439530e-03         -2.380818e-04 
#>  Pc2.Hydrophilicity.2      Pc2.CIDH920105.2      Pc2.BHAR880101.2 
#>         -1.413892e-04         -6.083517e-04          1.166533e-03 
#>      Pc2.CHAM820101.2      Pc2.CHAM820102.2         Pc2.MyProp1.2 
#>          1.928231e-03         -1.406107e-03         -2.380818e-04 
#>         Pc2.MyProp2.2         Pc2.MyProp3.2  Pc2.Hydrophobicity.3 
#>         -1.413892e-04          3.095179e-03          9.049982e-04 
#>  Pc2.Hydrophilicity.3      Pc2.CIDH920105.3      Pc2.BHAR880101.3 
#>          5.703830e-04         -6.052530e-04          1.407593e-03 
#>      Pc2.CHAM820101.3      Pc2.CHAM820102.3         Pc2.MyProp1.3 
#>          1.208629e-05         -3.818626e-03          9.049982e-04 
#>         Pc2.MyProp2.3         Pc2.MyProp3.3  Pc2.Hydrophobicity.4 
#>          5.703830e-04          7.898860e-05         -8.522536e-04 
#>  Pc2.Hydrophilicity.4      Pc2.CIDH920105.4      Pc2.BHAR880101.4 
#>         -3.463601e-04         -9.747006e-04         -2.234010e-03 
#>      Pc2.CHAM820101.4      Pc2.CHAM820102.4         Pc2.MyProp1.4 
#>          1.447898e-03         -3.222112e-03         -8.522536e-04 
#>         Pc2.MyProp2.4         Pc2.MyProp3.4  Pc2.Hydrophobicity.5 
#>         -3.463601e-04          1.759628e-03          1.361459e-03 
#>  Pc2.Hydrophilicity.5      Pc2.CIDH920105.5      Pc2.BHAR880101.5 
#>          1.133968e-03          2.993837e-03          2.675103e-03 
#>      Pc2.CHAM820101.5      Pc2.CHAM820102.5         Pc2.MyProp1.5 
#>          4.224830e-03          8.000519e-04          1.361459e-03 
#>         Pc2.MyProp2.5         Pc2.MyProp3.5  Pc2.Hydrophobicity.6 
#>          1.133968e-03          2.824775e-03         -1.111123e-03 
#>  Pc2.Hydrophilicity.6      Pc2.CIDH920105.6      Pc2.BHAR880101.6 
#>         -3.787263e-03         -1.619698e-03         -3.377592e-03 
#>      Pc2.CHAM820101.6      Pc2.CHAM820102.6         Pc2.MyProp1.6 
#>          3.047700e-03          5.464276e-04         -1.111123e-03 
#>         Pc2.MyProp2.6         Pc2.MyProp3.6  Pc2.Hydrophobicity.7 
#>         -3.787263e-03          4.399177e-03         -1.293396e-05 
#>  Pc2.Hydrophilicity.7      Pc2.CIDH920105.7      Pc2.BHAR880101.7 
#>          5.636157e-04         -1.375419e-03         -9.471565e-04 
#>      Pc2.CHAM820101.7      Pc2.CHAM820102.7         Pc2.MyProp1.7 
#>          2.510108e-03         -8.127858e-05         -1.293396e-05 
#>         Pc2.MyProp2.7         Pc2.MyProp3.7  Pc2.Hydrophobicity.8 
#>          5.636157e-04          2.327124e-03         -1.453212e-03 
#>  Pc2.Hydrophilicity.8      Pc2.CIDH920105.8      Pc2.BHAR880101.8 
#>         -1.486593e-03          9.121515e-04          2.923706e-03 
#>      Pc2.CHAM820101.8      Pc2.CHAM820102.8         Pc2.MyProp1.8 
#>          3.951525e-03          1.697524e-03         -1.453212e-03 
#>         Pc2.MyProp2.8         Pc2.MyProp3.8  Pc2.Hydrophobicity.9 
#>         -1.486593e-03          4.742073e-03         -2.259268e-03 
#>  Pc2.Hydrophilicity.9      Pc2.CIDH920105.9      Pc2.BHAR880101.9 
#>         -1.199058e-03         -1.997173e-04          1.563555e-05 
#>      Pc2.CHAM820101.9      Pc2.CHAM820102.9         Pc2.MyProp1.9 
#>          3.657590e-03         -2.088309e-03         -2.259268e-03 
#>         Pc2.MyProp2.9         Pc2.MyProp3.9 Pc2.Hydrophobicity.10 
#>         -1.199058e-03          3.186717e-03          2.248030e-03 
#> Pc2.Hydrophilicity.10     Pc2.CIDH920105.10     Pc2.BHAR880101.10 
#>          2.776775e-03          1.979571e-03          4.562424e-03 
#>     Pc2.CHAM820101.10     Pc2.CHAM820102.10        Pc2.MyProp1.10 
#>          3.953245e-04          1.488532e-03          2.248030e-03 
#>        Pc2.MyProp2.10        Pc2.MyProp3.10 Pc2.Hydrophobicity.11 
#>          2.776775e-03          7.015885e-05          8.132275e-05 
#> Pc2.Hydrophilicity.11     Pc2.CIDH920105.11     Pc2.BHAR880101.11 
#>          6.704110e-04          3.111663e-03          1.930726e-03 
#>     Pc2.CHAM820101.11     Pc2.CHAM820102.11        Pc2.MyProp1.11 
#>          2.920811e-03         -4.381601e-05          8.132275e-05 
#>        Pc2.MyProp2.11        Pc2.MyProp3.11 Pc2.Hydrophobicity.12 
#>          6.704110e-04          1.943423e-03         -7.149996e-04 
#> Pc2.Hydrophilicity.12     Pc2.CIDH920105.12     Pc2.BHAR880101.12 
#>         -1.542610e-03          2.052446e-04          7.637776e-04 
#>     Pc2.CHAM820101.12     Pc2.CHAM820102.12        Pc2.MyProp1.12 
#>          2.629436e-03          1.082862e-03         -7.149996e-04 
#>        Pc2.MyProp2.12        Pc2.MyProp3.12 Pc2.Hydrophobicity.13 
#>         -1.542610e-03          3.152375e-03          1.314200e-03 
#> Pc2.Hydrophilicity.13     Pc2.CIDH920105.13     Pc2.BHAR880101.13 
#>          3.363653e-03          3.154151e-03          3.770682e-03 
#>     Pc2.CHAM820101.13     Pc2.CHAM820102.13        Pc2.MyProp1.13 
#>          4.367833e-03          3.287165e-03          1.314200e-03 
#>        Pc2.MyProp2.13        Pc2.MyProp3.13 Pc2.Hydrophobicity.14 
#>          3.363653e-03          4.548631e-03          6.076342e-04 
#> Pc2.Hydrophilicity.14     Pc2.CIDH920105.14     Pc2.BHAR880101.14 
#>         -7.277389e-04          1.821700e-04          1.136616e-03 
#>     Pc2.CHAM820101.14     Pc2.CHAM820102.14        Pc2.MyProp1.14 
#>          2.725934e-03         -2.422624e-03          6.076342e-04 
#>        Pc2.MyProp2.14        Pc2.MyProp3.14 Pc2.Hydrophobicity.15 
#>         -7.277389e-04          3.644842e-03         -2.415164e-04 
#> Pc2.Hydrophilicity.15     Pc2.CIDH920105.15     Pc2.BHAR880101.15 
#>         -2.774263e-03          7.090685e-05          1.886016e-03 
#>     Pc2.CHAM820101.15     Pc2.CHAM820102.15        Pc2.MyProp1.15 
#>          3.086345e-03         -8.290011e-04         -2.415164e-04 
#>        Pc2.MyProp2.15        Pc2.MyProp3.15 Pc2.Hydrophobicity.16 
#>         -2.774263e-03          3.455495e-03          2.843438e-05 
#> Pc2.Hydrophilicity.16     Pc2.CIDH920105.16     Pc2.BHAR880101.16 
#>          1.914495e-03         -5.799552e-05          2.209231e-03 
#>     Pc2.CHAM820101.16     Pc2.CHAM820102.16        Pc2.MyProp1.16 
#>          9.447110e-04         -1.832162e-04          2.843438e-05 
#>        Pc2.MyProp2.16        Pc2.MyProp3.16 Pc2.Hydrophobicity.17 
#>          1.914495e-03          1.706065e-03          3.724984e-04 
#> Pc2.Hydrophilicity.17     Pc2.CIDH920105.17     Pc2.BHAR880101.17 
#>          1.900560e-03         -4.512885e-04          6.174000e-04 
#>     Pc2.CHAM820101.17     Pc2.CHAM820102.17        Pc2.MyProp1.17 
#>         -1.942794e-04          5.385367e-04          3.724984e-04 
#>        Pc2.MyProp2.17        Pc2.MyProp3.17 Pc2.Hydrophobicity.18 
#>          1.900560e-03         -2.383272e-04         -2.422325e-04 
#> Pc2.Hydrophilicity.18     Pc2.CIDH920105.18     Pc2.BHAR880101.18 
#>          1.558471e-03          2.694786e-03         -9.023092e-04 
#>     Pc2.CHAM820101.18     Pc2.CHAM820102.18        Pc2.MyProp1.18 
#>          3.617169e-03          2.025085e-03         -2.422325e-04 
#>        Pc2.MyProp2.18        Pc2.MyProp3.18 Pc2.Hydrophobicity.19 
#>          1.558471e-03          3.331545e-03          4.949343e-05 
#> Pc2.Hydrophilicity.19     Pc2.CIDH920105.19     Pc2.BHAR880101.19 
#>         -6.857316e-04          1.267582e-03         -3.028707e-04 
#>     Pc2.CHAM820101.19     Pc2.CHAM820102.19        Pc2.MyProp1.19 
#>          1.116845e-03         -9.942419e-04          4.949343e-05 
#>        Pc2.MyProp2.19        Pc2.MyProp3.19 Pc2.Hydrophobicity.20 
#>         -6.857316e-04          1.336517e-03         -2.302207e-04 
#> Pc2.Hydrophilicity.20     Pc2.CIDH920105.20     Pc2.BHAR880101.20 
#>          7.171290e-04          1.018011e-03          2.024248e-04 
#>     Pc2.CHAM820101.20     Pc2.CHAM820102.20        Pc2.MyProp1.20 
#>          3.103507e-03         -4.292626e-04         -2.302207e-04 
#>        Pc2.MyProp2.20        Pc2.MyProp3.20 Pc2.Hydrophobicity.21 
#>          7.171290e-04          2.234626e-03         -5.223873e-04 
#> Pc2.Hydrophilicity.21     Pc2.CIDH920105.21     Pc2.BHAR880101.21 
#>          1.269321e-03          2.767261e-03          6.379400e-04 
#>     Pc2.CHAM820101.21     Pc2.CHAM820102.21        Pc2.MyProp1.21 
#>          2.719312e-03         -5.515094e-04         -5.223873e-04 
#>        Pc2.MyProp2.21        Pc2.MyProp3.21 Pc2.Hydrophobicity.22 
#>          1.269321e-03          3.148752e-03          2.461464e-03 
#> Pc2.Hydrophilicity.22     Pc2.CIDH920105.22     Pc2.BHAR880101.22 
#>          2.520929e-03          4.365216e-03          4.823171e-03 
#>     Pc2.CHAM820101.22     Pc2.CHAM820102.22        Pc2.MyProp1.22 
#>          6.004310e-03         -3.525639e-04          2.461464e-03 
#>        Pc2.MyProp2.22        Pc2.MyProp3.22 Pc2.Hydrophobicity.23 
#>          2.520929e-03          5.044113e-03          1.952803e-03 
#> Pc2.Hydrophilicity.23     Pc2.CIDH920105.23     Pc2.BHAR880101.23 
#>          1.910167e-03          1.572223e-03          1.023679e-04 
#>     Pc2.CHAM820101.23     Pc2.CHAM820102.23        Pc2.MyProp1.23 
#>          2.158838e-03         -1.900535e-03          1.952803e-03 
#>        Pc2.MyProp2.23        Pc2.MyProp3.23 Pc2.Hydrophobicity.24 
#>          1.910167e-03          2.118122e-03         -1.918533e-03 
#> Pc2.Hydrophilicity.24     Pc2.CIDH920105.24     Pc2.BHAR880101.24 
#>         -7.946720e-04         -1.023436e-03          4.326760e-04 
#>     Pc2.CHAM820101.24     Pc2.CHAM820102.24        Pc2.MyProp1.24 
#>          8.041312e-04         -3.096339e-04         -1.918533e-03 
#>        Pc2.MyProp2.24        Pc2.MyProp3.24 Pc2.Hydrophobicity.25 
#>         -7.946720e-04          7.919708e-04         -3.077168e-03 
#> Pc2.Hydrophilicity.25     Pc2.CIDH920105.25     Pc2.BHAR880101.25 
#>         -2.000970e-03          1.267741e-03          1.367532e-03 
#>     Pc2.CHAM820101.25     Pc2.CHAM820102.25        Pc2.MyProp1.25 
#>          3.039638e-03          9.110689e-04         -3.077168e-03 
#>        Pc2.MyProp2.25        Pc2.MyProp3.25 Pc2.Hydrophobicity.26 
#>         -2.000970e-03          2.346114e-03          6.698600e-04 
#> Pc2.Hydrophilicity.26     Pc2.CIDH920105.26     Pc2.BHAR880101.26 
#>         -9.414037e-04         -1.228932e-04          6.990684e-04 
#>     Pc2.CHAM820101.26     Pc2.CHAM820102.26        Pc2.MyProp1.26 
#>          3.889658e-03         -3.211254e-05          6.698600e-04 
#>        Pc2.MyProp2.26        Pc2.MyProp3.26 Pc2.Hydrophobicity.27 
#>         -9.414037e-04          5.287878e-03          4.014679e-03 
#> Pc2.Hydrophilicity.27     Pc2.CIDH920105.27     Pc2.BHAR880101.27 
#>          3.558961e-03          2.958180e-03          3.952585e-03 
#>     Pc2.CHAM820101.27     Pc2.CHAM820102.27        Pc2.MyProp1.27 
#>          3.050843e-03          3.261111e-04          4.014679e-03 
#>        Pc2.MyProp2.27        Pc2.MyProp3.27 Pc2.Hydrophobicity.28 
#>          3.558961e-03          2.621260e-03         -8.390496e-04 
#> Pc2.Hydrophilicity.28     Pc2.CIDH920105.28     Pc2.BHAR880101.28 
#>         -1.936381e-03         -1.067034e-03          2.741618e-05 
#>     Pc2.CHAM820101.28     Pc2.CHAM820102.28        Pc2.MyProp1.28 
#>          2.754108e-03         -4.269171e-04         -8.390496e-04 
#>        Pc2.MyProp2.28        Pc2.MyProp3.28 Pc2.Hydrophobicity.29 
#>         -1.936381e-03          1.931571e-03          1.069554e-03 
#> Pc2.Hydrophilicity.29     Pc2.CIDH920105.29     Pc2.BHAR880101.29 
#>          1.588476e-03          1.768304e-03          2.975706e-03 
#>     Pc2.CHAM820101.29     Pc2.CHAM820102.29        Pc2.MyProp1.29 
#>          7.609157e-04          1.224138e-03          1.069554e-03 
#>        Pc2.MyProp2.29        Pc2.MyProp3.29 Pc2.Hydrophobicity.30 
#>          1.588476e-03          5.908359e-04          2.449004e-04 
#> Pc2.Hydrophilicity.30     Pc2.CIDH920105.30     Pc2.BHAR880101.30 
#>          1.119090e-03          7.795027e-04          1.672987e-03 
#>     Pc2.CHAM820101.30     Pc2.CHAM820102.30        Pc2.MyProp1.30 
#>          3.002246e-03          3.108433e-03          2.449004e-04 
#>        Pc2.MyProp2.30        Pc2.MyProp3.30 
#>          1.119090e-03          3.398358e-03