Conjoint Triad Descriptor

This function calculates the Conjoint Triad descriptor (dim: 343).

Usage

extractCTriad(x)

Arguments

x

A character vector, as the input protein sequence.

Value

A length 343 named vector

Note

For this descriptor type, users need to intelligently evaluate the underlying details of the descriptors provided, instead of using this function with their data blindly. It would be wise to use some negative and positive control comparisons where relevant to help guide interpretation of the results.

References

J.W. Shen, J. Zhang, X.M. Luo, W.L. Zhu, K.Q. Yu, K.X. Chen, Y.X. Li, H.L. Jiang. Predicting Protein-protein Interactions Based Only on Sequences Information. Proceedings of the National Academy of Sciences. 007, 104, 4337–4341.

Author

Nan Xiao <https://nanx.me>

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractCTriad(x)
#> VS111 VS211 VS311 VS411 VS511 VS611 VS711 VS121 VS221 VS321 VS421 VS521 VS621 
#>   0.1   0.3   0.6   0.2   0.4   0.0   0.3   1.0   0.6   0.5   0.0   0.2   0.3 
#> VS721 VS131 VS231 VS331 VS431 VS531 VS631 VS731 VS141 VS241 VS341 VS441 VS541 
#>   0.0   0.2   0.4   0.5   0.2   0.3   0.3   0.1   0.3   0.3   0.2   0.2   0.0 
#> VS641 VS741 VS151 VS251 VS351 VS451 VS551 VS651 VS751 VS161 VS261 VS361 VS461 
#>   0.1   0.2   0.2   0.2   0.5   0.1   0.2   0.0   0.0   0.1   0.4   0.2   0.3 
#> VS561 VS661 VS761 VS171 VS271 VS371 VS471 VS571 VS671 VS771 VS112 VS212 VS312 
#>   0.2   0.0   0.1   0.1   0.3   0.1   0.0   0.1   0.0   0.1   0.8   0.4   0.4 
#> VS412 VS512 VS612 VS712 VS122 VS222 VS322 VS422 VS522 VS622 VS722 VS132 VS232 
#>   0.6   0.1   0.5   0.2   0.8   0.5   0.2   0.3   0.2   0.0   0.2   0.1   0.3 
#> VS332 VS432 VS532 VS632 VS732 VS142 VS242 VS342 VS442 VS542 VS642 VS742 VS152 
#>   0.3   0.4   0.2   0.0   0.1   0.1   0.3   0.3   0.3   0.2   0.1   0.1   0.3 
#> VS252 VS352 VS452 VS552 VS652 VS752 VS162 VS262 VS362 VS462 VS562 VS662 VS762 
#>   0.3   0.2   0.2   0.2   0.3   0.0   0.2   0.0   0.3   0.2   0.1   0.0   0.2 
#> VS172 VS272 VS372 VS472 VS572 VS672 VS772 VS113 VS213 VS313 VS413 VS513 VS613 
#>   0.2   0.0   0.1   0.2   0.1   0.0   0.0   0.3   0.2   0.3   0.1   0.5   0.1 
#> VS713 VS123 VS223 VS323 VS423 VS523 VS623 VS723 VS133 VS233 VS333 VS433 VS533 
#>   0.0   0.4   0.4   0.5   0.2   0.3   0.1   0.0   0.2   0.5   0.1   0.4   0.4 
#> VS633 VS733 VS143 VS243 VS343 VS443 VS543 VS643 VS743 VS153 VS253 VS353 VS453 
#>   0.3   0.3   0.1   0.0   0.4   0.5   0.1   0.3   0.2   0.6   0.2   0.1   0.1 
#> VS553 VS653 VS753 VS163 VS263 VS363 VS463 VS563 VS663 VS763 VS173 VS273 VS373 
#>   0.1   0.2   0.0   0.4   0.1   0.2   0.2   0.1   0.1   0.0   0.1   0.1   0.3 
#> VS473 VS573 VS673 VS773 VS114 VS214 VS314 VS414 VS514 VS614 VS714 VS124 VS224 
#>   0.1   0.1   0.2   0.0   0.1   0.4   0.2   0.1   0.1   0.2   0.2   0.1   0.4 
#> VS324 VS424 VS524 VS624 VS724 VS134 VS234 VS334 VS434 VS534 VS634 VS734 VS144 
#>   0.0   0.3   0.2   0.2   0.3   0.4   0.4   0.3   0.1   0.2   0.0   0.1   0.2 
#> VS244 VS344 VS444 VS544 VS644 VS744 VS154 VS254 VS354 VS454 VS554 VS654 VS754 
#>   0.3   0.3   0.4   0.1   0.0   0.1   0.3   0.2   0.0   0.0   0.0   0.0   0.2 
#> VS164 VS264 VS364 VS464 VS564 VS664 VS764 VS174 VS274 VS374 VS474 VS574 VS674 
#>   0.0   0.2   0.2   0.0   0.1   0.2   0.0   0.2   0.0   0.3   0.3   0.0   0.0 
#> VS774 VS115 VS215 VS315 VS415 VS515 VS615 VS715 VS125 VS225 VS325 VS425 VS525 
#>   0.0   0.4   0.5   0.3   0.2   0.0   0.2   0.0   0.5   0.0   0.0   0.4   0.3 
#> VS625 VS725 VS135 VS235 VS335 VS435 VS535 VS635 VS735 VS145 VS245 VS345 VS445 
#>   0.1   0.0   0.2   0.0   0.5   0.1   0.0   0.3   0.0   0.3   0.1   0.1   0.0 
#> VS545 VS645 VS745 VS155 VS255 VS355 VS455 VS555 VS655 VS755 VS165 VS265 VS365 
#>   0.2   0.1   0.0   0.1   0.1   0.1   0.2   0.0   0.0   0.0   0.2   0.1   0.1 
#> VS465 VS565 VS665 VS765 VS175 VS275 VS375 VS475 VS575 VS675 VS775 VS116 VS216 
#>   0.1   0.0   0.1   0.0   0.0   0.1   0.2   0.0   0.0   0.0   0.0   0.2   0.4 
#> VS316 VS416 VS516 VS616 VS716 VS126 VS226 VS326 VS426 VS526 VS626 VS726 VS136 
#>   0.1   0.1   0.1   0.2   0.0   0.0   0.1   0.1   0.2   0.2   0.2   0.1   0.1 
#> VS236 VS336 VS436 VS536 VS636 VS736 VS146 VS246 VS346 VS446 VS546 VS646 VS746 
#>   0.2   0.3   0.2   0.1   0.0   0.3   0.1   0.3   0.2   0.0   0.1   0.1   0.0 
#> VS156 VS256 VS356 VS456 VS556 VS656 VS756 VS166 VS266 VS366 VS466 VS566 VS666 
#>   0.0   0.2   0.0   0.2   0.0   0.1   0.1   0.1   0.1   0.1   0.0   0.1   0.2 
#> VS766 VS176 VS276 VS376 VS476 VS576 VS676 VS776 VS117 VS217 VS317 VS417 VS517 
#>   0.0   0.0   0.0   0.1   0.0   0.0   0.1   0.1   0.0   0.4   0.1   0.0   0.0 
#> VS617 VS717 VS127 VS227 VS327 VS427 VS527 VS627 VS727 VS137 VS237 VS337 VS437 
#>   0.1   0.0   0.2   0.2   0.1   0.0   0.0   0.1   0.0   0.3   0.1   0.2   0.2 
#> VS537 VS637 VS737 VS147 VS247 VS347 VS447 VS547 VS647 VS747 VS157 VS257 VS357 
#>   0.1   0.2   0.0   0.2   0.2   0.0   0.0   0.0   0.0   0.2   0.1   0.1   0.2 
#> VS457 VS557 VS657 VS757 VS167 VS267 VS367 VS467 VS567 VS667 VS767 VS177 VS277 
#>   0.0   0.0   0.0   0.0   0.1   0.0   0.2   0.0   0.0   0.0   0.0   0.0   0.1 
#> VS377 VS477 VS577 VS677 VS777 
#>   0.0   0.0   0.1   0.0   0.0