Geary Autocorrelation Descriptor

This function calculates the Geary autocorrelation descriptor (dim: length(props) * nlag).

extractGeary(x, props = c("CIDH920105", "BHAR880101", "CHAM820101",
  "CHAM820102", "CHOC760101", "BIGC670101", "CHAM810101", "DAYM780201"),
  nlag = 30L, customprops = NULL)

Arguments

x
A character vector, as the input protein sequence.

props
A character vector, specifying the Accession Number of the target properties. 8 properties are used by default, as listed below:

AccNo. CIDH920105

Normalized average hydrophobicity scales (Cid et al., 1992)

AccNo. BHAR880101

Average flexibility indices (Bhaskaran-Ponnuswamy, 1988)

AccNo. CHAM820101

Polarizability parameter (Charton-Charton, 1982)

AccNo. CHAM820102

Free energy of solution in water, kcal/mole (Charton-Charton, 1982)

AccNo. CHOC760101

Residue accessible surface area in tripeptide (Chothia, 1976)

AccNo. BIGC670101

Residue volume (Bigelow, 1967)

AccNo. CHAM810101

Steric parameter (Charton, 1981)

AccNo. DAYM780201

Relative mutability (Dayhoff et al., 1978b)

nlag
Maximum value of the lag parameter. Default is 30.

customprops
A n x 21 named data frame contains n customized property. Each row contains one property. The column order for different amino acid types is 'AccNo', 'A', 'R', 'N', 'D', 'C', 'E', 'Q', 'G', 'H', 'I', 'L', 'K', 'M', 'F', 'P', 'S', 'T', 'W', 'Y', 'V', and the columns should also be exactly named like this. The AccNo column contains the properties' names. Then users should explicitly specify these properties with these names in the argument props. See the examples below for a demonstration. The default value for customprops is NULL.

Value

A length length(props) * nlag named vector.

Note

For this descriptor type, users need to intelligently evaluate the underlying details of the descriptors provided, instead of using this function with their data blindly. It would be wise to use some negative and positive control comparisons where relevant to help guide interpretation of the results.

References

AAindex: Amino acid index database. http://www.genome.ad.jp/dbget/aaindex.html

Feng, Z.P. and Zhang, C.T. (2000) Prediction of membrane protein types based on the hydrophobic index of amino acids. Journal of Protein Chemistry, 19, 269-275.

Horne, D.S. (1988) Prediction of protein helix content from an autocorrelation analysis of sequence hydrophobicities. Biopolymers, 27, 451-477.

Sokal, R.R. and Thomson, B.A. (2006) Population structure inferred by local spatial autocorrelation: an usage from an Amerindian tribal population. American Journal of Physical Anthropology, 129, 121-131.

See also

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractGeary(x)

#>  CIDH920105.lag1  CIDH920105.lag2  CIDH920105.lag3  CIDH920105.lag4 
#>        0.9361830        1.0442920        1.0452843        1.0563467 
#>  CIDH920105.lag5  CIDH920105.lag6  CIDH920105.lag7  CIDH920105.lag8 
#>        0.9406031        1.0765517        1.0675786        0.9991363 
#>  CIDH920105.lag9 CIDH920105.lag10 CIDH920105.lag11 CIDH920105.lag12 
#>        1.0316555        0.9684585        0.9353130        1.0201990 
#> CIDH920105.lag13 CIDH920105.lag14 CIDH920105.lag15 CIDH920105.lag16 
#>        0.9340933        1.0207373        1.0251486        1.0290464 
#> CIDH920105.lag17 CIDH920105.lag18 CIDH920105.lag19 CIDH920105.lag20 
#>        1.0414375        0.9494403        0.9905987        0.9987183 
#> CIDH920105.lag21 CIDH920105.lag22 CIDH920105.lag23 CIDH920105.lag24 
#>        0.9472542        0.9010009        0.9828848        1.0574098 
#> CIDH920105.lag25 CIDH920105.lag26 CIDH920105.lag27 CIDH920105.lag28 
#>        0.9897955        1.0290018        0.9400066        1.0584150 
#> CIDH920105.lag29 CIDH920105.lag30  BHAR880101.lag1  BHAR880101.lag2 
#>        0.9762904        1.0029734        0.9818711        1.0051730 
#>  BHAR880101.lag3  BHAR880101.lag4  BHAR880101.lag5  BHAR880101.lag6 
#>        0.9967069        1.1012905        0.9595859        1.1337056 
#>  BHAR880101.lag7  BHAR880101.lag8  BHAR880101.lag9 BHAR880101.lag10 
#>        1.0628740        0.9497400        1.0338201        0.9023549 
#> BHAR880101.lag11 BHAR880101.lag12 BHAR880101.lag13 BHAR880101.lag14 
#>        0.9802386        1.0151581        0.9292786        1.0066693 
#> BHAR880101.lag15 BHAR880101.lag16 BHAR880101.lag17 BHAR880101.lag18 
#>        0.9856150        0.9759971        1.0232794        1.0687209 
#> BHAR880101.lag19 BHAR880101.lag20 BHAR880101.lag21 BHAR880101.lag22 
#>        1.0491804        1.0351528        1.0232000        0.9016895 
#> BHAR880101.lag23 BHAR880101.lag24 BHAR880101.lag25 BHAR880101.lag26 
#>        1.0391900        1.0301878        1.0021211        1.0222207 
#> BHAR880101.lag27 BHAR880101.lag28 BHAR880101.lag29 BHAR880101.lag30 
#>        0.9260382        1.0403725        0.9530778        0.9884374 
#>  CHAM820101.lag1  CHAM820101.lag2  CHAM820101.lag3  CHAM820101.lag4 
#>        1.0133839        1.0022438        1.0551890        1.0169637 
#>  CHAM820101.lag5  CHAM820101.lag6  CHAM820101.lag7  CHAM820101.lag8 
#>        0.9413192        0.9730229        0.9865679        0.9428985 
#>  CHAM820101.lag9 CHAM820101.lag10 CHAM820101.lag11 CHAM820101.lag12 
#>        0.9521705        1.0440834        0.9744226        0.9842079 
#> CHAM820101.lag13 CHAM820101.lag14 CHAM820101.lag15 CHAM820101.lag16 
#>        0.9375430        0.9847534        0.9762810        1.0337143 
#> CHAM820101.lag17 CHAM820101.lag18 CHAM820101.lag19 CHAM820101.lag20 
#>        1.0637944        0.9600194        1.0264610        0.9730978 
#> CHAM820101.lag21 CHAM820101.lag22 CHAM820101.lag23 CHAM820101.lag24 
#>        0.9841629        0.8944897        1.0012565        1.0407044 
#> CHAM820101.lag25 CHAM820101.lag26 CHAM820101.lag27 CHAM820101.lag28 
#>        0.9774393        0.9556452        0.9771782        0.9855390 
#> CHAM820101.lag29 CHAM820101.lag30  CHAM820102.lag1  CHAM820102.lag2 
#>        1.0378770        0.9705165        0.9503346        1.0343446 
#>  CHAM820102.lag3  CHAM820102.lag4  CHAM820102.lag5  CHAM820102.lag6 
#>        1.0930293        1.0805000        0.9861535        0.9921047 
#>  CHAM820102.lag7  CHAM820102.lag8  CHAM820102.lag9 CHAM820102.lag10 
#>        1.0083331        0.9676778        1.0545499        0.9671392 
#> CHAM820102.lag11 CHAM820102.lag12 CHAM820102.lag13 CHAM820102.lag14 
#>        1.0000784        0.9749532        0.9242556        1.0615438 
#> CHAM820102.lag15 CHAM820102.lag16 CHAM820102.lag17 CHAM820102.lag18 
#>        1.0210459        1.0068944        0.9859525        0.9524320 
#> CHAM820102.lag19 CHAM820102.lag20 CHAM820102.lag21 CHAM820102.lag22 
#>        1.0254189        1.0110984        1.0153212        1.0091916 
#> CHAM820102.lag23 CHAM820102.lag24 CHAM820102.lag25 CHAM820102.lag26 
#>        1.0471259        1.0112268        0.9833002        1.0031331 
#> CHAM820102.lag27 CHAM820102.lag28 CHAM820102.lag29 CHAM820102.lag30 
#>        0.9958332        1.0151695        0.9763861        0.9332919 
#>  CHOC760101.lag1  CHOC760101.lag2  CHOC760101.lag3  CHOC760101.lag4 
#>        1.0150406        0.9729159        1.0498232        0.9944486 
#>  CHOC760101.lag5  CHOC760101.lag6  CHOC760101.lag7  CHOC760101.lag8 
#>        0.9503834        0.9408494        0.9576772        0.9247629 
#>  CHOC760101.lag9 CHOC760101.lag10 CHOC760101.lag11 CHOC760101.lag12 
#>        0.9614165        1.0452973        1.0171362        0.9890922 
#> CHOC760101.lag13 CHOC760101.lag14 CHOC760101.lag15 CHOC760101.lag16 
#>        0.9510173        0.9532281        0.9770492        1.0067054 
#> CHOC760101.lag17 CHOC760101.lag18 CHOC760101.lag19 CHOC760101.lag20 
#>        1.0677338        0.9598050        1.0069556        1.0038437 
#> CHOC760101.lag21 CHOC760101.lag22 CHOC760101.lag23 CHOC760101.lag24 
#>        0.9524803        0.9014090        0.9852450        1.0379942 
#> CHOC760101.lag25 CHOC760101.lag26 CHOC760101.lag27 CHOC760101.lag28 
#>        1.0006621        0.9339269        0.9963756        1.0052963 
#> CHOC760101.lag29 CHOC760101.lag30  BIGC670101.lag1  BIGC670101.lag2 
#>        1.0413666        0.9467525        1.0059350        0.9929376 
#>  BIGC670101.lag3  BIGC670101.lag4  BIGC670101.lag5  BIGC670101.lag6 
#>        1.0731913        1.0077837        0.9192554        0.9672734 
#>  BIGC670101.lag7  BIGC670101.lag8  BIGC670101.lag9 BIGC670101.lag10 
#>        0.9461109        0.9529949        0.9730305        1.0427910 
#> BIGC670101.lag11 BIGC670101.lag12 BIGC670101.lag13 BIGC670101.lag14 
#>        1.0039312        0.9984156        0.9450803        0.9733642 
#> BIGC670101.lag15 BIGC670101.lag16 BIGC670101.lag17 BIGC670101.lag18 
#>        0.9915806        1.0137285        1.0702942        0.9415800 
#> BIGC670101.lag19 BIGC670101.lag20 BIGC670101.lag21 BIGC670101.lag22 
#>        1.0139971        0.9625814        0.9565350        0.8898701 
#> BIGC670101.lag23 BIGC670101.lag24 BIGC670101.lag25 BIGC670101.lag26 
#>        0.9905643        1.0442658        0.9891278        0.9570413 
#> BIGC670101.lag27 BIGC670101.lag28 BIGC670101.lag29 BIGC670101.lag30 
#>        0.9987817        1.0008018        1.0376577        0.9536494 
#>  CHAM810101.lag1  CHAM810101.lag2  CHAM810101.lag3  CHAM810101.lag4 
#>        0.9927308        0.9615010        1.0668733        1.0353024 
#>  CHAM810101.lag5  CHAM810101.lag6  CHAM810101.lag7  CHAM810101.lag8 
#>        0.9916586        0.9749523        0.9871645        0.9444802 
#>  CHAM810101.lag9 CHAM810101.lag10 CHAM810101.lag11 CHAM810101.lag12 
#>        1.0291907        1.0194290        1.0607408        0.9999449 
#> CHAM810101.lag13 CHAM810101.lag14 CHAM810101.lag15 CHAM810101.lag16 
#>        0.9931695        0.9297656        0.9611833        0.9814565 
#> CHAM810101.lag17 CHAM810101.lag18 CHAM810101.lag19 CHAM810101.lag20 
#>        1.0321136        0.9892586        1.0563039        1.0206550 
#> CHAM810101.lag21 CHAM810101.lag22 CHAM810101.lag23 CHAM810101.lag24 
#>        0.9041110        0.9103649        0.9775755        1.0068787 
#> CHAM810101.lag25 CHAM810101.lag26 CHAM810101.lag27 CHAM810101.lag28 
#>        1.0079916        0.9596887        1.0191264        1.0140882 
#> CHAM810101.lag29 CHAM810101.lag30  DAYM780201.lag1  DAYM780201.lag2 
#>        0.9151374        0.9253788        0.9646010        1.0522299 
#>  DAYM780201.lag3  DAYM780201.lag4  DAYM780201.lag5  DAYM780201.lag6 
#>        1.0335338        1.1041881        0.9693172        1.0428607 
#>  DAYM780201.lag7  DAYM780201.lag8  DAYM780201.lag9 DAYM780201.lag10 
#>        1.1097225        0.9169442        0.9054205        0.9649199 
#> DAYM780201.lag11 DAYM780201.lag12 DAYM780201.lag13 DAYM780201.lag14 
#>        1.0263117        1.0189425        0.9578808        1.0286703 
#> DAYM780201.lag15 DAYM780201.lag16 DAYM780201.lag17 DAYM780201.lag18 
#>        1.0104818        1.0764483        0.9954050        0.9180787 
#> DAYM780201.lag19 DAYM780201.lag20 DAYM780201.lag21 DAYM780201.lag22 
#>        0.9481601        1.0128404        0.9719377        0.9898303 
#> DAYM780201.lag23 DAYM780201.lag24 DAYM780201.lag25 DAYM780201.lag26 
#>        0.9977120        0.9509454        1.0878960        1.0429411 
#> DAYM780201.lag27 DAYM780201.lag28 DAYM780201.lag29 DAYM780201.lag30 
#>        0.9938437        0.9506562        0.9532393        1.0463685 

myprops <- data.frame(
  AccNo = c("MyProp1", "MyProp2", "MyProp3"),
  A = c(0.62, -0.5, 15), R = c(-2.53, 3, 101),
  N = c(-0.78, 0.2, 58), D = c(-0.9, 3, 59),
  C = c(0.29, -1, 47), E = c(-0.74, 3, 73),
  Q = c(-0.85, 0.2, 72), G = c(0.48, 0, 1),
  H = c(-0.4, -0.5, 82), I = c(1.38, -1.8, 57),
  L = c(1.06, -1.8, 57), K = c(-1.5, 3, 73),
  M = c(0.64, -1.3, 75), F = c(1.19, -2.5, 91),
  P = c(0.12, 0, 42), S = c(-0.18, 0.3, 31),
  T = c(-0.05, -0.4, 45), W = c(0.81, -3.4, 130),
  Y = c(0.26, -2.3, 107), V = c(1.08, -1.5, 43)
)

# Use 4 properties in the AAindex database, and 3 cutomized properties
extractGeary(
  x,
  customprops = myprops,
  props = c(
    "CIDH920105", "BHAR880101",
    "CHAM820101", "CHAM820102",
    "MyProp1", "MyProp2", "MyProp3"
  )
)

#>  CIDH920105.lag1  CIDH920105.lag2  CIDH920105.lag3  CIDH920105.lag4 
#>        0.9361830        1.0442920        1.0452843        1.0563467 
#>  CIDH920105.lag5  CIDH920105.lag6  CIDH920105.lag7  CIDH920105.lag8 
#>        0.9406031        1.0765517        1.0675786        0.9991363 
#>  CIDH920105.lag9 CIDH920105.lag10 CIDH920105.lag11 CIDH920105.lag12 
#>        1.0316555        0.9684585        0.9353130        1.0201990 
#> CIDH920105.lag13 CIDH920105.lag14 CIDH920105.lag15 CIDH920105.lag16 
#>        0.9340933        1.0207373        1.0251486        1.0290464 
#> CIDH920105.lag17 CIDH920105.lag18 CIDH920105.lag19 CIDH920105.lag20 
#>        1.0414375        0.9494403        0.9905987        0.9987183 
#> CIDH920105.lag21 CIDH920105.lag22 CIDH920105.lag23 CIDH920105.lag24 
#>        0.9472542        0.9010009        0.9828848        1.0574098 
#> CIDH920105.lag25 CIDH920105.lag26 CIDH920105.lag27 CIDH920105.lag28 
#>        0.9897955        1.0290018        0.9400066        1.0584150 
#> CIDH920105.lag29 CIDH920105.lag30  BHAR880101.lag1  BHAR880101.lag2 
#>        0.9762904        1.0029734        0.9818711        1.0051730 
#>  BHAR880101.lag3  BHAR880101.lag4  BHAR880101.lag5  BHAR880101.lag6 
#>        0.9967069        1.1012905        0.9595859        1.1337056 
#>  BHAR880101.lag7  BHAR880101.lag8  BHAR880101.lag9 BHAR880101.lag10 
#>        1.0628740        0.9497400        1.0338201        0.9023549 
#> BHAR880101.lag11 BHAR880101.lag12 BHAR880101.lag13 BHAR880101.lag14 
#>        0.9802386        1.0151581        0.9292786        1.0066693 
#> BHAR880101.lag15 BHAR880101.lag16 BHAR880101.lag17 BHAR880101.lag18 
#>        0.9856150        0.9759971        1.0232794        1.0687209 
#> BHAR880101.lag19 BHAR880101.lag20 BHAR880101.lag21 BHAR880101.lag22 
#>        1.0491804        1.0351528        1.0232000        0.9016895 
#> BHAR880101.lag23 BHAR880101.lag24 BHAR880101.lag25 BHAR880101.lag26 
#>        1.0391900        1.0301878        1.0021211        1.0222207 
#> BHAR880101.lag27 BHAR880101.lag28 BHAR880101.lag29 BHAR880101.lag30 
#>        0.9260382        1.0403725        0.9530778        0.9884374 
#>  CHAM820101.lag1  CHAM820101.lag2  CHAM820101.lag3  CHAM820101.lag4 
#>        1.0133839        1.0022438        1.0551890        1.0169637 
#>  CHAM820101.lag5  CHAM820101.lag6  CHAM820101.lag7  CHAM820101.lag8 
#>        0.9413192        0.9730229        0.9865679        0.9428985 
#>  CHAM820101.lag9 CHAM820101.lag10 CHAM820101.lag11 CHAM820101.lag12 
#>        0.9521705        1.0440834        0.9744226        0.9842079 
#> CHAM820101.lag13 CHAM820101.lag14 CHAM820101.lag15 CHAM820101.lag16 
#>        0.9375430        0.9847534        0.9762810        1.0337143 
#> CHAM820101.lag17 CHAM820101.lag18 CHAM820101.lag19 CHAM820101.lag20 
#>        1.0637944        0.9600194        1.0264610        0.9730978 
#> CHAM820101.lag21 CHAM820101.lag22 CHAM820101.lag23 CHAM820101.lag24 
#>        0.9841629        0.8944897        1.0012565        1.0407044 
#> CHAM820101.lag25 CHAM820101.lag26 CHAM820101.lag27 CHAM820101.lag28 
#>        0.9774393        0.9556452        0.9771782        0.9855390 
#> CHAM820101.lag29 CHAM820101.lag30  CHAM820102.lag1  CHAM820102.lag2 
#>        1.0378770        0.9705165        0.9503346        1.0343446 
#>  CHAM820102.lag3  CHAM820102.lag4  CHAM820102.lag5  CHAM820102.lag6 
#>        1.0930293        1.0805000        0.9861535        0.9921047 
#>  CHAM820102.lag7  CHAM820102.lag8  CHAM820102.lag9 CHAM820102.lag10 
#>        1.0083331        0.9676778        1.0545499        0.9671392 
#> CHAM820102.lag11 CHAM820102.lag12 CHAM820102.lag13 CHAM820102.lag14 
#>        1.0000784        0.9749532        0.9242556        1.0615438 
#> CHAM820102.lag15 CHAM820102.lag16 CHAM820102.lag17 CHAM820102.lag18 
#>        1.0210459        1.0068944        0.9859525        0.9524320 
#> CHAM820102.lag19 CHAM820102.lag20 CHAM820102.lag21 CHAM820102.lag22 
#>        1.0254189        1.0110984        1.0153212        1.0091916 
#> CHAM820102.lag23 CHAM820102.lag24 CHAM820102.lag25 CHAM820102.lag26 
#>        1.0471259        1.0112268        0.9833002        1.0031331 
#> CHAM820102.lag27 CHAM820102.lag28 CHAM820102.lag29 CHAM820102.lag30 
#>        0.9958332        1.0151695        0.9763861        0.9332919 
#>     MyProp1.lag1     MyProp1.lag2     MyProp1.lag3     MyProp1.lag4 
#>        0.9987014        1.0042430        0.9768573        1.0215035 
#>     MyProp1.lag5     MyProp1.lag6     MyProp1.lag7     MyProp1.lag8 
#>        0.9654430        1.0173820        0.9899070        1.0259654 
#>     MyProp1.lag9    MyProp1.lag10    MyProp1.lag11    MyProp1.lag12 
#>        1.0470257        0.9354884        0.9901335        1.0102046 
#>    MyProp1.lag13    MyProp1.lag14    MyProp1.lag15    MyProp1.lag16 
#>        0.9604603        0.9777933        0.9986361        0.9935197 
#>    MyProp1.lag17    MyProp1.lag18    MyProp1.lag19    MyProp1.lag20 
#>        0.9863639        1.0013734        0.9944359        1.0021568 
#>    MyProp1.lag21    MyProp1.lag22    MyProp1.lag23    MyProp1.lag24 
#>        1.0101742        0.9369555        0.9493721        1.0463017 
#>    MyProp1.lag25    MyProp1.lag26    MyProp1.lag27    MyProp1.lag28 
#>        1.0762954        0.9830056        0.9012154        1.0225609 
#>    MyProp1.lag29    MyProp1.lag30     MyProp2.lag1     MyProp2.lag2 
#>        0.9707321        0.9912091        0.9900593        1.0132171 
#>     MyProp2.lag3     MyProp2.lag4     MyProp2.lag5     MyProp2.lag6 
#>        0.9939543        1.0220681        0.9749927        1.1142163 
#>     MyProp2.lag7     MyProp2.lag8     MyProp2.lag9    MyProp2.lag10 
#>        0.9886743        1.0465832        1.0369295        0.9221438 
#>    MyProp2.lag11    MyProp2.lag12    MyProp2.lag13    MyProp2.lag14 
#>        0.9835647        1.0490332        0.9073427        1.0270654 
#>    MyProp2.lag15    MyProp2.lag16    MyProp2.lag17    MyProp2.lag18 
#>        1.0853336        0.9506871        0.9516677        0.9625250 
#>    MyProp2.lag19    MyProp2.lag20    MyProp2.lag21    MyProp2.lag22 
#>        1.0281791        0.9887198        0.9739604        0.9363962 
#>    MyProp2.lag23    MyProp2.lag24    MyProp2.lag25    MyProp2.lag26 
#>        0.9529674        1.0334644        1.0674287        1.0375601 
#>    MyProp2.lag27    MyProp2.lag28    MyProp2.lag29    MyProp2.lag30 
#>        0.9084307        1.0695042        0.9657679        0.9768113 
#>     MyProp3.lag1     MyProp3.lag2     MyProp3.lag3     MyProp3.lag4 
#>        1.0146280        0.9722862        1.0497983        1.0080631 
#>     MyProp3.lag5     MyProp3.lag6     MyProp3.lag7     MyProp3.lag8 
#>        0.9823002        0.9406788        0.9923895        0.9274328 
#>     MyProp3.lag9    MyProp3.lag10    MyProp3.lag11    MyProp3.lag12 
#>        0.9692848        1.0514912        1.0026759        0.9732718 
#>    MyProp3.lag13    MyProp3.lag14    MyProp3.lag15    MyProp3.lag16 
#>        0.9387388        0.9636088        0.9700880        1.0131516 
#>    MyProp3.lag17    MyProp3.lag18    MyProp3.lag19    MyProp3.lag20 
#>        1.0611696        0.9700136        1.0189479        0.9968361 
#>    MyProp3.lag21    MyProp3.lag22    MyProp3.lag23    MyProp3.lag24 
#>        0.9737338        0.9262255        1.0027198        1.0388112 
#>    MyProp3.lag25    MyProp3.lag26    MyProp3.lag27    MyProp3.lag28 
#>        0.9966404        0.9222564        0.9889491        1.0064019 
#>    MyProp3.lag29    MyProp3.lag30 
#>        1.0376997        0.9617269