Protein Sequence Alignment for Two Protein Sequences

This function implements the sequence alignment between two protein sequences.

twoSeqSim(
  seq1,
  seq2,
  type = "local",
  submat = "BLOSUM62",
  gap.opening = 10,
  gap.extension = 4
)

Arguments

seq1
Character string, containing one protein sequence.

seq2
Character string, containing another protein sequence.

type
Type of alignment, default is "local", could be "global" or "local", where "global" represents Needleman-Wunsch global alignment; "local" represents Smith-Waterman local alignment.

submat
Substitution matrix, default is "BLOSUM62", can be one of "BLOSUM45", "BLOSUM50", "BLOSUM62", "BLOSUM80", "BLOSUM100", "PAM30", "PAM40", "PAM70", "PAM120", or "PAM250".

gap.opening
The cost required to open a gap of any length in the alignment. Defaults to 10.

gap.extension
The cost to extend the length of an existing gap by 1. Defaults to 4.

Value

An Biostrings object containing the alignment scores and other alignment information.

See also

Examples

if (FALSE) {

# Be careful when testing this since it involves sequence alignment
# and might produce unpredictable results in some environments
library("Biostrings")
s1 <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
s2 <- readFASTA(system.file("protseq/P10323.fasta", package = "protr"))[[1]]
seqalign <- twoSeqSim(s1, s2)
summary(seqalign)
score(seqalign)
}