Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0279739983 -0.0020918015  0.0425815589 -0.0847757482  0.0227391696 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0061394080 -0.1207328906 -0.0715040102  0.0401558089 -0.0766798847 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0610113826  0.0508748800 -0.0601825200  0.0084891047 -0.0422963281 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0746103465 -0.0293246821  0.1113546736 -0.0446722389 -0.0763393061 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0187233121  0.0971954763 -0.0446937401 -0.0224212760  0.0178701713 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.2443096245 -0.0327995421  0.0247085141  0.0455420849 -0.0599844409 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.1158157811 -0.1649275389 -0.0678219299 -0.0992152253  0.0208425242 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0554599127 -0.0230028860  0.0810874687  0.0274003695  0.0451798794 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0172033940  0.0052982739  0.0094640562  0.0174695281  0.0506851020 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0592798028 -0.0343188142 -0.0223585063  0.0380096005 -0.0345234409 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0532107714  0.0025769225  0.0036248784 -0.0667561001  0.0353872743 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.1689795898 -0.0443677436 -0.0892094141  0.0454322216 -0.0697139898 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0625245523  0.0068099737  0.1012440212 -0.0315787121  0.0853952374 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0713067033 -0.0964910811  0.0003531308  0.0094964126 -0.0670162832 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.1133411176  0.1180800434  0.0543208997  0.0489749841  0.1161446989 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0778138116 -0.0720908188  0.0008953970 -0.0474954673 -0.0707700153 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0598321304  0.0482886710 -0.0737226586 -0.0102659846 -0.0103892946 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0156778971 -0.0369018572 -0.0322006712 -0.0444842771  0.0023377265 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0214943246  0.0331837917  0.0441585954 -0.0570711485  0.0519439795 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0727446719  0.0276408234 -0.0020727563 -0.1139735786  0.0553307648 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0556795144  0.1336014187 -0.0525087837 -0.0124167939 -0.0854331990 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0162305162 -0.0292140561 -0.0906752465  0.0291508161 -0.0737884451 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0441747343  0.1155757063 -0.0897624903  0.0438788768  0.0679022928 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0764251562  0.0189355308 -0.1714441188 -0.0708539488  0.0167676633 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0567347828 -0.0681448966 -0.1243959690 -0.1827669486 -0.0816802576 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0906345850 -0.0030962579  0.0299848972 -0.0091392093 -0.0054954009 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0133079972  0.1038798577  0.0840473698  0.0461123487  0.1325136148 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0523099274  0.0675325364  0.0415156724  0.0015897482  0.1228980230 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0295981594 -0.0073617030  0.0012327513  0.1203518680 -0.0533359407 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0420835698  0.0860629730  0.0297835114 -0.0603901364  0.0896406710 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0493336066  0.0238522395  0.0507965830 -0.0194252889  0.0309990063 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0542677013  0.0095838732  0.0477333413  0.1384449062 -0.0356012715 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.0752405716 -0.0802330745 -0.0167883339 -0.0284248859 -0.0986274992 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0342861735 -0.0861844702 -0.0565018927  0.0803040158 -0.1598760627 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0554863677 -0.0958521478 -0.1079713254 -0.0160289029 -0.0357116922 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.1113567050  0.0395610296 -0.0131549893 -0.0126838959  0.0123709919 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0997995138 -0.0278157100  0.0097325259  0.0614093351  0.1820619265 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.1042121431  0.0153313160 -0.1461806977  0.0792996888  0.0567701457 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0205458646  0.0874874665 -0.0914771883  0.0427944684 -0.1502504427 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0086439534  0.1251988565 -0.0231627872 -0.0981369824 -0.0201834170 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0553606852 -0.0707889888 -0.0651256298  0.0748757688  0.0514750028 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0350225633  0.0354886941  0.1591869739  0.0135418041  0.0366907616 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0094952716 -0.0092938560 -0.0590919571 -0.0432110090  0.0690205527 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0908415331  0.0481287445  0.0604000889  0.0510147244  0.1524584939 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0296220668 -0.1070012055  0.0251715343  0.1076313631  0.0568518503 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0498794268 -0.0366486725  0.0748286578  0.0437493331 -0.0223310961 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.1209202918  0.0176231359 -0.0018744766  0.0191799246  0.0374659149 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0566067826 -0.0094126125  0.1103468915  0.0595324711 -0.0133292608 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0362931465 -0.0085282194 -0.0195490359 -0.0867427789  0.0666888225 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0296417019 -0.0758627701 -0.1041841339  0.0046445418 -0.0542958282 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0086625294 -0.0080847722 -0.0996527510 -0.0422222697  0.0667778351 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.1399234318 -0.0031708986  0.0057028104 -0.0406206689 -0.0545284516 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0154544389 -0.2086483172 -0.0742645268 -0.0575544617 -0.0155152597 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0361769074 -0.0462465330 -0.1601234590  0.0038545373  0.0189750821 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0824281499 -0.0790151984  0.0407328214 -0.0223416010 -0.0120860042 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0452924242 -0.1172849669  0.0076729929 -0.0965788869 -0.0320801253 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0988165660  0.1010159049 -0.1163365369 -0.0218262998  0.0175075624 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0818030738  0.0264671847  0.1169310786 -0.1128362292  0.0482160130 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0298988194 -0.1265763629  0.0553340634  0.0566636837 -0.0594191719 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.1234232562  0.0376038297 -0.1718348925  0.0044671392 -0.1092369690 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0213952997  0.0084096688  0.0939262456 -0.0260783931 -0.0459629716 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0066103665 -0.0918624604  0.0842582169  0.0059505807  0.0292444723 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0399312236  0.0220234237  0.1249895702 -0.0024949532 -0.0277098076 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0033865551 -0.0025830284  0.2021669104 -0.0220518597  0.1361908134 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0384247134 -0.1308702610 -0.0581848309 -0.0669550184  0.0272856712 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0956985150  0.0063788302  0.0532239249  0.0177077618 -0.0258613004 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0846375652 -0.0064074311 -0.0280843531 -0.0333598779  0.0281108710 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0583471806  0.0818137713 -0.0643450925 -0.1103478124 -0.0648928162 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.0309468992  0.1141178220  0.1282276221 -0.0302311339 -0.0529142004 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0267136575  0.0841449835  0.1010865877  0.0562574088  0.0106489601 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0259473420 -0.0631407608 -0.0463118652  0.1920954643 -0.1198512832 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.1284225675 -0.0557035834  0.1047722282  0.0142279373  0.0684155889 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0014964813  0.1029518083  0.0723876117  0.0675398783 -0.0442121450 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0376657030 -0.0063644271  0.0166409556  0.0502715111  0.0842760368 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0406664631  0.1079897638  0.0234323002  0.0041654418  0.0035901219 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.1324477608  0.0376536416 -0.1001659943 -0.0669384893 -0.0350814136 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.0115347233  0.0121103891 -0.0169107857  0.0735040052  0.1742114484 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0004629229  0.0875217192  0.0510399336  0.1180983062 -0.0185664915 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0013720929  0.1456697460 -0.0099729931  0.0465650229  0.0394721301 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0979599796  0.0351667171  0.2088817175 -0.0792237830  0.0399565660 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0255757386 -0.0412328091  0.0122676862 -0.0781092103 -0.0477525992 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0382779463  0.1499125776  0.0352315231  0.0257294416  0.0997800331 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0537406840 -0.0754250148  0.0794679197 -0.0515902044 -0.0317074024 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  0.1135685667 -0.1171336216 -0.1855849201 -0.0350909518 -0.0062422385 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0144676184  0.0196037023 -0.0141289171 -0.1081265459 -0.0326911928 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0630950275  0.0197970004 -0.0661335573  0.0777871806 -0.1266279836 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.2248061304 -0.0068864705 -0.0794946003  0.0381147473  0.1632824808 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.0589675479  0.0066580642 -0.0299274625 -0.0563274577  0.0096290548 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0074125223 -0.0403559937 -0.0600161721  0.0086831985  0.0476986124 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0704071404  0.0104676106 -0.1153824263