Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -3.076270e-02 -3.417895e-02 -9.871545e-02  9.801138e-02  1.585970e-03 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -1.109051e-01  4.443967e-02 -3.287684e-02  4.629667e-03  2.738418e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -9.015174e-03 -5.267326e-02 -4.247783e-02  1.829844e-02 -9.064261e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -5.800294e-03 -8.911462e-03 -1.441639e-01  9.745219e-02 -2.661092e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  7.206143e-02  6.452228e-02 -1.620458e-02 -3.518110e-02 -7.547698e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  9.667823e-02 -2.863324e-04 -1.396054e-01  1.712350e-03 -6.997606e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  3.562674e-02 -2.994791e-02 -2.973608e-02  1.103294e-01  4.207431e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  8.221974e-02  7.357815e-02 -1.446750e-02  1.184368e-02 -9.558821e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  2.492003e-02 -7.639390e-02 -3.987895e-04  4.598964e-02  2.454163e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  1.572256e-01  5.243928e-03  7.084985e-03  1.350412e-01 -1.667490e-03 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  1.059246e-01 -5.072349e-02 -2.079481e-02 -6.332024e-02 -7.476395e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -5.326463e-02  3.558912e-02 -4.777021e-02 -3.305663e-02  9.209430e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -9.642551e-02 -4.528046e-02 -3.406077e-02  7.192532e-02  3.669821e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  3.848668e-02 -7.848832e-02 -4.129342e-02  3.277418e-02  1.991034e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -9.666493e-02  1.983792e-02  6.623759e-02  8.463246e-02 -1.855775e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.179952e-01  3.125575e-02  8.152728e-02  1.470199e-01 -2.299890e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -2.589335e-02 -1.742951e-02  4.759269e-02  2.034850e-01  5.618569e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  2.992037e-02  5.655650e-02 -1.427289e-02 -4.981936e-02  2.406626e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  4.993367e-02  7.240899e-02  5.661677e-02 -1.356989e-01  7.454912e-03 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -1.339608e-01  3.673511e-02 -6.293494e-02 -8.416412e-04  2.023674e-01 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  6.800202e-02  1.004410e-01  2.809525e-02 -2.345120e-02  8.843853e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  1.001102e-01 -3.738481e-02  4.254549e-02  7.406632e-02 -9.100214e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  2.195037e-02 -4.263271e-02  3.252638e-02 -2.340236e-02  9.490958e-03 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  4.294209e-02 -7.413738e-02  2.222329e-03 -1.172507e-02  3.312470e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -3.872970e-02 -5.259085e-02 -3.056198e-02  6.292210e-02 -2.857985e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -5.529672e-02  3.655633e-02 -5.565878e-02 -4.345821e-02  1.572460e-01 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -8.012626e-02 -3.238278e-02  1.454963e-01 -1.186710e-02 -1.182668e-01 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -7.336687e-02 -6.944443e-03  1.002371e-01 -1.568587e-02 -2.588933e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  1.068261e-01 -1.086998e-02 -8.463093e-02  1.285550e-02  2.967956e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  7.205838e-02 -4.299819e-02  1.256890e-02 -7.475495e-02 -1.469053e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  4.436038e-02 -3.182289e-02 -6.959437e-02  1.163019e-01  4.155228e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  3.903785e-03 -1.502896e-02 -2.264501e-02 -6.433063e-02  6.538363e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  1.167445e-01  7.490553e-02  3.249334e-02  1.578427e-03 -1.130740e-01 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  1.078970e-01  1.029392e-01 -8.280580e-02 -1.006821e-01 -3.484369e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -3.806351e-03 -6.344292e-02  2.451615e-01  5.206765e-02  1.820162e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -9.753298e-02  2.282671e-03 -3.414487e-02 -2.581204e-02 -1.210539e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -6.508889e-02 -4.817417e-02  2.191227e-02  2.174255e-02 -1.554066e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  2.966583e-03 -8.997370e-02 -3.653258e-02  6.665725e-02 -4.799284e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -2.415583e-02 -1.795564e-02 -5.291002e-02 -1.043147e-01  1.946036e-01 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -3.724900e-02 -5.580243e-02  4.386787e-02  6.252917e-02 -1.369430e-01 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  1.000005e-01 -3.568594e-02 -9.573722e-03  7.198374e-02 -6.639663e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  9.552393e-02 -3.687854e-02 -3.998862e-02  2.503511e-02 -8.099104e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -1.208756e-01  1.669847e-01  3.822294e-02  8.930343e-02 -5.373180e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -1.576764e-01  1.538408e-02  4.783030e-03 -6.148078e-03 -7.962783e-03 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -1.147946e-01  2.507386e-02  2.093259e-02 -3.938994e-02  9.474398e-04 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  8.941842e-02 -7.665164e-02 -7.570740e-02 -1.127163e-01  3.522575e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.617303e-01  4.122820e-02 -4.036021e-07  2.637582e-03  1.563981e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -3.987775e-02 -6.511888e-02 -1.461789e-01 -7.570825e-02  9.977254e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -5.515305e-02 -7.019306e-02  1.594398e-02  1.112938e-01 -2.351860e-03 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -7.936725e-03  9.314956e-02 -4.869003e-02  1.664279e-01  3.525773e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  8.394705e-02 -7.154352e-02  9.293344e-02 -9.356564e-02 -7.241245e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  7.606406e-02 -4.679135e-02  4.366612e-02 -4.386647e-02 -1.274887e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  3.903345e-02  5.082549e-02  9.521588e-03  1.553075e-01  2.707587e-03 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -2.582173e-02 -5.707363e-02 -4.513647e-02 -1.590857e-01  1.004647e-01 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -1.856515e-01  1.057353e-01 -4.357120e-02 -2.067433e-02  9.554538e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -4.012810e-02  8.031409e-02 -5.660641e-02  3.429622e-02 -6.875869e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  4.054744e-02 -2.767843e-02 -3.060547e-02  3.692081e-02  1.497327e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  9.309050e-02 -6.104312e-04  2.638385e-02 -1.014036e-01  1.089374e-01 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -4.333109e-02  1.153654e-01  5.052171e-02 -2.589516e-02 -1.353830e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  7.665507e-03 -1.108214e-01 -4.260460e-02  7.778423e-03  5.505475e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -6.829848e-02 -3.362908e-02 -4.336870e-02  6.226195e-02 -3.018807e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -4.619469e-02  1.117819e-01 -3.631563e-02  2.291139e-02 -8.355142e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -3.352369e-02 -7.746660e-02  4.625887e-02 -1.600779e-02  1.115890e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -2.605137e-03 -6.100831e-02  5.101146e-02 -8.295171e-02 -6.503908e-03 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  8.433416e-02 -3.072503e-04  1.706902e-02 -1.242878e-01  5.174825e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -8.226412e-02 -7.376616e-02 -2.927121e-02 -6.854412e-02  3.376775e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  2.794607e-03  1.116769e-01  6.997334e-02 -7.801010e-03  6.695885e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -1.642659e-02  1.276285e-01  1.690857e-01 -1.080874e-02 -7.317670e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  5.314039e-02  1.054054e-01  6.897347e-02  4.096726e-02  3.737950e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  9.049313e-02 -3.759594e-02  1.071546e-01 -1.026466e-01  2.274073e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  5.001845e-02 -3.646443e-02  1.051612e-01 -5.623500e-02  1.311815e-01 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  1.020758e-01  9.142932e-02 -6.008339e-02  3.261402e-02 -1.815687e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -3.421003e-02 -1.227536e-01 -1.009777e-01  2.510088e-02  2.768933e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  9.880233e-03 -2.255666e-02 -1.242938e-02  1.155967e-02 -5.270538e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -3.539830e-02  1.872501e-02  7.873646e-02 -2.344594e-02  8.611980e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  9.839401e-03 -4.898463e-02 -3.963498e-02 -5.902504e-02  1.775046e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -3.447197e-02  3.603933e-02  4.000935e-02  9.054628e-02 -2.458606e-03 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  8.413845e-02 -1.309537e-01 -3.352114e-02  2.521707e-02  5.180796e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  1.933889e-02  5.352188e-02  1.186777e-01  8.400953e-03  5.481151e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  2.607412e-02 -5.785851e-02  6.653860e-02  3.424506e-02 -2.661531e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  6.981742e-02  6.612456e-02  4.585749e-02  5.980682e-02  1.095215e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -1.815440e-01  9.821366e-02 -1.176907e-01  7.957340e-02  6.710767e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  3.908923e-02 -5.857399e-02  4.256124e-02 -9.823899e-02  2.091321e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -1.417451e-01 -2.240833e-03 -7.356138e-03 -5.293989e-02 -1.950335e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  1.411112e-01  4.377462e-03  1.198883e-02  9.767136e-02 -3.597915e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  1.764095e-02  8.476731e-02  2.953837e-02  1.054545e-02  9.265804e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -1.348711e-03 -4.226632e-02  5.630336e-02  2.034749e-02  2.375899e-03 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  5.858381e-02  3.609494e-02 -3.194872e-02  8.808775e-02  5.411586e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  2.668889e-02  5.175655e-02 -8.191358e-03 -1.755356e-02  3.717070e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -1.488084e-02 -1.568682e-02 -5.763566e-03