Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  5.910468e-02  6.718357e-02 -1.766928e-01  6.114378e-02  8.212752e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  1.296977e-02 -1.755218e-01 -5.537243e-02 -2.216321e-02  1.147852e-01 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -8.908454e-02  1.985432e-02  1.843468e-02  7.063190e-02  2.467995e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -1.035447e-01 -5.004724e-02 -4.727554e-04  2.725259e-02 -1.475443e-01 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  1.106059e-01  1.540912e-01 -8.837844e-03  9.471980e-02 -5.256170e-03 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  5.410691e-02  9.720836e-02 -6.096003e-02 -3.452499e-02 -3.838370e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  3.356054e-03 -7.600903e-02  7.371590e-02  1.760295e-01 -5.013086e-03 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  1.067968e-01 -7.196902e-03 -1.768589e-02 -4.876523e-02  1.875494e-01 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -3.879383e-02 -7.245989e-02 -7.983472e-02 -9.924824e-02  3.746081e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -6.471746e-03  1.656006e-01  1.518155e-01 -2.380007e-02  2.525675e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -6.752475e-03 -8.107497e-02  8.690987e-03 -3.148500e-02  1.188253e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -4.586872e-02 -4.935517e-02  1.333292e-02  3.837351e-03  1.562464e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  2.882795e-02 -7.885119e-02 -7.673055e-02  5.402816e-02 -3.490299e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  4.875986e-02  9.135939e-02 -8.984449e-02  4.505268e-02  1.598431e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -5.605141e-02 -8.569721e-03  1.293089e-02 -3.276800e-02  2.299937e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  9.377272e-02 -2.425016e-02  1.141534e-02 -1.039525e-01 -9.600584e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  2.039584e-02 -4.247679e-02  1.644952e-01 -2.766378e-02  2.786514e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -6.892343e-03  6.210780e-02 -1.882354e-02  7.304571e-02 -5.501095e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  2.350080e-02  5.568984e-04 -1.250720e-04 -6.018433e-02 -1.156896e-04 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -5.065191e-02 -2.372770e-02  2.433751e-02  8.609002e-02  4.494221e-03 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  2.288987e-02 -6.684415e-02  6.881570e-02  9.870925e-02 -6.736992e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -5.392220e-02 -4.011948e-02  2.480936e-02  3.527815e-02  9.858552e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -5.441891e-02  3.314716e-02  6.882581e-02 -6.427711e-02  7.461550e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  4.262263e-02  1.054285e-01  3.857759e-02 -3.576829e-02  7.375120e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -7.433969e-02 -5.119764e-02 -4.130480e-02 -1.599343e-02  2.058974e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -1.452296e-02  1.462266e-03 -7.127797e-02  1.826784e-02 -7.276753e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  2.500193e-02 -2.372198e-02  2.800575e-02 -9.100556e-02 -1.155296e-01 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -7.735018e-02  1.046721e-01  7.643483e-02 -1.286898e-01  3.130385e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -2.491981e-02  5.382481e-02 -6.020407e-02 -1.398038e-01  1.188000e-01 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -1.759565e-02  1.032029e-02 -1.448665e-02  1.572283e-02 -8.277707e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  8.639596e-02 -3.607698e-02 -7.764100e-02  4.194898e-02  8.390340e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  6.914584e-02 -5.168028e-03 -1.106141e-01 -8.246308e-02 -5.600653e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  1.323259e-02 -1.601635e-01 -2.868925e-02 -5.776616e-02 -8.717065e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -4.900912e-02  1.050505e-01  8.540155e-02  1.034893e-01  1.412701e-01 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -2.154025e-02  3.120152e-02  9.310038e-02 -1.533233e-01  7.359293e-03 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  1.124893e-01  1.126808e-02 -2.437809e-02  1.567290e-01  5.573403e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -1.427875e-01  1.371851e-01  8.802636e-02  2.538143e-02 -3.443330e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -2.139498e-01 -6.043041e-02  2.598998e-02 -1.101050e-01  1.219678e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -7.852887e-02  9.799069e-03  7.751918e-02 -1.046714e-01  6.536448e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -1.064953e-01  5.492463e-02 -5.569961e-02 -8.776832e-02 -7.034295e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  8.802876e-02 -9.937372e-03 -1.192615e-01 -4.529161e-02 -5.125217e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -1.342284e-01  3.937738e-02 -6.907521e-02 -1.324720e-01  2.400353e-01 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -5.833794e-02 -1.344625e-01  3.429581e-02 -1.068237e-01 -1.170179e-01 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  3.472338e-02 -7.295565e-02 -6.459467e-02 -1.717250e-02  5.218004e-03 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  8.636907e-02 -7.176141e-02 -1.165322e-01 -4.577984e-02  1.015259e-01 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -6.607934e-02  5.522451e-02 -2.061207e-03  3.993006e-03  1.003201e-01 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.154283e-01 -2.096781e-02 -4.441242e-02 -8.762846e-03 -8.573838e-03 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  1.342395e-02  5.645682e-02 -2.661926e-02 -4.301592e-02 -4.205610e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  2.255262e-02  1.061639e-01  4.443121e-02 -1.156213e-02  3.239053e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  9.514149e-03 -6.154561e-02 -5.640086e-02  2.653141e-02 -6.617489e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  4.455371e-02  3.519655e-02 -1.036318e-01 -7.953066e-02 -8.283881e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -8.270215e-02  1.218050e-02  1.297837e-01  3.579854e-03 -3.148657e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  8.812370e-02 -6.803803e-02  5.994602e-02  6.185621e-02 -9.665658e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -4.651302e-02 -7.177493e-02  2.590649e-02 -1.110369e-02 -2.673174e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -3.041765e-02 -1.586035e-02  1.020505e-01  3.786558e-02  3.962505e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  1.062483e-01  1.346318e-02 -1.816163e-02 -3.809442e-02 -1.168476e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  1.813317e-03  3.742594e-02 -4.218616e-02 -2.874285e-04 -7.455399e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  6.883807e-02  2.423542e-02  5.377856e-02  5.755552e-02 -2.048938e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -3.911661e-02  8.775948e-02 -6.908130e-02  5.976090e-02 -5.593888e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -1.222045e-02  1.409815e-01 -7.964224e-02 -7.136837e-02  8.617263e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -6.106854e-02 -4.440663e-02 -1.262959e-02  8.659459e-02  4.380018e-03 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  1.275670e-01  2.344510e-03  5.316152e-02 -6.009412e-04 -4.461646e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  1.209990e-01  1.608273e-02 -5.337337e-03 -1.195712e-01  1.368145e-01 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -1.711680e-03  1.089939e-01  8.725418e-03  2.058222e-02 -3.945181e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  4.155264e-02  2.512661e-02 -1.758073e-02  5.418672e-02 -7.109975e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -6.785241e-02  7.307491e-02 -1.084777e-01 -1.507997e-03  2.782866e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -4.011990e-02 -4.602415e-02  1.525622e-01  3.291026e-02  7.430676e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -2.526091e-03  2.915736e-02 -1.282211e-02 -8.667517e-02 -9.266755e-03 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  5.469480e-06 -3.680138e-03  4.215302e-02 -8.172351e-02  4.245129e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  4.305394e-02 -8.393199e-03 -1.184987e-01 -2.556147e-02 -4.580705e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -1.592811e-01  1.158620e-01  4.716136e-02 -3.965575e-02  9.175543e-03 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -2.136531e-01  1.042685e-02  1.281699e-01  4.336539e-02 -1.629719e-01 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  8.793420e-02 -4.907590e-04  9.948001e-02 -5.811253e-02  7.079095e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -1.656401e-02 -3.020762e-02  1.199008e-01  8.244611e-02 -7.584256e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  1.232806e-01 -9.972887e-02 -4.525895e-02 -8.185671e-02 -2.215006e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  1.075883e-01 -5.386439e-03  8.757156e-02  5.483680e-04 -2.617881e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  3.421626e-02 -6.678683e-02  7.928139e-02  4.761266e-03  8.743964e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -1.277337e-01  7.806273e-03 -7.279454e-02  1.276564e-01 -2.186068e-03 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  6.893094e-03  1.351755e-01 -5.417870e-02  9.150295e-02 -1.549159e-01 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -2.355543e-02  1.163630e-01 -2.208166e-02  3.490949e-02 -5.012293e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  8.226985e-02  1.284246e-01  4.202435e-02  2.380457e-02  1.103078e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  2.058872e-02  8.053060e-03 -7.813297e-02 -1.013511e-01 -8.575051e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -1.684663e-01  7.410583e-02 -5.156458e-02 -5.247318e-02 -7.027513e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -1.546998e-02  4.991111e-02  8.101432e-03 -3.001381e-02 -5.313644e-03 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  6.184960e-02 -2.782808e-02  1.033886e-01 -5.122207e-02 -3.824804e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  5.859106e-02 -5.642325e-02  1.823741e-04  3.390229e-03 -8.098962e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -5.094283e-02 -1.546727e-02 -1.350017e-02 -4.945962e-02 -3.115226e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -2.244115e-03  1.398373e-01  1.718681e-01  4.129949e-02 -6.093187e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  3.793370e-02  7.041331e-03 -4.735605e-02 -1.634578e-03  9.628023e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  4.563270e-03 -6.647458e-03 -1.700919e-01