Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  3.022723e-02 -1.661893e-02 -8.245823e-02  1.179014e-01  2.066570e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -1.095478e-01  3.909496e-02  7.934174e-02 -4.944818e-02  9.313482e-03 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  3.674530e-02  5.200789e-02 -7.629650e-02 -5.787919e-02 -9.480367e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -5.729934e-02 -5.414480e-02 -8.702492e-03 -9.750665e-02  6.451396e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -6.905945e-02  2.928160e-02  6.673064e-03 -1.143110e-01 -2.381622e-03 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -1.760713e-02  7.250322e-02  1.635026e-01 -1.945236e-02 -1.950175e-01 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -1.681022e-02 -9.504882e-02 -9.220764e-02 -7.049151e-02 -1.415687e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  7.425505e-03 -5.396145e-02  2.258128e-02  3.698350e-02 -4.631336e-03 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -1.452111e-03 -9.888752e-03 -5.496403e-02  4.924086e-02 -4.122663e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  3.893504e-02 -6.821656e-02 -6.669971e-02  1.993395e-02  1.016769e-01 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  7.531629e-02 -2.901597e-02  1.955489e-02  8.705623e-02 -9.846029e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -2.238674e-02  7.029554e-03  1.061255e-02  2.477233e-02  1.721409e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -1.222020e-02  3.489548e-02  1.654687e-02  4.673076e-02 -1.113663e-01 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -1.510970e-01  6.468112e-03  9.692606e-02 -7.480213e-02 -3.878376e-03 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -2.084190e-02 -5.073679e-02  5.098778e-02  7.499400e-02 -1.299282e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.466123e-02 -1.482264e-01  7.687079e-02  7.087566e-02 -1.994806e-01 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  4.959273e-03  2.733517e-03 -8.558882e-02 -1.678026e-01 -9.015260e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -3.828835e-02 -3.584088e-02 -4.555706e-02 -4.743195e-02  4.606698e-03 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -1.034350e-01  4.407632e-02  2.520927e-02  2.573890e-02 -6.071100e-03 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  3.645520e-02 -4.282425e-02  4.572766e-02  1.027139e-01 -1.816406e-01 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  5.899635e-02 -3.612380e-02  8.873944e-02  6.704358e-02  1.909514e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  1.324086e-02 -1.047547e-01 -9.870676e-02 -1.192866e-01  7.457524e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -2.564272e-02  3.220177e-03  3.370511e-02 -4.803756e-02  1.372075e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -5.101721e-05 -8.762828e-02  9.134057e-02  6.422923e-02 -7.715539e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -9.300586e-03 -2.377876e-02  6.928918e-02  1.684593e-01 -3.128952e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  7.560442e-02 -1.023264e-01 -6.545496e-02  5.218861e-02  1.651803e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -5.600443e-02  7.686457e-02 -9.377187e-02  1.421806e-01  2.798538e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -5.740614e-02  2.285967e-02 -1.021859e-01  7.466063e-02 -7.691594e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -5.402427e-02 -1.588992e-02 -5.404087e-02  8.397899e-02  1.111134e-01 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  6.019436e-02 -1.785431e-01  3.678670e-02 -1.134512e-02 -1.009602e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -1.083807e-01  1.436431e-01  5.451480e-02  5.319865e-02 -9.142936e-03 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -5.037995e-02 -1.440543e-01 -4.411671e-02 -1.332928e-01 -9.225530e-04 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -8.013739e-02  9.109114e-03 -1.865675e-02 -6.895976e-03 -8.893030e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  1.236931e-01  3.272533e-02  7.133135e-03  9.826413e-02  3.000761e-04 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -2.951628e-02 -1.120841e-02 -8.727655e-02  7.723761e-02 -4.553155e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -2.825431e-02  3.014062e-02  5.347734e-03  1.393984e-01  5.001615e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -2.329337e-02 -3.674307e-02  5.516594e-02  8.760011e-02  8.216725e-03 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -1.406280e-01 -2.281303e-02 -6.579588e-02  1.045030e-01  1.579505e-01 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -1.289694e-02 -4.351629e-02  7.259392e-02 -2.482958e-02  3.946894e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -6.724259e-02  1.705407e-01 -1.082265e-01  4.306083e-04 -1.535584e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  9.007582e-02  4.722978e-03 -2.844441e-02  1.370860e-01 -2.373794e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -2.383361e-02  1.045138e-01 -6.540888e-02 -7.885386e-02  2.593970e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  7.883700e-02  1.036132e-01 -3.345443e-02 -8.741785e-02  3.955204e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -6.209461e-02  6.896979e-03  4.490194e-02  8.412467e-02  4.198646e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  9.798159e-02 -4.192350e-02 -3.128446e-02  6.406174e-02 -1.303238e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  4.361143e-02 -2.593768e-02  5.309227e-03  1.182856e-01 -5.396436e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  1.923942e-02 -2.195860e-01  7.747568e-03 -1.227701e-02 -2.257060e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -1.655514e-01 -1.656159e-01 -7.397195e-03  2.245454e-05 -7.694573e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  8.475752e-02  4.093335e-02 -3.204473e-03 -9.332004e-02  1.014701e-01 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  4.865661e-02  3.497791e-02 -3.194869e-02  6.477551e-02  6.361523e-03 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  6.158477e-02  5.500474e-02 -1.470611e-02 -2.636683e-02 -5.057170e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -4.273574e-02  1.127419e-02  7.741297e-02  3.181640e-02 -1.640683e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -7.112339e-02  1.139003e-01  8.377577e-02 -1.123085e-02  1.932015e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  6.132613e-03 -1.502967e-02  1.008462e-01  3.326558e-02  8.251729e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  1.581833e-03 -4.215678e-02  4.697021e-02 -3.628630e-03 -1.149569e-01 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  8.672618e-02  1.133848e-01 -1.074463e-02  2.627020e-02 -7.968206e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -6.769922e-02 -2.010594e-02  1.063009e-01 -9.144088e-02 -3.430746e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  1.348381e-01 -4.639399e-02  3.011490e-02 -3.081994e-02 -2.629230e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  7.035349e-02  3.498164e-02 -1.484491e-02 -2.920351e-02  4.042018e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -5.441559e-03 -2.852581e-03  8.232178e-03  7.003312e-02  1.433149e-01 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  7.401278e-02 -8.145281e-02 -1.248504e-02 -1.169677e-01  7.750822e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  4.821591e-02  7.574092e-02 -1.982434e-02  1.686688e-03 -5.150162e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -9.772513e-02  4.409704e-02  8.395378e-02  2.562913e-02  2.602075e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -2.008485e-02 -4.634686e-02  1.248445e-02  6.736573e-02  8.897501e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -3.187004e-03  9.548727e-02 -1.752107e-02 -2.824501e-02  9.834135e-03 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  1.782183e-02  6.892958e-02 -4.106489e-02  6.611034e-02  5.129735e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -8.924910e-02 -6.872401e-02  3.874089e-02 -1.762674e-02  8.947776e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  2.103929e-02 -1.344924e-01  1.458583e-04  2.406366e-02 -4.939674e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  4.116900e-02 -7.083625e-02 -1.168045e-01  5.239434e-02  6.207620e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  3.992778e-02 -2.617493e-01  3.460001e-02  2.658414e-02  5.884367e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -3.883041e-02  6.956564e-02 -4.199425e-03 -5.222253e-02  1.220847e-01 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  1.520070e-01 -2.163180e-02 -8.711582e-02  5.276865e-02 -5.713439e-03 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  1.038261e-01  1.439985e-01 -9.556126e-02 -1.528153e-02  1.105228e-01 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  2.352358e-02 -7.204671e-02 -4.948779e-02  4.569367e-02 -6.607223e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -7.232468e-02 -4.901579e-02  3.261335e-02 -2.384089e-03  9.052760e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  5.914941e-02 -6.653316e-02 -9.197030e-03 -9.048575e-04  6.913575e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -1.458481e-03  1.622719e-02 -1.187117e-01 -2.562135e-02  1.023016e-01 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  9.917215e-02  5.477991e-02 -4.043543e-02 -2.718816e-02 -5.300728e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -3.699902e-02  9.466140e-02  2.333647e-02  4.291717e-02 -2.329504e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -1.312902e-01 -5.744492e-02 -7.817905e-02  1.374108e-02  1.257196e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  5.437291e-02  4.867222e-02 -9.754924e-03 -1.556867e-01  1.611624e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -1.215956e-02 -1.382743e-02  6.602806e-02  8.510224e-03 -6.883371e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  1.172210e-01  5.638756e-02 -7.317409e-03  1.208545e-01 -6.884179e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  4.139808e-02 -4.481704e-02  8.519938e-02 -5.083453e-03  2.677748e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  6.862280e-02 -1.148291e-03 -1.108618e-01 -4.790569e-02 -6.591725e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  5.525031e-02  6.485708e-02 -2.873261e-02 -1.186686e-01 -7.153383e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  5.882880e-03 -2.384800e-04  2.382220e-02 -2.319576e-02  9.079706e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -2.159766e-02  7.254143e-02  8.620997e-02 -1.112118e-01 -2.582954e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  9.861681e-02  6.268136e-02 -4.812472e-02  2.077176e-04 -7.081708e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -4.284679e-02  1.144860e-01 -8.423173e-02