Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sjorstrom, M., Sandberg, M., & Rannar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sjostrom, M., Rannar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -2.378154e-02  1.616346e-02  1.510745e-01  1.350682e-02  6.507510e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -3.854147e-03 -6.966810e-02  6.461185e-03  7.120816e-02 -4.905284e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  8.496775e-02 -6.661472e-02  3.036421e-02  4.906345e-02 -9.423086e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -1.500548e-01  1.609746e-03  3.306131e-02 -1.217777e-01  5.309548e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  3.264065e-02  4.172285e-02  7.718406e-02  4.339640e-02  1.131529e-01 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  7.450966e-02 -6.060084e-02  3.433017e-02  4.304686e-02  3.423741e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -4.318474e-02  7.559752e-02  8.378399e-03 -4.037241e-02 -4.129359e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  5.751480e-03 -9.607509e-02 -1.437869e-01 -7.830721e-02 -2.433023e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  1.247346e-02  1.060447e-02 -1.152719e-01 -2.713255e-02 -7.561834e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  2.673118e-04  2.173993e-01  8.437089e-02  9.736402e-02  2.622468e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -5.104863e-02 -1.638899e-01 -5.603456e-02  4.218483e-02 -1.825385e-01 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  3.352544e-02  1.251674e-01  3.105282e-02 -6.349298e-02 -2.513566e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -7.851790e-02  9.744577e-03 -2.688466e-02 -2.620110e-02 -1.014002e-01 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  2.565637e-02  9.835211e-02  6.847551e-02 -4.339471e-02 -9.746491e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -1.156611e-01 -7.386885e-02  3.491987e-02  1.185141e-01 -2.516182e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  1.213352e-02  2.525860e-02 -1.273275e-01  4.169518e-03 -3.454842e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  9.832859e-03 -5.444787e-02 -1.637631e-02 -4.079656e-02  9.135960e-03 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -3.278657e-02 -4.661392e-02  6.702765e-02 -2.212424e-02 -2.918723e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  8.390456e-02  2.809721e-02  8.963361e-02 -5.509758e-02 -5.398185e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  3.796264e-02 -4.177923e-02 -1.199491e-02 -2.530341e-05 -7.833759e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  9.129335e-02  9.419866e-02  2.884141e-02 -6.222774e-02 -3.806628e-03 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  2.939685e-02 -7.653849e-02  9.193155e-02 -7.232301e-03 -2.329555e-03 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -1.364652e-01  9.693763e-03 -4.942307e-03 -3.230371e-02 -8.517994e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  5.665605e-02  9.901049e-03  1.043796e-01 -3.183322e-02  7.150624e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  2.048636e-02 -8.103480e-02  1.729519e-02 -4.785343e-02  3.837429e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  6.186973e-02  5.856284e-02 -6.452601e-02  3.300800e-02  1.895021e-01 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -5.073301e-02  3.882350e-02  1.021021e-02 -6.186309e-02 -5.515414e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  7.823002e-02  2.416030e-02 -7.296051e-02  4.976617e-02 -1.446622e-01 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -1.822817e-02  3.731089e-02 -1.613876e-01 -5.208498e-02 -5.244953e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -4.252773e-02  1.717882e-02  1.122052e-02  1.182869e-01 -7.252452e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  2.599955e-03  6.839116e-02 -3.002915e-02  1.954154e-03  2.758980e-03 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -6.022017e-02  1.587562e-03 -4.825403e-03  1.051635e-03  1.029889e-01 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -1.128060e-01  2.326098e-02 -2.438107e-02  9.076277e-02 -7.564512e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -2.055635e-01  6.765535e-03 -4.485009e-02 -5.571658e-02 -4.564079e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  9.496626e-02 -1.212035e-02  3.050101e-02  1.407273e-02 -5.736403e-03 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -2.399361e-02  3.450500e-02 -8.541520e-02  6.515029e-02  4.010890e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -3.029441e-02 -8.036140e-02 -6.076005e-03 -1.072572e-02 -3.440876e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -8.959166e-02 -4.118370e-02  6.329688e-02  6.013995e-02 -4.896015e-03 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -4.573840e-02 -9.476172e-02  9.460238e-02 -2.056061e-02  4.835954e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  1.441791e-01  2.854042e-02 -4.102382e-02  1.018251e-02 -6.322645e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  1.391465e-02 -3.246863e-02  1.158064e-04 -4.245238e-02  1.478706e-01 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  3.157529e-02 -5.352600e-02 -1.337289e-01 -3.760267e-02  1.332130e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  4.848053e-02 -7.004088e-02  7.100465e-02  3.361856e-02  2.206772e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  1.005220e-01 -4.170384e-02  1.016096e-02  5.831325e-02  8.239771e-03 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  8.078889e-03 -5.870504e-02  1.460870e-01  1.579320e-01  8.755741e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -5.317744e-04  4.312650e-02 -5.438589e-02  1.165579e-01 -1.542687e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.580723e-01 -2.014356e-02 -8.594046e-03  5.848151e-02 -5.530597e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -4.476934e-02 -6.854218e-02 -1.246404e-02  7.924143e-02  2.013215e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -3.447620e-02  3.216773e-02 -5.164674e-02 -4.580822e-02 -1.949547e-01 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -3.718600e-02 -4.849791e-02 -3.480485e-02  8.937212e-02 -1.140616e-01 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  7.307441e-02  6.441852e-02 -4.040068e-02 -9.431480e-02 -1.408371e-01 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  7.658464e-02  1.075126e-01  6.380517e-02 -2.640063e-02  4.150709e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -5.152054e-04 -6.482486e-02 -1.284722e-01 -4.064445e-02  8.409306e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  1.732456e-02 -5.231727e-02  6.095155e-02  2.043132e-01 -3.536678e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  2.687027e-02 -5.669859e-02 -3.126496e-02 -2.961025e-02 -8.660060e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  1.089242e-02 -5.473348e-02 -4.471085e-02 -6.048187e-02 -5.326678e-03 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -7.278730e-02 -8.327736e-02  1.220281e-01  2.877014e-02 -7.774233e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -4.059473e-03 -1.173901e-02 -1.480329e-01 -5.737306e-02  3.150179e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  8.128701e-03 -3.179892e-02 -9.008551e-02  2.411676e-02  1.377879e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  1.239610e-01  5.549966e-02  4.118724e-02 -1.159212e-01 -2.637503e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -1.042280e-01  3.768165e-03 -7.554085e-02  9.897019e-02  6.426452e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  9.019073e-02 -8.580756e-02 -2.494138e-02  2.320902e-02 -2.904855e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  3.690824e-02 -2.866734e-02  3.872857e-02 -6.004093e-02 -3.872463e-03 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -5.538278e-02 -2.313984e-03 -2.557960e-03 -9.792253e-03  1.208574e-01 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  3.755430e-02  6.308323e-02  4.960751e-02  1.549349e-01 -3.861740e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -3.480056e-02  1.132809e-01  8.468269e-02  7.429328e-02  5.637465e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -5.176942e-03  4.727698e-02 -1.514655e-02 -8.457050e-02  1.400117e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -2.676894e-02  1.441938e-02 -3.306125e-02  2.185860e-02  4.779100e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -1.151416e-01  6.218970e-02 -5.122380e-02 -9.254738e-02  1.144143e-01 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  8.570035e-02 -8.171576e-02 -5.409420e-02 -1.526252e-01 -9.935035e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -3.987092e-02  9.281755e-02  8.133392e-02  9.267195e-02 -1.041408e-01 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  8.100290e-02  3.126185e-02  6.151736e-02  3.049130e-02 -1.552932e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -1.260855e-01 -6.826473e-02  1.447744e-01 -7.715543e-02 -2.154964e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -6.878121e-02  5.953547e-02 -5.982693e-02  7.331432e-03 -2.507333e-03 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -2.183105e-02  1.863070e-04  6.350171e-02 -3.102340e-02  2.933859e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -7.984955e-02 -3.098728e-02  9.698753e-02 -8.640628e-03 -1.838152e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  1.246361e-02  7.522712e-02  2.610559e-02 -9.735075e-02 -1.040327e-01 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  2.273928e-02 -9.090985e-02  6.474260e-02 -8.561166e-02 -6.785685e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -5.084894e-02  6.182155e-02  5.601282e-02  4.486735e-02  3.037279e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  8.074031e-02 -4.303355e-02 -1.259293e-01 -1.131793e-01 -9.990141e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -1.802156e-01  1.114590e-01  3.839938e-02 -9.238271e-02 -1.704487e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -4.309987e-02 -5.367593e-02  6.528600e-02 -8.639008e-03 -2.691733e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  1.491798e-01  2.556034e-02 -3.159239e-02 -1.188435e-01 -5.025763e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  8.549287e-02  1.649911e-02  5.279356e-02  2.701088e-02 -9.600010e-03 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -1.944061e-02 -8.306112e-02  5.817218e-03  1.072555e-01  1.796041e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -9.168665e-02  3.524826e-02 -1.174500e-01 -1.473749e-02  1.233289e-01 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -5.869576e-02  2.394643e-02  3.415500e-02 -4.488363e-02 -2.090670e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -5.753371e-02 -7.076470e-02  1.554258e-01 -6.918480e-02 -5.098261e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -9.269911e-02 -9.932426e-02 -1.847403e-02  2.049421e-02 -6.056141e-03 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -6.971940e-03  3.569535e-02  3.970636e-02