Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -1.083627e-01 -6.605681e-02 -6.911017e-02 -3.504486e-02  2.527860e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  1.239368e-01 -7.832865e-03  2.020495e-02 -1.724648e-01 -1.033924e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  4.675826e-02  9.159451e-02  1.048439e-01 -3.210039e-02 -2.051161e-04 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  5.669749e-03  7.436229e-02 -2.729485e-02  3.804478e-02 -2.168968e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  7.674347e-02 -7.819477e-02  4.748771e-02 -7.574010e-02 -3.493939e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  8.432080e-02 -2.043381e-02 -8.280438e-02 -1.746947e-02 -2.435434e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -2.922370e-02  1.673826e-02  1.287036e-01 -9.680082e-03 -9.330285e-03 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  5.565057e-02 -5.050702e-02 -1.249763e-01  4.061697e-03 -8.461602e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -9.791467e-02  7.757059e-02 -1.038124e-01  5.912426e-03  8.747041e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -5.459499e-02 -1.475555e-01 -1.063825e-01  1.938361e-02  6.433693e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -1.129780e-01  1.383441e-02  1.073835e-01 -4.241973e-02  7.184500e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -2.398069e-02 -4.621197e-02  9.775088e-02  5.929294e-02  7.871746e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -4.709195e-02 -2.106908e-02  1.117495e-01 -6.829683e-02 -1.569702e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -3.681313e-02  1.553285e-02 -2.042316e-02  2.034599e-02 -1.548758e-01 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  1.124628e-02  3.052897e-02  1.831689e-04  1.002044e-02 -4.785174e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  1.217042e-01  4.488219e-02  2.430203e-03 -2.674341e-02  5.932137e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -1.263688e-01  2.504096e-02 -2.376351e-02 -6.384701e-02 -2.363321e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -1.208421e-01  1.891320e-02  5.535640e-02  5.860247e-02  3.516008e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -3.514072e-02  5.422793e-02 -1.006837e-02 -1.819933e-02  6.663726e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -4.018131e-02  1.941510e-01 -5.582672e-02 -3.544676e-02 -2.028582e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  1.184397e-01 -6.930834e-02 -7.812226e-02 -7.165421e-03  4.751210e-03 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -4.946782e-02 -1.019449e-01 -6.514896e-02 -1.291145e-01  9.821167e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -1.081415e-01 -4.485736e-02  6.113036e-02 -1.347698e-02  5.427015e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -3.934057e-02  1.030089e-01  4.025260e-02  8.911593e-02 -3.226668e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  6.811233e-02  1.113213e-01  2.492096e-02 -3.291060e-02 -6.181285e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  3.381806e-02 -1.321044e-01 -8.647244e-02 -7.512364e-02 -1.059255e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  1.250819e-02 -4.343611e-04  5.146473e-02  1.370273e-02 -6.928878e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  4.113624e-02 -6.859962e-02  1.085280e-01  9.057208e-02 -4.208866e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -2.398327e-02  7.318470e-02  7.491949e-02  1.130303e-01 -1.713429e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  7.939652e-02 -2.384672e-02 -4.414849e-02  1.662026e-01 -8.939992e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -9.879026e-02  1.075108e-02  6.869992e-02  4.251907e-02 -8.585816e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  1.177466e-01 -2.074682e-02  8.784276e-02  1.155013e-02 -2.886709e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -6.837014e-02 -3.473822e-02  4.479872e-02  4.306376e-02 -4.292008e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  5.993506e-02 -5.781335e-02  1.094300e-01  6.850047e-02 -2.719415e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -2.377780e-02  2.270219e-02  4.083585e-02  3.338501e-02  5.003374e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  8.530511e-03 -8.785271e-02  4.532275e-02 -9.831274e-02 -1.695853e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  2.245979e-02 -2.949502e-02  2.018817e-02 -8.487896e-02  2.912942e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  8.313683e-03  3.556010e-02 -9.839437e-03  1.180123e-01  1.587406e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  2.400365e-02  4.022118e-03 -3.377820e-02 -3.608671e-02 -1.433669e-01 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  8.800742e-02  1.412583e-01 -6.370933e-03  2.374985e-02 -8.577099e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  4.243349e-02 -6.340585e-02  2.125480e-02 -6.036670e-02 -1.727112e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  3.988322e-02 -4.575164e-02 -6.826103e-02 -1.656695e-03  3.826360e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  9.472449e-02  6.091600e-02 -3.217487e-02  6.687937e-04  2.583228e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -6.073908e-03 -1.395524e-03  7.183396e-02  5.801118e-02  1.921115e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  9.092684e-02 -1.172102e-01 -1.446832e-01  5.333569e-02  6.553052e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  5.708385e-02  7.559207e-02  2.288646e-02  1.219164e-02 -4.879584e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.018936e-01  5.999563e-03  1.020778e-01  6.441365e-02 -1.852354e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -5.522752e-02  3.676240e-02  3.566168e-02 -1.478165e-02 -5.797687e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -4.936382e-02 -2.162711e-02 -2.105850e-02 -4.987982e-02 -3.364776e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  1.167226e-01  4.173630e-02 -4.549374e-02  4.290125e-02 -4.168860e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  9.034659e-02 -9.533333e-02 -4.706265e-02  8.854051e-02  1.132422e-01 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  1.180976e-02  1.582603e-01 -2.291959e-02 -4.268594e-02  3.232677e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -3.990098e-02  2.748326e-02 -1.040853e-01 -1.207910e-01 -4.901538e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  7.545526e-02 -5.892197e-02  2.396781e-02 -3.561219e-02 -6.903978e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -1.197378e-01 -1.151602e-01 -3.361562e-02 -1.212592e-01  9.152463e-03 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  2.393163e-02 -4.352749e-02 -1.042422e-01  3.675775e-02 -1.140409e-01 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  2.741638e-02 -1.185668e-01 -1.326981e-01 -3.666414e-02 -7.369416e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  6.125882e-02 -1.598337e-01 -4.512534e-02 -1.578909e-02  4.601404e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  9.513022e-02 -1.502781e-01 -2.436304e-01 -1.312370e-02  3.797205e-03 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  2.798518e-02  5.511619e-02  4.250490e-02  3.560953e-02 -3.871743e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  5.534292e-02  3.386890e-02  2.002782e-02 -1.109030e-01 -1.165128e-01 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  5.877493e-02  6.907324e-02 -6.467029e-03 -2.659088e-02  1.235278e-03 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -1.030433e-01 -2.891840e-02  1.067866e-01  4.331153e-02 -1.627451e-01 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  6.674899e-02  9.271419e-02 -3.379320e-03 -4.581849e-02 -1.901182e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -7.966554e-02  3.414431e-02 -1.248936e-02  2.251419e-02  1.349422e-01 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -5.803156e-02  6.439625e-03 -2.381647e-02  1.546486e-02  1.750502e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  1.997001e-02  7.167108e-02  1.070128e-01  1.271227e-01  3.008860e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -3.851679e-02 -2.409385e-01  1.260602e-01  4.466355e-02  2.069249e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  1.657111e-02  6.535875e-02  3.490099e-02 -7.314759e-02  7.385109e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  1.608465e-01  8.950218e-02  5.030577e-03 -1.442860e-02  1.780315e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -1.278046e-01 -4.823574e-02 -3.635242e-02 -5.762079e-03 -2.504038e-03 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -8.153777e-02 -1.637040e-01  1.319943e-02 -3.074068e-03 -6.188299e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  4.676667e-03 -4.482924e-02 -1.540046e-02  2.929718e-02 -2.257554e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  8.222289e-02 -5.516410e-02  3.118634e-02 -4.844069e-02 -4.901648e-03 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -3.561802e-02  1.755339e-02  6.116413e-02  9.189857e-02 -2.083209e-06 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  1.669454e-01  9.553444e-02  1.906161e-03  1.357529e-01 -1.309280e-01 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -2.739622e-02  1.660607e-02 -5.399755e-02 -6.262589e-02  4.588603e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  1.385055e-02  7.154779e-03 -1.483355e-02  9.677839e-02  3.278150e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -2.071985e-02  1.203902e-02  1.632430e-02 -6.846546e-02 -4.015991e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  9.050562e-03  9.060744e-02 -7.182217e-02  5.623090e-02 -1.394004e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -1.359384e-02  6.037452e-02 -2.382161e-02 -2.036577e-02 -3.496412e-03 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -5.009490e-02  1.330226e-01  2.336805e-02  1.138030e-01  7.918928e-03 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  1.038962e-01 -5.352258e-02 -2.351218e-02 -1.316665e-01  5.244603e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -2.112519e-02  9.876155e-03  9.386394e-02 -5.566031e-02  2.810365e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  1.767093e-01  6.389297e-02 -1.189163e-02  9.972763e-04 -2.328432e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -1.724119e-02 -7.821570e-04  7.231102e-02 -7.051619e-02 -6.180652e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -7.343384e-02 -3.391118e-02 -4.678558e-02  5.269440e-02 -1.259454e-03 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  2.386623e-02 -5.106428e-03 -7.765110e-02 -2.613124e-02  1.507035e-01 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  1.403615e-03 -1.173905e-01  7.710084e-02 -7.167199e-03  3.185759e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -1.479761e-01  2.633919e-02 -1.326046e-01