Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sjorstrom, M., Sandberg, M., & Rannar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239–253.

Sjostrom, M., Rannar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295–305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0750749167  0.0848505337 -0.0615447044  0.0098169196 -0.0189978812 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0375747208 -0.0296800547  0.0121787322 -0.0205410613 -0.0571280738 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0344969761  0.0025624596  0.0309846123  0.0808168198  0.0387693626 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0633576732 -0.0229616567 -0.0828736603 -0.0337376129  0.0669482854 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.1426241178 -0.0211844688  0.0011674845  0.0480043743 -0.1180292423 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.0154076726  0.0201476982 -0.0025319687  0.0434883557 -0.0714196814 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  0.0500182575 -0.0186436259 -0.0143916861  0.0031414177 -0.1784597170 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0240363479  0.0957533338  0.0111294811  0.0768656315  0.0320514172 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.1091659678 -0.0149953903 -0.0126056597 -0.0120725838  0.0605767745 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0620945185 -0.0716845218 -0.0951027579 -0.0255788936 -0.1110512460 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0170336554 -0.0414674213 -0.1149259816 -0.0395082043 -0.0354321255 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0801397400 -0.0248520229 -0.1272264150 -0.0194550487  0.0244147745 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.1364954100  0.0810963565  0.0628289327 -0.1373782473  0.1061819325 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0363968716  0.0682951550  0.0282681954  0.0395000265  0.0053039786 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0235548245 -0.0405076958  0.0024245140 -0.1438209269 -0.0867089780 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0111226025  0.0308991138 -0.0280348217  0.0783131387 -0.0382110878 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0442062932  0.0483991073 -0.1156126079  0.0457847522 -0.0159129553 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0736723855 -0.0154592482 -0.0006653833  0.0048535984  0.0540377815 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0602021453 -0.0503742017  0.0503980764 -0.0180560269 -0.1057703181 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0958068703 -0.0468550924 -0.0599732167 -0.0347883872 -0.1177173540 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0183715897 -0.0396960498  0.0296899517  0.0794404014 -0.0256318960 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0186809704  0.0539064902  0.0992498020 -0.0138968429  0.0515905318 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.1111030464  0.1069783463 -0.0054025200  0.0860211900 -0.0853655653 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0939460104 -0.0076773992 -0.0217498418 -0.0421976879  0.0854115319 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0795411608 -0.0180841784 -0.0768465511  0.0083596326  0.0331801882 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0508366317 -0.0634185584  0.0804733452 -0.0796886568 -0.0923995923 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0602184565  0.0703893071  0.0049776687  0.0593312119 -0.0104019935 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0292711560 -0.1061338052  0.0231934375  0.1089184942  0.0703087156 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0516312411 -0.1043995745 -0.0396162383  0.0750264819 -0.0474724011 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.1264689864 -0.0883531949 -0.0004002042  0.0095663206  0.0155890181 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0113007563  0.0821423236 -0.0150505784 -0.0514835863 -0.0257126744 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0053087059 -0.0266620435  0.1190891308  0.1059167046  0.0195010132 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0203161213  0.0647378446 -0.0228588509 -0.0512710616  0.0783432455 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0004213126 -0.0579852956 -0.0542145921  0.0248967930 -0.0552462046 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0807094268 -0.0096928333  0.0235028642 -0.0151401348  0.1592056144 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0196500485 -0.0890580458 -0.0247318313 -0.0556488619  0.0154648545 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.1365196091  0.0040679927 -0.0281934196  0.0860517445  0.0281057099 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.1082216186  0.0273566333  0.0790254261  0.0210287059 -0.0581457923 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0224607887  0.0111080549 -0.0722774820  0.1151019796 -0.0969365318 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0856048969  0.1625402466  0.1272119736  0.0325247926 -0.0130680984 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0925062200  0.0141291333  0.0095810294  0.0291313074  0.1436421609 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0237038914 -0.1001886690  0.0125506867 -0.0241819506  0.0555041641 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.1236825335 -0.0128199418 -0.0912686730  0.1200237984  0.1233580114 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0423884183  0.0601168342  0.0764057694  0.0702884281  0.0492540633 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.1512475609 -0.0163364225 -0.0214162689 -0.0574113144 -0.0031114011 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0373914150  0.0045055289 -0.0555754655  0.0093731852  0.0482270648 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.1152183037 -0.0173228180 -0.0305631577  0.0211728647  0.1336502105 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0204202222  0.0346654919  0.0313058704 -0.0481720109 -0.0892439439 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0492653269  0.0141581774 -0.0279323650  0.1025730401  0.0016526276 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0009027473 -0.0073437973 -0.0807967268 -0.0283610933  0.0599505687 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0233566354  0.0814850502 -0.0396266605 -0.0035028165  0.0124356262 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.0130836106 -0.0815037803  0.0061831199 -0.1326869521 -0.0220830355 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0434228525  0.0110804228  0.0955020475  0.0679545619 -0.0867197996 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.1455640029  0.0196687714 -0.0045710613 -0.0817822125  0.0312150122 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0063292442  0.0291248548 -0.0075790280  0.0688577286  0.1595466307 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0092456656  0.0081026916  0.1316805433  0.1178220996 -0.0521615101 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0440540679 -0.0250678264 -0.0322170510  0.0186735058 -0.0415203229 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0291200215 -0.0002804389 -0.0385342430  0.0412550039 -0.0299134961 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0508337255 -0.0292301645  0.0244460858 -0.1987400288  0.0364009358 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0052617139  0.0786639496  0.0152617383  0.0170291513  0.0046948882 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0094720140 -0.0288440229 -0.0292318008  0.0893239179 -0.0123649547 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0207190257  0.0457774991  0.0599270771  0.0268403679 -0.0597575679 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0198972704 -0.0957971776  0.1065109004  0.1649254199 -0.1394064931 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0235809384 -0.0045086385 -0.0720961823  0.1903239351 -0.0629214479 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0019108549 -0.0590843100 -0.0661819079  0.0707373080  0.0027763209 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0688539702  0.0147834844  0.0964657129  0.0870268781  0.0098427634 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0113288400 -0.0393408144 -0.0394333532 -0.1046570959 -0.0054636931 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0567874847  0.0192822661  0.0048684463  0.0923967279  0.0389993664 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.0640241149 -0.0044147355 -0.0071742657  0.1319565048  0.1015062142 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0191537334 -0.0825288547 -0.0311186438 -0.0416762375 -0.0123191693 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  0.0278724264  0.0804889755  0.0258253637 -0.0966050140 -0.0103958704 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0625276653 -0.0651988992  0.0082932702  0.0092559623 -0.0655063851 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0097800484 -0.0732940516 -0.0466052207  0.0697716411 -0.1123779546 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0385114982  0.0777371152 -0.0400864100 -0.0098903437 -0.1580303267 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0371292016  0.0502117245 -0.0341469316 -0.0500379974  0.0684642554 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0267630851  0.0105258125 -0.0203515245  0.1381757545 -0.0051176222 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0774099798 -0.0975210478  0.0208717597  0.0333753640 -0.0391774658 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.1093663549  0.0905878459  0.0942138701 -0.0323917529  0.0936525685 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.1491901686  0.0389391023 -0.0038218978 -0.0558881942  0.0387331482 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.1447996986 -0.0506487059 -0.1285705617 -0.0501242375  0.1243887730 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0020198074  0.0441865542 -0.0392187524 -0.1295615370  0.0419010311 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  0.0140740224 -0.0642015598 -0.0377993368 -0.0663306058  0.1289224409 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0037886759 -0.0297655947 -0.0240082243  0.1086263076  0.0418688175 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0114805464 -0.0700975805  0.0447490074  0.0450889447 -0.0229499541 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.2283277472  0.0696877548 -0.0422644305  0.1095367128  0.0295807158 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0330404579 -0.1203486999  0.0259309805  0.0965258342  0.0168379359 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0750380264  0.1267340899  0.0010283601  0.0724076967 -0.0834645838 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0939667295 -0.0172670307 -0.0449637980  0.0123231920 -0.0081187667 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.0286353278 -0.0645978208 -0.0127335079  0.0118020654  0.1335730214 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0019784276  0.0441755440  0.1247734238