Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0928136421  0.0660134785 -0.0342344499  0.0534988005  0.0686528270 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0694629446  0.0214891777 -0.0272650782 -0.0340021184 -0.0177425279 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0086266144 -0.0354453741 -0.0431007967 -0.1013360213 -0.0381060066 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0705354774  0.1798176487  0.0652687740  0.0498249512  0.1304701304 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0569669239  0.0271410472 -0.0906094579 -0.0115415294 -0.0841886912 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0126837172 -0.0676716168  0.0122033225  0.0023396879 -0.0458136979 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0892059354 -0.0668461018 -0.1460125987 -0.0674656584  0.0866326275 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0527374673 -0.0574556957 -0.0585387710  0.0390104747  0.0822081364 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0590454134 -0.0495919700  0.0485224777  0.1371554720 -0.0597765623 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0026552235  0.0441669853 -0.0901276417 -0.0721534449 -0.1625429080 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0665325036  0.0250687641  0.0681894328 -0.0799914404 -0.0320222060 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0844974398 -0.0173386596  0.0340734487  0.0426793200 -0.1547953657 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0501573437  0.0961617186 -0.0419440627  0.1008956842  0.0503204846 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0423601000 -0.0409944362  0.0973856216 -0.0190125919  0.1374729885 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0074667835 -0.1536082860  0.0212648331  0.0921431023 -0.0239374490 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0767459290  0.0584051014  0.0056829339  0.0087749338  0.0487127032 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0014839478  0.0654368893 -0.1623498600  0.0957569724 -0.0757913820 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.0812709471  0.1110438422 -0.0139470018  0.0876993267 -0.1214043378 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0838645182  0.1538897491  0.0712055878 -0.1606518118  0.0272883592 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.1035199671  0.0650050030  0.0333013245 -0.0586317699  0.0182633207 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0864470491 -0.0679940078  0.0223002212  0.1042070036  0.0982415994 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0182101204  0.0508502948 -0.0712744281 -0.0725492632  0.0370489342 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0952713942  0.0715430394 -0.1051689668  0.0663373485  0.0207541580 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0760303690  0.0731322988  0.0928873637  0.0659406911  0.0445598580 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0051758878  0.0243910435  0.0462376977  0.0797555045  0.0323872141 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0180523420 -0.0025885848  0.0305925024 -0.0214100576  0.0684157671 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0064156735  0.0763065505  0.0100194430 -0.1272616334  0.0218826889 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0306945859 -0.0058617289  0.0286505441  0.0672672109 -0.0307526747 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -0.0174167332  0.2122838495  0.0132988379  0.0065888989 -0.0727726034 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.1029497465 -0.0519615154 -0.0647025073  0.0611781360  0.0701024535 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0277385585  0.2158047943 -0.0801266401  0.0838342035  0.0476356877 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0126965847 -0.0798442076 -0.1608826971 -0.1187621172 -0.0188355189 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0982293876  0.0040049212 -0.0015600288 -0.0018080718 -0.0446567002 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.1861703882 -0.0144942826 -0.1190304896 -0.0329596285  0.0506753949 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0576992345 -0.0321319319 -0.0217271815 -0.0155700010  0.0047614487 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.1580038027  0.0223176345  0.1009017342 -0.0372719439 -0.0659195762 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.1397777249 -0.0174247778  0.0563395324  0.0001065575  0.0141097695 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0901103476  0.0331895978  0.0555996168 -0.0262511412 -0.0728474472 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0765506839 -0.0006975334  0.0421364889 -0.0165484664  0.1056238967 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.1056334916  0.1519068558  0.1009748515  0.0200744743 -0.0993438274 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0149524981 -0.1518137454 -0.0809542400  0.0643684279  0.0978599490 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0943502118 -0.0182577467  0.0089639713  0.0451270159  0.0442521191 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0615002705  0.0877203973 -0.0771308298 -0.0069279997 -0.0347116739 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0681177906 -0.0447462517  0.0045008979 -0.0560673906  0.1054411367 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0328714271  0.0259252771  0.1049353957 -0.0696405847  0.0095171844 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.1407189884 -0.0727241277 -0.0663155875 -0.1340712471 -0.0487093340 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.0546271782 -0.1030281338 -0.0662626322 -0.1031306007 -0.0775069252 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0213967515 -0.1627097617 -0.0275670729  0.1223120153 -0.0303530974 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0650153560  0.0172913906  0.0261459755 -0.0833789690  0.1172487163 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0473188237  0.0754533378  0.0389302709  0.1208329087  0.0042560593 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0272449024  0.0429155883 -0.0448883774  0.0588765060  0.1313193709 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.1529578277  0.0095375941 -0.0645934145  0.1040028509 -0.0017063792 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0394536152 -0.2341515726  0.0891709583  0.0099652668  0.1457088782 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.0082278449  0.1321884620 -0.0866154897 -0.0012315916  0.0596320135 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0566345352  0.0794165345 -0.0183430677  0.1145103172 -0.1300538058 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0083846469  0.0406619212  0.0975926323 -0.0935007877  0.0326610472 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0839420354 -0.0125312824 -0.0298617159 -0.1186166788 -0.0796195845 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0630733069  0.0949189075  0.0293895181  0.0266606340  0.1177733192 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0700233806 -0.0455841632  0.0244372293  0.0388313688 -0.0366646399 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0640241121  0.0807946548 -0.0830559555  0.0031604433 -0.0218835019 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0368639150  0.1485740861 -0.0741306733 -0.0319417097  0.0867313587 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0205675282 -0.1474070516  0.0221568304 -0.0244128985 -0.0177875762 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1092107493  0.1023822922  0.0803498296  0.0311059449  0.0754647442 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0451656372 -0.0234971401  0.0139956366 -0.0187291445 -0.0198777850 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.1105355118  0.0523307787 -0.0370420201 -0.0827767581  0.0729672431 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.1275143800  0.0286542723 -0.1267620357 -0.0048569356  0.0111041399 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0880638284  0.0242701761  0.0020345654 -0.0037342156 -0.0196553681 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0872823845  0.0223545661  0.0555332535 -0.0016772321  0.0741108284 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.0207659353 -0.0683677749  0.0976674818  0.0434654174 -0.0128220149 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0337450889  0.0702127829  0.0032671055  0.0411290979 -0.0494401541 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  0.0015998713 -0.0824370006 -0.0828708088  0.0784501767  0.0002184029 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.1037604881 -0.0932434379  0.0603845140  0.0369093344 -0.1333687747 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.1651067445 -0.0892257608  0.0360345419  0.0290038477  0.0841094182 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0196508875 -0.0705923829  0.0548195052 -0.0580882413 -0.1377847973 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0565400097 -0.0931252702  0.0439806531  0.0062904863  0.0661356016 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.0469637654 -0.0251110638  0.0654435535 -0.1401412966  0.1172547667 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.1239955767  0.0034568619 -0.0635514036  0.0517273929 -0.0161178545 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0665456121 -0.0186365323 -0.0386295721 -0.0665346117  0.0277078619 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0142041188  0.0999517844  0.0527462303  0.0263034842  0.1072283874 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.0234691995 -0.1533651255  0.0244589026  0.0421866708 -0.0027954512 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0061453288 -0.1151843164 -0.0159970513 -0.0841248314 -0.0750012310 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.2016100254 -0.1731422306  0.0359209347  0.0486135454 -0.0724748647 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0371544837 -0.0536086031 -0.0192238967 -0.0519509126 -0.0641625821 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0457271427 -0.1069663559  0.0045578742 -0.0005422571 -0.0792788336 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0191993780 -0.1119679553 -0.0149772322 -0.1007577769 -0.0024211877 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0158227881 -0.1533894039 -0.0819044760 -0.0228150312 -0.0245748818 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0876479053  0.0553673222  0.0546010604 -0.0937043140 -0.1016742226 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0346441340 -0.0275078716  0.0073354873 -0.0877397577  0.1416351323 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.0494853885  0.0006576911 -0.0268830287  0.1207864385 -0.1350048160 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.0475288384  0.0360754351 -0.0046841156