Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -5.420593e-04  1.490260e-01 -5.429176e-02  9.563079e-02 -3.758346e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -1.322286e-02 -5.615933e-02 -1.580634e-01 -2.325052e-02  1.110731e-01 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  9.485463e-03  9.595728e-03 -1.592733e-02  3.543482e-02 -2.083254e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  7.652117e-02 -8.562855e-03 -6.619251e-03 -1.224663e-01  8.510876e-03 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -7.668979e-02 -3.282286e-03 -1.014214e-01 -8.730855e-02 -4.464508e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  4.459624e-02  4.006915e-02  3.842476e-02  1.161836e-01 -2.740608e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  1.487662e-01  4.862363e-03 -7.241181e-03  5.745644e-02 -3.086880e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  3.960444e-02  7.641318e-02 -9.593230e-02  7.029626e-02  3.292032e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -7.851453e-02 -1.181723e-02 -4.166120e-02 -1.166790e-01 -9.108880e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -8.088429e-02 -2.538687e-02 -3.883969e-02  8.941426e-02 -8.004839e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -3.416938e-02 -4.156103e-02 -1.735220e-01 -8.848146e-02 -3.350074e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -1.568580e-01 -9.791302e-02 -2.747683e-02 -1.889805e-01  7.087901e-03 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -4.054537e-02  1.788256e-01 -2.402160e-02 -1.233643e-01  8.548172e-03 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  1.942850e-02  7.455982e-02 -1.425874e-02 -1.565516e-01  1.744785e-01 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -1.767574e-01 -8.385365e-03 -1.530634e-03 -9.795976e-03 -1.052272e-01 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -8.036218e-02  4.259985e-02 -8.322276e-02  3.497524e-02  9.798692e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  5.289939e-02 -8.749676e-02 -3.927190e-03 -4.244098e-02  1.941486e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -8.510882e-02  3.626045e-02  1.937260e-01 -1.252201e-01  1.072012e-01 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  6.551831e-02 -1.520247e-01  3.290812e-02  1.107773e-02  3.881200e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -7.888705e-02 -1.396531e-01  6.115925e-02  1.665701e-01  2.279460e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  7.859268e-02  1.375189e-03 -3.077403e-02 -1.357264e-02  5.829573e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -9.881096e-03 -3.617391e-02 -1.747801e-02 -4.345553e-02  5.587692e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  1.192142e-01 -3.738191e-02  1.195235e-02  5.437023e-03  5.705000e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  9.722119e-02 -4.385725e-02 -8.664756e-02 -1.740706e-02 -2.426360e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -2.121607e-02 -2.652052e-02  2.852263e-02 -1.251582e-01 -2.752532e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -8.078602e-02 -2.569379e-02  9.584123e-02  3.064385e-02  2.690695e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -3.571239e-03  2.038811e-02 -1.761991e-02  1.585840e-02 -4.329767e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  1.009867e-01  1.028650e-02 -6.063231e-03  3.059003e-02  2.306003e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  7.319647e-02 -1.099566e-01 -1.815091e-01  3.399751e-02 -2.236745e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -8.439686e-03  4.633276e-02 -3.458356e-02  8.455483e-02 -1.380416e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -1.966553e-02 -1.505274e-02  1.241381e-02 -1.353436e-01  2.609079e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -2.086005e-02  5.779303e-02  5.871593e-02  3.873955e-02  1.417532e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -7.803887e-02  7.018965e-02  8.813560e-02  3.747995e-02 -5.966525e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  5.641241e-02 -8.366642e-02 -3.395181e-02 -2.557432e-02  6.391626e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -9.801057e-02  2.627675e-02  8.036328e-02  4.502809e-02 -8.200561e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -1.163399e-01 -1.753244e-01 -4.706279e-03 -5.399940e-03 -1.097291e-01 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -6.000971e-02  3.642325e-02  2.224338e-01 -2.539797e-03  8.350303e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -1.783835e-01 -8.566847e-03  2.074534e-01 -4.304085e-02  1.360816e-05 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -1.260800e-01  9.256317e-03  4.906258e-02  5.909934e-02 -7.660391e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  7.261602e-03  7.301958e-02 -7.113231e-02 -1.020382e-02  4.537573e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  5.339710e-02 -3.760444e-02 -1.244793e-01  3.899262e-02  1.054230e-01 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -7.828245e-02 -7.507803e-03  5.171451e-02 -1.310155e-01 -5.965623e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  3.028708e-03  8.869718e-02 -6.571043e-02  9.818278e-02 -2.073972e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -6.003858e-02  9.796127e-02 -1.093502e-01  4.096917e-03 -2.849652e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  3.582824e-02 -8.694378e-03  2.364211e-02  1.102801e-01 -7.239449e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -8.535296e-02  2.243671e-02  6.747847e-03 -1.584457e-01 -7.840151e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  3.535884e-02 -7.485049e-02 -4.340166e-02  3.803668e-02 -4.022339e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -2.548932e-01  7.680251e-02  2.448535e-02  8.501374e-02  4.978020e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  3.573539e-02  3.764389e-02 -1.823716e-02  6.358295e-02  4.326795e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  8.423820e-03  5.049461e-02  2.828131e-02 -5.258237e-02 -4.341492e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -1.659763e-02 -3.424549e-02  1.332379e-01  2.195037e-01  2.441121e-03 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  4.775066e-02 -5.422614e-02  6.235765e-02 -1.356098e-02 -2.827735e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  5.762422e-02 -3.999559e-02 -7.859094e-02 -1.388739e-02  9.592826e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -7.539311e-02 -7.130042e-02  8.684557e-02 -6.227985e-02  8.388994e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  6.629688e-02  2.347615e-02  1.080077e-02 -5.374961e-02  3.887759e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  8.081190e-04 -7.267766e-02  6.063558e-02  1.672151e-01 -1.188478e-01 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  6.544791e-03 -4.574918e-02 -1.189311e-01 -5.320247e-02 -1.549080e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -5.619281e-02 -2.344425e-02 -9.196659e-03 -1.192254e-01  1.452441e-01 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -7.852628e-03  5.659577e-02  4.581270e-02 -6.532361e-02 -6.186188e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -6.113534e-02  1.731253e-01 -1.405770e-01 -3.930455e-03 -3.443584e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -1.604839e-02 -1.051900e-01 -8.757975e-02  9.811647e-03 -1.166988e-01 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  8.926780e-02  1.345406e-01 -1.620683e-02  3.412677e-02  3.562193e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  1.098264e-01 -4.283658e-02 -2.261924e-02  6.498719e-02 -1.285748e-01 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  5.560813e-02 -3.974547e-02 -4.206502e-02 -3.535246e-02  6.076681e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -6.797015e-02 -4.762470e-02  9.449612e-02 -7.640994e-02 -1.185975e-01 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  1.262793e-01  1.509732e-02  1.275258e-01  2.586690e-02 -3.983790e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -1.011408e-01  4.430925e-02 -8.232207e-02  3.369894e-02 -4.022683e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  4.371175e-02  5.067350e-03 -5.164390e-02  1.250305e-02 -5.236835e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  1.493228e-02 -2.047473e-03  1.800692e-02 -7.414194e-04 -8.009270e-03 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  2.702629e-02 -3.308482e-02 -4.621830e-02  8.983795e-03 -3.563563e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -8.601162e-02 -1.007421e-01  1.509716e-02  1.279885e-01  1.296139e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -1.300803e-03  8.754694e-02 -7.888195e-02 -5.358997e-02  4.721136e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  2.339621e-01  1.250658e-02 -1.646592e-01 -7.710522e-02 -5.769174e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  1.174494e-01  2.377421e-02  9.668495e-03  8.195306e-02  8.764670e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  1.372067e-01  3.254185e-02  8.539817e-02  1.268644e-01 -1.746081e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  1.255847e-01  1.518144e-02 -6.230789e-02 -3.810535e-03  1.023464e-01 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  6.351075e-02 -6.077550e-03  9.775032e-02 -2.439327e-02 -1.923166e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  9.958872e-02  9.068220e-02 -1.189617e-02 -9.389195e-04 -9.487926e-03 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -6.625028e-02  4.906920e-02 -1.442948e-02  7.169865e-02  4.726783e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  4.142996e-02 -1.103588e-01 -1.210207e-02 -1.135589e-01  1.858188e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  4.924552e-02  7.087920e-02 -6.848424e-02 -9.964630e-02 -7.414302e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -8.700125e-02  6.194588e-02  5.714989e-03  6.158321e-02 -9.873232e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  6.258317e-02  1.441909e-01 -9.451181e-02  1.318580e-01  1.067355e-01 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -5.020939e-02 -4.290127e-03  1.124075e-01 -3.743530e-02  1.139830e-01 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -1.438275e-01  3.819494e-02 -6.895585e-02  1.987415e-01  8.170512e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -2.173167e-02 -5.520973e-02 -1.385513e-02 -3.663035e-02  7.598956e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  5.330919e-03  2.297252e-02 -1.657783e-01  9.331632e-02 -1.694376e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -1.207882e-02  1.139623e-02  4.290195e-02 -9.015456e-02  9.727554e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  5.395574e-02 -2.117296e-02 -1.098926e-01 -3.719386e-02  1.070303e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -7.090733e-02  1.211893e-01 -3.086436e-02