Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -3.866037e-02  7.904711e-03  1.725198e-02 -1.028387e-01 -1.306323e-01 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -6.969637e-02  3.327112e-02  7.210649e-03  6.817201e-02  1.005868e-01 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  9.462342e-02  8.945033e-02  1.043935e-02  1.265467e-01 -7.330927e-03 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  5.889535e-02 -1.042306e-01  1.271789e-01 -9.006074e-03  8.557693e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -2.597208e-02  6.664972e-02  4.143926e-02  4.867799e-02  1.943971e-01 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -2.219595e-02 -6.984306e-03 -2.101514e-02 -6.194916e-02  4.838950e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -1.814158e-02  1.180377e-02 -1.604120e-02  1.672132e-01  4.768891e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  1.340735e-01 -4.692735e-04 -1.700516e-02  9.275213e-02  1.286272e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  8.462668e-02  1.327813e-01  6.639362e-02 -5.300304e-02 -8.930449e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -4.890787e-03  4.459701e-02  1.885358e-01 -2.792608e-02  4.464640e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  1.601812e-02  1.135837e-01 -1.126531e-01 -8.864230e-02  1.324976e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  4.207667e-02 -1.128576e-01 -1.727750e-02 -6.737170e-02  9.385001e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  9.045084e-02  9.480850e-02  7.319690e-02  3.560219e-02  7.860519e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  9.756902e-02 -3.136040e-02  2.763458e-02 -5.085329e-02 -3.863139e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  8.684461e-03  5.957456e-02  4.367802e-02 -2.777644e-02  8.643458e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.041131e-02 -4.999880e-02  9.806184e-02 -1.094435e-01  1.126085e-01 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  1.095702e-01 -1.230883e-02  3.139579e-02  2.578561e-02 -4.542728e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  7.094907e-02 -2.348464e-02  5.454956e-03 -6.787546e-02  6.478982e-03 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  1.865177e-02 -6.282763e-02 -4.321281e-02 -1.322924e-01 -3.983675e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -2.621609e-02 -1.144968e-02 -4.823403e-02 -5.682026e-02  6.947305e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  2.575120e-03  3.462864e-02 -7.019066e-02  2.439502e-02 -4.495861e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -2.615389e-02 -2.751598e-02  3.418442e-03  2.200434e-02 -4.007728e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -4.878996e-02  2.347960e-02  4.086726e-02 -1.128043e-01 -4.643590e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -1.003228e-01 -4.460022e-02 -9.670900e-02  6.842571e-02  3.966520e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -1.131841e-01  4.673973e-02 -7.156104e-02  7.940546e-02  9.307575e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -8.238672e-02  6.327037e-02  1.174469e-02  5.944677e-02 -2.314888e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -9.147651e-02 -5.704462e-02 -4.637396e-02 -5.033534e-02  3.664535e-03 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  5.766413e-02  5.418803e-02 -2.424499e-02 -4.417954e-02 -3.180222e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  1.766751e-02  4.130066e-02 -2.396350e-03 -9.410791e-02 -2.031535e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -2.913640e-02  5.615585e-02 -5.715337e-03  5.391789e-02 -9.805838e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -5.753670e-02 -7.978918e-02 -8.082933e-02  4.462938e-02  1.979483e-01 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -6.577280e-02 -6.726884e-02 -4.569451e-02 -2.339877e-02 -6.325621e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  1.799005e-02  1.268631e-02  1.227672e-02  7.493253e-02 -1.244597e-03 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  5.128951e-02  3.577908e-03 -1.008024e-01 -1.501587e-01  3.453566e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -5.896104e-02  6.881613e-02  1.135300e-02  9.904675e-02 -1.284856e-01 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -3.718792e-03  1.176827e-01 -1.323379e-02  6.811098e-02 -1.023721e-03 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  4.790886e-02 -1.089482e-01 -3.195066e-02 -4.757520e-02  6.096617e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  2.710059e-02 -1.257487e-01  3.401514e-02  7.114237e-02  3.953133e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  1.489764e-02 -5.296156e-02 -8.409643e-02 -4.743224e-02  2.819547e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -6.337670e-02 -7.913935e-02  1.830718e-02  4.103072e-03 -3.700082e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -5.713238e-02  2.342789e-02 -9.286913e-02 -8.064159e-02 -2.831176e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  1.845020e-02  9.753632e-02  4.369218e-02  8.974031e-02  1.113944e-01 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  1.031706e-01  1.300174e-01  2.984573e-03 -2.554493e-02  1.228084e-01 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  1.863425e-02  1.199913e-01 -8.609670e-02  6.159672e-02 -7.633585e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  7.543206e-02 -1.366635e-03 -6.650629e-02 -1.076603e-01  2.541836e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  2.577340e-02  7.134073e-02 -2.340347e-02  4.908058e-02  2.124842e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  7.494113e-02 -1.037419e-01  2.427550e-02 -5.261121e-02  1.023721e-03 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -3.340594e-02  6.511415e-02 -1.783699e-02  5.211810e-02  5.237277e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -4.269720e-02 -3.904672e-05  2.240700e-02 -1.906455e-03  5.322650e-03 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  2.964202e-02  1.955055e-02  4.926097e-02  4.940011e-02 -2.024151e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  5.571310e-02  3.797458e-02 -2.745766e-02 -1.250520e-02  2.197070e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -2.823974e-02 -1.192151e-01 -5.703013e-02  2.164960e-02 -1.136443e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -2.550011e-02 -7.874593e-03 -2.712176e-02 -9.071589e-02 -2.584786e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  8.198614e-02 -1.789251e-01 -8.582526e-03  5.178989e-03 -9.707630e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -5.027597e-02  2.922458e-02 -3.974731e-04  8.306548e-02  1.089410e-01 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  5.892110e-02 -7.343007e-02  6.718716e-02 -8.540275e-02 -2.268926e-03 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -1.167503e-01  4.747736e-02 -1.775253e-02  2.123080e-01  5.700126e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  5.703506e-02  1.258576e-02  6.995888e-02 -2.504448e-02  5.412514e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -8.374473e-03 -5.721200e-02 -9.375447e-02  4.071683e-03  7.422666e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -3.489211e-02  1.575167e-01  3.856236e-03  7.238512e-03  7.868769e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  3.283580e-02 -1.374234e-02 -1.846084e-01 -2.323777e-02 -4.433219e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  5.883201e-02 -4.191962e-02  6.725999e-03 -7.995146e-02 -8.124615e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -3.644275e-02  7.142430e-02 -2.745373e-02  3.069979e-02  1.245809e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -4.375684e-02 -5.517060e-02  1.101796e-01  6.768215e-02  3.828465e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -6.578799e-02 -1.522200e-01  3.787328e-02  6.527884e-03 -4.879963e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -7.529430e-02 -7.955345e-02 -6.648881e-02 -1.514660e-02 -1.886287e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -3.913750e-02 -5.000812e-02  1.418153e-02  1.975923e-02  1.293228e-01 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -9.151825e-02 -1.414229e-01  5.164327e-02 -3.429418e-02 -1.280446e-01 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -4.952414e-02 -3.758670e-02  4.074735e-02  4.094110e-02  1.766788e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  6.450435e-03  2.935413e-02  4.305869e-02 -1.322304e-01 -5.260669e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  4.840226e-03  7.393319e-02  7.458053e-02 -1.244814e-02  5.707764e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  4.323416e-02 -4.205585e-03 -8.790476e-02  6.768689e-02 -5.750771e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  3.497626e-02 -3.043259e-02  2.566813e-02  7.418244e-02  2.708586e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  6.771061e-02 -5.778103e-03  4.351973e-02 -1.327834e-01 -6.452522e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  4.412804e-02  8.219499e-02 -2.411994e-02 -5.023808e-02  4.605551e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -1.263699e-01  5.139105e-02  3.637766e-02 -4.837223e-02  3.629306e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -2.749844e-02 -5.023274e-02  3.973585e-02  4.723064e-02 -3.918988e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  6.092463e-03  1.374099e-01  1.715541e-02  1.270404e-02 -1.921318e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -1.072194e-03  4.328452e-02 -1.164020e-01  4.061506e-02 -3.897185e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  6.329983e-02  4.989030e-02  3.549444e-02  4.362855e-02  7.518170e-03 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -4.562694e-02  3.585546e-02  4.377759e-03 -3.910188e-03 -2.353194e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  2.874868e-03  3.568141e-02 -1.173990e-02  4.730300e-02  5.673367e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -2.210454e-02 -1.625615e-02 -1.683621e-01  1.685232e-02 -3.894850e-03 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -6.345031e-02 -7.506935e-02  3.065284e-02  9.607237e-02  7.533877e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -1.017908e-01 -4.557228e-03 -1.468591e-02 -6.156208e-02 -1.037939e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  4.343355e-02  2.187716e-02 -1.332524e-01 -7.985559e-02  5.986050e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -1.029097e-01 -2.490224e-02 -9.866853e-02  3.268752e-02 -9.167672e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -6.646105e-03 -3.977698e-02 -9.709527e-02  6.326932e-02 -2.201547e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -2.365601e-02  6.366304e-02 -3.300287e-02  1.137538e-02 -7.479126e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -1.151805e-02 -1.899448e-02 -1.514411e-02