Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0758582845  0.0890746624 -0.1031733721 -0.1503676382  0.0664710901 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0035847379 -0.0227154003 -0.0809471541  0.1056134454 -0.1300383710 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0983594736  0.0096402342 -0.0416864876 -0.0949141442 -0.0516132950 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0759776444 -0.0519687152  0.0711225916  0.0352607985  0.0710159911 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0128281258 -0.1288331461 -0.0019736777 -0.0756097161  0.1344672516 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.0212270724 -0.0695883626 -0.1533531066  0.0715552532  0.0728301155 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  0.0686470737 -0.0685382246 -0.0536599210  0.1095234955 -0.0666125041 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.2233096204  0.0961244233 -0.0444613213 -0.0007646249 -0.0161923279 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0654595337  0.0498325991  0.0080744210 -0.0588454387 -0.0005494256 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0124181153  0.1018088234  0.0690409693 -0.0725315848 -0.0391861131 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.0792406216 -0.0221122187 -0.0847414440 -0.0427386005  0.0181276578 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0347472879 -0.0202570123  0.0965555057 -0.0851014579 -0.1015271519 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.1197935299 -0.0741447098 -0.0426853947  0.0166340988  0.0141944836 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0184126860  0.0355309010 -0.0047235953 -0.0139940858  0.1365264658 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0173154783 -0.0743992100 -0.1106918832  0.0410314497 -0.0205184293 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0812698408 -0.0603668639  0.0264803055 -0.0750374274  0.0322938719 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0676631444 -0.0248316562 -0.0541320878  0.1386987629 -0.1428300290 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.1300356207  0.0301669325 -0.0249214056 -0.1499658949 -0.0190844578 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0589615183  0.0933642289  0.0343463916 -0.0330495259  0.0091724604 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.0411516492 -0.0939347186 -0.0890899413 -0.0340670074 -0.0266732955 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0450969184  0.0777113604  0.0955546472  0.0279384339  0.0104054735 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.1390857946 -0.1064743327 -0.0445312924  0.0079892350  0.0787715513 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0231932124 -0.0499854333 -0.1064627843 -0.0167154735 -0.0485016221 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0844865644  0.1434590168  0.0953060173  0.0320668350  0.0467951134 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0120800475  0.1324223711 -0.0463952765  0.0006476993  0.0750275534 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0189104597  0.0003198221  0.0010321935  0.0006522209 -0.0182497195 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0497194941 -0.0236581778 -0.0748678332 -0.0315604634 -0.1231834247 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0168043670 -0.0195988762 -0.0485062949 -0.1101177061 -0.0009243550 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -0.0884152778  0.0941814410 -0.0058337365  0.1069601522 -0.0285926929 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0111518836  0.0793467932  0.0059497351  0.0443511590 -0.0161524855 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0353917334  0.0204864175  0.0020326566  0.1060552591 -0.0785307985 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0560147429 -0.1067712529 -0.1322359331  0.0354638940  0.1366835492 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0730803504  0.1308637126 -0.0944560962 -0.0843855376 -0.0356491261 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0442307157 -0.1294188038 -0.0370769771 -0.0413680981  0.0138269962 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0576879800 -0.0149826019  0.0208477053 -0.0549808360  0.0101972044 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0743173885  0.0699159012 -0.0900639344 -0.0634915442 -0.1473964516 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0701157728 -0.0831389067 -0.0657736707 -0.1237723442 -0.0891298138 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0366564874  0.0787573985  0.1022231656 -0.0601439181  0.0260315003 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.0243075530  0.0355143883  0.0210708825  0.0003615466 -0.0079745286 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0277701524  0.0286201266 -0.0796978336  0.0103290804  0.0192501648 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0063932048  0.0585956967  0.0159570250 -0.1635681929 -0.0716527490 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.1624974557 -0.0219567409 -0.0185675637 -0.0240120669 -0.2244031369 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0147372022  0.0399928750  0.0385049174  0.0223419034  0.1935104276 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0528139883  0.0705691696 -0.0406979262  0.0005260498 -0.0475875752 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0494623543  0.0450450997 -0.0138991555  0.0695317943  0.0378368329 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0532820500 -0.0822719169 -0.0150380129  0.1587383739 -0.0337177932 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0457567364 -0.0581801082 -0.0199546602  0.0368496220 -0.0327812958 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.0835338241  0.0323097528  0.1272730818  0.0944347325  0.0900232197 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.1101171175  0.0475867679  0.0273467484 -0.0213035700  0.0031575062 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0581994352 -0.0118254752  0.0871058651  0.0260198338 -0.1325857406 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0402801258 -0.0879508167  0.0290857787  0.0668723533 -0.0599129289 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0052585693 -0.1337912018  0.0310926104 -0.0104309355 -0.0964063130 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0266143485 -0.0201591607 -0.0176535095 -0.0424562969  0.0525548982 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.0785209932 -0.1036653770  0.0702350520 -0.1339626527 -0.0268049434 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.1872035825 -0.1531579045  0.0128018688 -0.0282318694  0.0068172898 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0897507048  0.0084023573 -0.0349638014  0.0112969198 -0.0760581399 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0460184195  0.0789062923  0.0895048846 -0.1000883508 -0.0468494637 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0353976330  0.0064371024  0.0176345588  0.0885283798  0.0927102863 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0336087408  0.0919815580 -0.0433847141 -0.0488885939 -0.0264030312 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0158599991 -0.1165661573 -0.0346531772 -0.0237521700 -0.0443444730 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0416920565 -0.0024773857  0.0882547250  0.0873944198 -0.0337537042 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0898893325  0.0891447344 -0.0129871454 -0.0741214384 -0.0201457012 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.0020522822 -0.0400481601 -0.0996742761  0.0141610683  0.0253336759 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.1074453549  0.0017677413 -0.0313007647  0.0932566443 -0.0698367760 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0443883973 -0.0351264077 -0.0752036158  0.0476816744  0.0132568292 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.1086972825 -0.0423035747  0.0974113950 -0.0159788512 -0.0372417644 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0110993118 -0.1245496860 -0.0369408806 -0.0516372413  0.0726267958 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0218340205  0.0547580837  0.0799043435 -0.1693480488  0.0290039326 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0928940360 -0.1824044716 -0.0303193800 -0.1026630586  0.0346224415 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.1024180203  0.0492495835  0.0441763826  0.0512533357 -0.0053245377 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  0.0030469173 -0.0959798367  0.0984382466  0.0857120954  0.1175521784 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.1267365222 -0.0306636725  0.0314569883 -0.1153621327  0.0384390319 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0036155764  0.0355768144 -0.0242214602 -0.1352600035 -0.0272405462 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0204142900 -0.0128356259  0.0603926622  0.0276663019  0.0155042042 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0541736520  0.0239541181  0.0094932867  0.0132870857 -0.0531628622 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.0409491181  0.0786735537  0.0313965705 -0.0930408634  0.0638067113 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0115144259  0.2526793845  0.0044548517  0.1586064698 -0.0364389744 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0966334206  0.0318629751 -0.1205838611  0.0133159333  0.0371236278 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0317420037 -0.0446311141  0.0102626087  0.0876261202  0.0246511128 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0365794218  0.0195891612  0.0041439493  0.0509531130 -0.0627686800 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0080186411 -0.0593319139  0.1180203444 -0.0127767990  0.0313225695 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0212130261  0.0463463084 -0.0676123242 -0.0667156710 -0.0658101060 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0202517477  0.0239607845 -0.0720884013  0.0316716795 -0.1248634368 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0276037484  0.0800854350  0.0031747970 -0.0393630321  0.1448203107 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0072180178 -0.0582899510  0.0339248873  0.0515768127 -0.0012301386 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0198683743  0.0811519388 -0.0726370067  0.1780196490  0.0429624979 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0374068487  0.1338163331  0.0668199864  0.0306289958  0.1122093638 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.1320144034 -0.0410686789 -0.0118497761  0.0470646718  0.1139230769 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0512617661  0.0778588755  0.0532874492 -0.0166919354 -0.0504267224 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0843286730  0.0525217090  0.0579942427