Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.1202799502 -0.0181562021 -0.0629203946 -0.0322302996  0.0235497227 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0945414080 -0.0421318365  0.0472992632 -0.0156507870 -0.0666115297 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0390419832  0.0246501572 -0.1196404089  0.0132020877 -0.0668824533 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.1317278300  0.0377762458 -0.0424849145  0.1181807313  0.0806208420 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0163805380 -0.0555319664 -0.0217307466 -0.0448972836  0.0186862567 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0983462390  0.0851451592 -0.0027253135  0.1081322107 -0.0364507093 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  0.1315667315  0.0210893024 -0.0383397028  0.1450646818 -0.0195608966 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0599670085  0.1217217115  0.0726746294 -0.0617502681  0.0334008573 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0356589185 -0.1679564009  0.1013096296 -0.0294736651 -0.1351465742 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0582287651  0.0222916529  0.0216665573 -0.0513437345 -0.0791790411 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0044188782  0.0904462959 -0.0809688620  0.0680128850 -0.1452127288 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0601983186  0.0260401681  0.0446657518 -0.0356500534  0.0117849404 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -0.1130333049  0.1110787808  0.0485890987 -0.0553934682 -0.1128152133 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0825777044 -0.1317609821  0.0909030537 -0.0152761664  0.0244443357 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0651058693 -0.0010145738  0.0717981526 -0.0321476938 -0.0129557241 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0273095318  0.0149841661  0.0463100676 -0.0356048221  0.0372470196 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0837422445  0.0001818662  0.0795124933  0.0333687543 -0.1619705398 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.0598611395 -0.0179516509 -0.0311220913 -0.0338724669 -0.0366123239 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0126543575 -0.0969028882 -0.1122234977 -0.0726887197  0.0635684397 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0172505995 -0.0359800997 -0.0620205246 -0.0848442527  0.0942150410 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -0.0824698106  0.0898246915  0.1240190960  0.0584103176  0.0609658269 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0121973047 -0.0331859925  0.2375195733 -0.1385197026 -0.1164814602 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0120976329  0.0579112298  0.0054338592 -0.0624920921  0.0014003626 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0956412207 -0.0338869778  0.0466121200 -0.0304211223 -0.0887625623 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.1088418401 -0.0022165178  0.0532409627  0.0436548138 -0.0177382786 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0626980740 -0.0449977924  0.0094474783  0.0713800159  0.0788185578 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0246667219  0.0503413602 -0.0788531218 -0.0734636629 -0.0512010721 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0900323950  0.0131901135  0.0938370276  0.0042556859 -0.0125666642 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -0.0252462972  0.0531969263  0.0601965248  0.0747727750 -0.0270632435 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0258737756 -0.0622690773 -0.0334729460  0.0620586610 -0.0364439928 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0013463589  0.0139887806 -0.0784396845 -0.0486813121  0.0098372581 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.1115928520  0.0639558503  0.0202148053 -0.0091175187  0.1443061395 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.0860071649  0.0060980772 -0.0929176194  0.0431049882  0.0460025367 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0538252689  0.0390893648  0.0160567736 -0.0902767949  0.0770743076 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0215862373  0.0737398871  0.0365792532 -0.0588637287 -0.0482258431 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0314178545 -0.0559681970 -0.0072848268 -0.0588450437 -0.0543305701 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0762985920  0.0085814780  0.0114710749 -0.0001789387 -0.0713759051 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0223132381 -0.0422863255 -0.0277848373 -0.0031765519  0.0338353416 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.1535593838 -0.0071088166  0.0166160561  0.0423445518 -0.0377593664 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0697188775  0.0914765418  0.0651981934 -0.0303306292  0.0612512064 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0002505189  0.0793808184 -0.0875202870  0.0715065561  0.0529015180 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0504042642 -0.1238035507  0.0318669886  0.0134353347  0.0899968462 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.1266732159  0.0469102037 -0.0757550051 -0.0527396954  0.0202176396 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0063924994 -0.1768248781 -0.0876298086  0.0306707513 -0.0026634835 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0325069153 -0.0944277059  0.0075452631  0.0124013663  0.0471253298 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.1297789830  0.0517159098 -0.0597044977  0.0242767231 -0.0011810094 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0656286777  0.0183405694 -0.0098783249  0.1478602022 -0.1041138467 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0197910922 -0.0551651983 -0.0105173781  0.0918703055  0.0127516702 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0746329665 -0.0012244077  0.1478127802  0.0307349076  0.0287300992 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0244614832  0.0600876366 -0.0165596868 -0.1152007117 -0.0386134845 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0328252001  0.1093335187  0.1423696994 -0.0430871948 -0.0571853236 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0006736002 -0.0410310168  0.0678953488  0.0933289030  0.0722148431 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0249207403  0.0167401182 -0.0092056072  0.0112243912  0.0305151132 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0790026054  0.0091032043 -0.0621303378  0.0359967308 -0.0355691454 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0061607430 -0.0937448971 -0.0761292938 -0.0791994499 -0.1127251902 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0646749416 -0.0065606724 -0.0992357215 -0.0062724025 -0.0386575424 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0140596530  0.0456014149 -0.0659189144  0.0196044591 -0.0788708591 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0606039974 -0.0955444799 -0.0224615618  0.0656758762  0.0790054937 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0342504120 -0.0059949509  0.1736569750 -0.2174037986 -0.0941671583 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0970039695 -0.1335101069 -0.0353890397 -0.0507464580 -0.0613009735 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0424185907  0.0096673141  0.1597413712  0.0104713273  0.1422363294 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.1694360785  0.0301634821  0.0432775331  0.0878015696  0.0469225558 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0045190909 -0.0559798132 -0.0770679253  0.0949792461 -0.0384632620 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0330449089  0.0167561958  0.0088140674  0.1388102355  0.0273855781 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0252433497 -0.1357401766 -0.0968495334  0.0236240722 -0.0206874171 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0391687160  0.1787311309 -0.0389811158  0.1591416579  0.0008833663 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0681685107  0.0556766726 -0.0232724315  0.0906280283  0.0009195202 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -0.0336045035  0.0390043359 -0.0315013412  0.0915641429  0.0154701774 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.0327891072 -0.0273360994  0.0339325084 -0.0428395156  0.0408075034 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0624501813  0.0003188818  0.1610672125  0.0173748445 -0.0513275528 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0343774551  0.0126745323 -0.0218065216 -0.0558966872  0.0005338360 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.1479429906  0.0891368901  0.0288699006 -0.0386891783 -0.0937987907 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0885826442 -0.0694669844 -0.0359748051  0.0358094483  0.0664065235 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0718134091 -0.0534460898 -0.0210069679  0.0428827843 -0.0511008900 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0267151163  0.0439891471  0.0803531393 -0.2031830083 -0.0104676598 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0385982445 -0.0677668660 -0.0135191156 -0.1053562753  0.1622056329 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0369740029 -0.0114065174 -0.0779594127 -0.1724519274  0.0719372030 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  0.1174009053 -0.0608626585  0.0848769056 -0.1407904023  0.1531171432 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0557982706 -0.0247311291 -0.0566488427 -0.0782882334  0.1471041243 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0808998851  0.0662308743  0.0698963719  0.0688443952 -0.0807108122 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.1384584263  0.0597207867 -0.0793506254 -0.0722281717 -0.0865803822 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  0.0624142457 -0.1176239113  0.0797640293  0.1172128074  0.0799935729 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0806260446 -0.1293338906  0.0535239549  0.0938728886 -0.0836860418 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0800862620 -0.1874429044 -0.0174565544  0.0393893870  0.0621938297 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0310012958 -0.0508237184 -0.1697107820  0.1040620665  0.0619087202 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0244091141  0.0275994050  0.0256027429  0.0097454826  0.0063657156 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0652709824 -0.0845045045  0.0942421487 -0.0322095452  0.1406017972 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0768623078 -0.0976302040 -0.0828238568  0.0839438461 -0.1003933232 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0683564472 -0.0476712143 -0.0597364227 -0.1083388826  0.0765742344 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.0012766285 -0.0111712553  0.0186380640