Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -9.448348e-02  1.087169e-01 -5.910698e-02  3.158911e-02  2.436466e-01 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -6.283266e-03 -2.611781e-02  2.581126e-03  3.532009e-02  7.468889e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  1.282357e-04 -2.263780e-02 -1.133222e-02 -1.220238e-03  3.723054e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -4.101948e-02  2.207914e-02  3.260286e-02 -2.673294e-02 -1.166101e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -2.256115e-02  4.394761e-02  2.830798e-02 -1.404082e-01 -2.451631e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -5.905415e-02  1.363156e-02 -5.119309e-02 -5.407504e-02  3.573693e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  8.085635e-04 -3.285082e-04  3.276505e-02  2.988069e-02 -6.257472e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -3.676364e-02 -2.083482e-02  1.405071e-02 -4.768104e-02 -2.131249e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -8.321060e-04 -1.157305e-02 -3.508979e-02 -4.134508e-02 -9.229433e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  5.011221e-02 -1.435552e-01 -7.755699e-02  7.427005e-02  8.318962e-03 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  7.299453e-02  9.949130e-02 -3.694119e-02 -3.243330e-02  4.438493e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -7.585308e-02  1.648423e-02 -7.769450e-02  7.347510e-02 -1.041482e-01 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -1.329386e-02  6.445398e-02 -3.409832e-04 -1.498217e-02 -1.766042e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -6.846370e-02  3.020045e-02  6.423372e-02 -4.161724e-02 -2.804870e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  1.604436e-02 -7.354056e-02  4.792932e-02  1.566391e-01 -9.635262e-03 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.034311e-02  5.474724e-02  5.577861e-02  9.564683e-03  7.741509e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -1.348871e-01  1.699020e-01 -6.701102e-02 -6.900531e-02  1.103719e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -2.381295e-02 -5.354944e-02 -4.070210e-02  1.035523e-01 -1.546472e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  2.328618e-03  1.141890e-01  2.447221e-02  2.280771e-01  2.742598e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  2.027024e-02 -6.119167e-03  1.749907e-02  3.340790e-02 -1.032631e-01 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  7.858530e-02 -4.330077e-02  1.091554e-01  6.390033e-02  3.834865e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  6.205920e-02  8.036372e-02 -3.281713e-02  8.649439e-02  9.598423e-05 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -7.588495e-04 -2.037797e-02 -6.980282e-03  1.709647e-02 -9.305797e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -1.117580e-01 -5.744528e-02 -1.787687e-02  9.911052e-02  2.743819e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -6.107046e-02 -3.681871e-02 -3.834198e-03  6.866735e-02 -3.004551e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  1.759050e-02  1.284933e-02  7.137527e-02  2.032546e-02 -3.616660e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  3.407190e-02 -1.881706e-02  1.901881e-02 -2.186723e-02  1.482967e-01 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  3.234882e-02  1.231634e-01 -8.793128e-03 -4.225929e-02 -5.128715e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  7.332758e-02 -4.157657e-02 -7.036842e-02  9.689568e-02  2.744033e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -1.441285e-01  6.671653e-02 -4.976346e-02 -1.090292e-02  2.531338e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  7.501819e-02 -8.946058e-02  3.765910e-02  1.159815e-02 -3.852322e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -2.626736e-02 -3.812650e-02  6.906992e-02 -1.143778e-02  2.273484e-03 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -5.060628e-02 -4.002380e-02  1.883589e-02  1.313864e-02  3.863885e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  5.445990e-02  2.492583e-02  6.578452e-02  7.417023e-03  4.693960e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  1.235800e-02  3.451412e-02  5.528156e-02  2.717885e-02  7.085354e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -2.712251e-02 -5.947435e-02 -1.692534e-01 -8.708628e-03 -4.022092e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -3.015218e-02  1.795472e-02  1.633116e-01 -8.240633e-02 -5.028579e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -8.354818e-03  8.629082e-02  8.859007e-02  1.223007e-01 -9.719447e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  9.800464e-02  3.209755e-02 -2.904627e-02 -1.517759e-01 -1.201700e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  5.552833e-02  1.076360e-01  5.011011e-02 -9.379068e-02 -1.643826e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -7.059117e-02  4.771997e-02 -4.388925e-02  3.668828e-02  2.718994e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  1.131434e-01 -8.760335e-03 -5.037717e-02  4.115342e-02 -9.726103e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  3.156330e-02  4.519560e-02  8.821274e-02 -8.680422e-05 -4.387591e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  4.915623e-02 -7.555366e-02 -5.095075e-02 -5.332784e-02  1.016867e-01 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  5.775395e-02  7.704192e-02  3.155754e-02  1.711709e-02 -1.904712e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -6.678616e-02 -3.940838e-02 -9.644501e-02  2.976433e-02  1.459561e-01 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -3.369364e-02 -4.699842e-02 -1.021811e-01 -1.007738e-01  5.571861e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -5.820642e-02  1.269710e-01 -2.456364e-02 -8.214698e-02  1.225468e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  1.170728e-01  1.060022e-01  1.664610e-02  1.901111e-02  7.158559e-03 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -7.407992e-02 -1.240132e-01 -1.799315e-02 -1.778065e-02  4.384460e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -9.991837e-04  1.052525e-01  5.220331e-02 -1.554366e-01 -3.257701e-03 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  3.377279e-02  4.920568e-02 -6.370161e-02  5.019172e-02  1.267322e-01 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  6.456995e-02  7.939027e-02 -2.828625e-02  1.543073e-02  2.080498e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  5.068405e-02  1.198432e-01 -8.798801e-02 -1.524379e-02  5.447850e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -1.726370e-01 -3.536813e-02  4.876049e-02  1.818017e-02  8.434376e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  1.532363e-01 -1.256191e-01 -5.884068e-03  3.588812e-02  9.053138e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -6.128806e-02 -6.971977e-03  1.966370e-02 -4.693216e-02 -3.645892e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  4.244962e-02  1.423050e-02  8.575004e-02  3.919756e-02 -1.703410e-03 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  5.817203e-03 -4.871657e-02  2.410101e-02  5.633915e-03 -3.509580e-03 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  5.073457e-02 -9.718685e-04  5.547175e-02 -1.183552e-02  2.556072e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  3.072959e-02 -7.409055e-02  7.638890e-03 -5.867955e-02 -4.540723e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -1.155324e-01  4.444538e-02  6.424512e-02  5.491521e-02 -3.051405e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  5.738222e-05  7.699929e-02  4.075117e-02  5.328105e-02  4.914675e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -3.027756e-02  8.323313e-02  8.826275e-02 -2.815340e-02  2.077761e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  8.113067e-02 -1.883064e-02  7.104985e-02 -1.073519e-02  5.995000e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  7.276227e-02 -1.848216e-01 -3.579897e-02  1.053147e-02  6.564842e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -7.678178e-03  6.537190e-02 -2.116757e-02  3.304395e-03 -5.696128e-03 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -1.563690e-02 -7.362994e-02  1.755269e-02  5.001224e-03 -5.463894e-03 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  1.386948e-01 -2.341832e-02  4.619779e-02 -8.011398e-02  9.925670e-04 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  4.978532e-02  1.048413e-01 -4.914553e-02  1.021103e-01 -1.558989e-03 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  5.393456e-02  1.823963e-01  3.573089e-02 -8.538371e-02  1.398364e-03 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -1.917072e-02 -1.005292e-01 -2.511710e-02  1.144835e-03  9.686957e-04 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -4.889350e-03 -3.437431e-02 -1.261922e-01 -1.135159e-02  5.776735e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -3.709146e-02  8.363164e-03  5.879749e-02 -1.741862e-02 -4.716188e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -1.389816e-01 -2.494891e-02  5.113222e-02 -7.793562e-02  1.980318e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  5.151450e-02  5.810371e-02 -6.544159e-02 -1.783214e-01 -4.855406e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  2.683066e-02 -8.222417e-02 -7.686221e-02  1.596719e-02  5.849891e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  2.147421e-02 -1.104477e-01  8.626146e-02  4.976816e-02 -8.680703e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -4.843571e-02 -9.048907e-03 -1.729769e-02 -7.650848e-02  2.794897e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -2.109869e-02 -4.955850e-02  7.960932e-02 -1.903786e-02  7.002937e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  2.248774e-02  7.774284e-03  7.420304e-04 -3.524737e-02 -9.606225e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  8.133573e-03  2.547461e-02 -3.700384e-02 -4.256978e-02 -8.965299e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  6.677900e-02 -9.007284e-02 -5.464008e-02 -1.060927e-01  6.067837e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -8.268382e-02 -1.480299e-02  7.948704e-03  7.829560e-02 -1.049975e-01 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  8.342696e-02 -6.394169e-02  1.273621e-01  1.439701e-01  7.313180e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -1.562353e-02 -6.597385e-02 -5.651646e-02 -4.181427e-02 -5.189133e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -3.312358e-02  2.135983e-02 -1.157778e-01 -3.475440e-02 -3.936745e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  1.908736e-02  6.222762e-02  1.877905e-02 -1.408987e-02 -4.158617e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -7.534369e-03 -1.970254e-02 -9.493076e-02 -4.954339e-02  4.924597e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -3.112244e-02 -9.131151e-04 -1.328974e-02