Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

acc(mat, lag)

Arguments

mat
A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag
The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)

#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -1.014626e-02 -4.689789e-02 -2.692515e-02  1.647534e-02  9.724370e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  7.484145e-02 -7.700089e-02  5.507884e-03  1.177610e-02  5.797200e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  2.777720e-02  7.371168e-02 -3.096055e-02 -6.863102e-02  5.809776e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  4.190925e-03 -1.620536e-01 -1.914391e-03  5.151140e-03  4.239143e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -3.490024e-02 -8.021212e-02 -4.018082e-02  6.264454e-02  2.969457e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -7.984651e-02  4.907431e-02 -1.382956e-02 -1.340997e-01 -4.668484e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  2.855372e-02 -2.130800e-03  8.552905e-02  9.787946e-03  7.878780e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  3.191921e-02 -2.942088e-02  4.191549e-02 -1.930977e-01  1.801244e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  8.467722e-02 -4.763093e-02 -9.402049e-02 -1.060703e-01  1.122288e-01 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -1.474603e-02  6.833816e-04 -2.074423e-02 -7.696818e-02 -3.507624e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -3.618370e-02  8.265259e-02 -1.182280e-02 -1.087773e-01 -4.273125e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -4.820032e-02  9.072186e-02 -4.034793e-02  3.855694e-02  3.272032e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  3.445096e-02 -2.929789e-02  1.288788e-01  4.403265e-02 -1.214084e-01 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -4.718246e-02  1.498830e-02 -3.510517e-03  2.588731e-02  3.759873e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  1.041074e-01  9.830565e-02 -2.995253e-02 -4.107438e-03 -8.757631e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -9.577378e-03  1.003160e-01  6.609862e-02 -4.093969e-02 -1.337836e-01 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -3.133033e-02 -1.163018e-01  8.296535e-02  2.601980e-02  4.565739e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  1.010374e-01 -3.934506e-02 -3.167190e-02 -4.123814e-02 -1.036662e-01 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  1.102415e-01  4.812668e-02 -1.014820e-01 -7.001576e-02 -3.997563e-04 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -7.309766e-02 -2.621820e-02  1.148827e-01  1.541341e-01  4.523372e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -1.500824e-01  2.905324e-02  5.775549e-02  6.885754e-02  3.235401e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -5.394662e-02  2.923125e-03 -1.001891e-01 -2.537464e-02 -2.821271e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  4.127258e-02 -4.412768e-02 -1.829914e-02 -2.753872e-02 -1.034009e-01 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  4.420047e-03  3.396719e-02  1.309730e-01  1.084799e-01  2.735404e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  8.781395e-02 -1.045412e-01 -9.428870e-02 -1.798644e-02 -3.550348e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  7.206837e-02 -4.099878e-03  1.096715e-03 -7.730774e-02  5.162370e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  1.367656e-02 -3.298740e-02  1.290353e-01  6.666055e-02 -3.461625e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  9.996531e-02 -8.396944e-02  6.112895e-02 -8.136800e-02 -6.971209e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -1.846996e-02 -4.372762e-02  8.690213e-02  4.804077e-02  9.540462e-04 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  2.965491e-02 -9.818870e-02 -2.112124e-03 -4.883648e-02  1.451089e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  8.361294e-02 -4.952688e-02  1.852868e-02 -5.045060e-02  7.113146e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  4.344614e-02 -2.519331e-03  4.097570e-02  7.662437e-02  3.948970e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -1.316233e-01 -2.391191e-02  5.628492e-02 -8.817427e-02 -3.587799e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -2.501454e-03  2.470557e-02 -5.544666e-02  2.189575e-02  2.256658e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -4.442616e-02 -1.587000e-02  1.730559e-02  1.373491e-02  9.869325e-03 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  1.082218e-01 -1.050241e-01 -7.049303e-02  7.566065e-02  1.593674e-03 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  1.589099e-02  7.996965e-02  2.445346e-02  8.723594e-02 -9.814877e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -1.907348e-02 -1.283348e-01 -1.051488e-01  4.423537e-02  6.285219e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  2.435132e-02 -5.038947e-02  9.147137e-02 -8.602451e-03 -9.754225e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  4.795032e-03  3.563519e-02  6.424525e-03 -1.372916e-01 -8.136544e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  2.392972e-04 -8.045395e-02 -5.663838e-02 -1.326506e-02 -2.952993e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  4.738841e-04 -7.433937e-02  8.054848e-02  3.764598e-02 -6.196252e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -3.336751e-02 -6.633212e-03 -7.027689e-02 -2.005246e-04 -4.974386e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  4.039876e-02 -7.930518e-02  5.055768e-02  1.667628e-01  1.295357e-01 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  5.501598e-02  2.527351e-02  1.634447e-01  5.991775e-02  6.803580e-03 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -8.177647e-02 -3.367828e-02 -5.736340e-02 -9.135188e-02 -4.004469e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -6.371581e-03  5.606253e-02  3.023385e-02  9.332307e-02  1.703150e-01 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  5.389973e-02  9.713901e-02  9.197559e-02  1.690463e-01 -6.529361e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -4.243294e-03  7.388237e-02 -1.365675e-01  1.577828e-02  5.854146e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  1.647634e-02  4.134614e-02  2.087563e-02  2.770891e-02 -2.279915e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -3.585366e-02  2.065098e-02  1.424844e-01  8.308389e-03 -4.552844e-03 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -4.353104e-05 -1.186247e-02 -3.916131e-02  1.712992e-02  1.021958e-01 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  1.142540e-01 -9.140194e-03 -6.561867e-02  9.564153e-02  7.441308e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -6.693345e-02 -2.296764e-02  2.599986e-02  2.201391e-02  2.193636e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  1.455573e-01  1.310553e-01 -1.036613e-01 -6.369097e-02 -5.003286e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  2.493700e-02  5.808592e-03 -1.288403e-02  2.246542e-02 -1.551220e-01 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -4.918705e-02  9.181147e-02 -1.645229e-04 -3.704220e-02 -7.805927e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  1.163643e-01  2.498178e-02 -1.763281e-02  4.081747e-02 -4.399147e-03 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  5.879293e-02 -3.103823e-03  2.093428e-02  7.314053e-02 -1.531432e-01 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -7.230911e-02  9.272714e-02 -6.402741e-02  1.625264e-02  2.102454e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -1.132282e-01 -9.877661e-02 -2.540298e-02 -2.729005e-02 -2.303502e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -2.667938e-02  4.561081e-02 -1.956544e-01 -1.331554e-02 -5.113536e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -3.080750e-02 -5.806834e-02  2.991227e-02 -3.239143e-02 -4.126889e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -6.308123e-02 -9.039233e-03 -7.812553e-03  1.487613e-01 -1.115144e-01 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  1.425205e-01  6.783791e-02  4.977117e-02 -8.917744e-03 -3.679629e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -1.766771e-03  8.767048e-02  4.206977e-02  1.384143e-02  1.738082e-03 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  2.177773e-02 -6.917411e-02  1.602737e-02 -4.822309e-02 -1.190594e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -6.194225e-02  2.470740e-02 -6.069955e-02  9.024074e-03  1.028006e-01 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  3.273739e-02 -9.077855e-02  1.574003e-02 -7.274030e-02 -2.341530e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  1.578845e-02 -7.166288e-02  6.383553e-02 -1.420652e-02  8.507359e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -3.052375e-02  1.231110e-01 -1.496741e-02  2.099733e-02 -1.121743e-01 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -4.276103e-03 -1.949380e-02 -7.622341e-02 -4.937604e-03  5.710194e-03 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  8.226911e-02  1.124395e-01  1.063482e-01 -1.907698e-02  8.616094e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  9.865814e-02 -2.312361e-02 -3.455099e-02 -3.096164e-02  4.662142e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  3.257820e-02  1.000918e-01  1.659531e-01 -4.281339e-03 -8.774312e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  4.154192e-02 -6.935180e-03  4.088713e-02  3.174958e-02 -2.481217e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  7.497311e-02  2.743034e-02  2.257679e-03  5.528845e-03  1.899622e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -2.146237e-02 -1.255438e-01  4.577646e-02  7.786086e-02  6.027946e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -4.806514e-02 -2.847502e-02  3.193948e-02 -3.720124e-03 -8.489616e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  1.868112e-02  9.167665e-02 -6.882979e-02  6.076987e-02  4.250670e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  5.911103e-02  5.108975e-02 -9.541006e-02 -4.186642e-02  2.501601e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -2.614642e-02  1.519533e-02  1.604408e-02 -7.161328e-02  1.021471e-01 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -1.187321e-01  7.877069e-02  6.309676e-02 -3.031132e-02 -4.266212e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -4.365925e-02  2.874007e-02  1.618527e-01  5.262294e-06  2.018572e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -1.235325e-02  3.068055e-02  1.191081e-01  1.373631e-02 -3.742718e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  5.564077e-02  1.399507e-02 -4.229821e-02  2.776238e-02 -6.197035e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -5.754315e-02  5.627585e-02  7.608026e-02  1.124065e-01  3.936790e-03 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -4.405565e-02  2.825133e-02 -6.906910e-02  9.554245e-03 -3.502412e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -1.007280e-01  3.196162e-02 -2.285670e-01  5.462874e-02 -3.553518e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  3.798382e-02  1.101128e-01  9.460069e-02