Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Note

Please see the references for details about auto cross covariance.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Author

Nan Xiao <https://nanx.me>

Examples

p <- 8 # p is the scales number
n <- 200 # n is the amino acid number
lag <- 7 # the lag paramter
mat <- matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0505765132  0.0587386522  0.0857473270 -0.0174697940 -0.0878160760 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0003876378  0.0227759058  0.0063646486 -0.0249146513 -0.0121988439 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0397287110  0.0998822835  0.0449380493 -0.0590044800 -0.0565522807 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.1026444946  0.1053617799 -0.0018691674 -0.0677001168  0.0579078827 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0455814528  0.1409144386  0.0619570559  0.0826018659 -0.0100917172 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.1170843957  0.0368273041 -0.0520337012 -0.0844844268 -0.0007323069 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0803304973 -0.0363464163  0.0720756235 -0.0364792062  0.1341650895 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.1142080062  0.0784990618  0.0395352188 -0.0233266706 -0.0612949952 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.1041405400 -0.0920258998  0.0062502703 -0.0444833670  0.0298529785 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0387600665  0.0864574616  0.0636809904 -0.0285088499  0.0781705977 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.1090344684  0.0741555302  0.0954444684  0.0268446970 -0.0181716764 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0928802485 -0.0564941908 -0.1593646915 -0.0067693805  0.0360670870 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0524238054 -0.0322750665  0.0800070013  0.0192907228  0.1226980239 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0177933700 -0.0590217738  0.0544297914 -0.1097546060 -0.1526782186 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0173853018  0.0833907116  0.0371701245  0.1411539294 -0.0840801667 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0774181304  0.0536929783  0.0144782389 -0.0406425790 -0.1007105914 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0780233194  0.0410798835 -0.0089898057  0.1126423681  0.0241819672 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.1195437804 -0.0167274115  0.0997210507 -0.0661462134  0.0424725990 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0825921848 -0.0408676184  0.0164368643 -0.0329267388  0.0422724414 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0023787782 -0.0260623643 -0.0346418927  0.1761543112  0.1087225372 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0540773090  0.0129210393 -0.0655752723 -0.0130374097 -0.2458309279 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0571819882  0.0070770385  0.0387923804 -0.0646049236  0.0833526307 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.1162900188  0.0585570061  0.0053972302 -0.0565445166 -0.0952982180 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.1447302572 -0.1305382722  0.1140290221 -0.0690373244 -0.1552126182 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0132219786  0.0172991520  0.0189170666 -0.0332114568 -0.0520671210 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0366146762 -0.0185218523  0.0653308116 -0.0804276082  0.0030663934 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0248613400  0.0695369958 -0.1526712065  0.0741044693  0.0217509643 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0302953682 -0.0397993375 -0.0314194319  0.0423291366 -0.0188984151 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0007963555  0.0176001433 -0.0175632485 -0.1290784319  0.0355730735 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0053026615  0.0764866243  0.1115845887 -0.0350964645 -0.0711303784 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.1102114627  0.1162814535 -0.1552385439  0.0343154864 -0.1450733364 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0256658486  0.0852944387 -0.1116085559  0.1897372185 -0.1222397331 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0464084240 -0.0013331308 -0.0592763878 -0.0008856232  0.0130064988 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0689867602 -0.0607907530 -0.0040515426 -0.0664861430  0.0150632834 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0036965366  0.0166985804  0.0002740181  0.0527051316  0.0464253919 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  0.0512562052  0.2676140159  0.1243255774 -0.0756995042 -0.1673616185 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0143522934  0.1505841183 -0.0773731720 -0.0319857686  0.0485811940 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.1233361737  0.1412080004  0.0543616284 -0.0887713602 -0.0327899780 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.1144670104 -0.0502850767  0.1092597992 -0.0846546750 -0.1100110042 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0553809816  0.0261692605 -0.0050013269 -0.0094810442 -0.1246365985 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0025486275  0.0512392267 -0.0591645295 -0.0803850014 -0.0435523829 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0103805835  0.0411966624  0.0374537599  0.0625382347  0.0518463524 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.1441567362 -0.0285658424  0.0160803885 -0.0710171161 -0.0035100470 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0903285777  0.0748771853 -0.1050565890  0.0089976629  0.1307499081 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0857623238  0.1144047198 -0.1273577110 -0.1733323446 -0.0573199824 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0820734489 -0.1024809868  0.0488442771  0.0304306207 -0.0010297361 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.1410616551  0.0075456831  0.0652947464  0.0516156625 -0.0588992751 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0593915086  0.0156850062 -0.0694976773 -0.0862886886 -0.0174024174 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0401711234 -0.0613726647 -0.0441926812 -0.0721634871 -0.0401162843 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0352209160  0.0845176779  0.1085371967 -0.0475444666 -0.1004456740 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0602905767 -0.1174203156 -0.0036877844  0.0106962935  0.1221269247 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.1140948791 -0.0443854664 -0.0574185109 -0.0373504097  0.0072744547 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.1194683959 -0.0384528347  0.0349349567  0.1776358310 -0.0212374746 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.1566096944  0.1039027962 -0.0636469746 -0.0567452697  0.0200269499 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0291960131  0.0635678468 -0.0451611602  0.0071534913 -0.0898365418 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.1145541871  0.0752016126 -0.0887964013  0.1148675167  0.0769073825 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0816501721 -0.1380668559 -0.0813455419 -0.0132909404 -0.0549163394 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0409248330 -0.0228369088 -0.0415503527 -0.0151279228 -0.0087515439 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  0.1926809859  0.0140073849  0.0179008320 -0.0068941397  0.0023968575 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0145306920 -0.1390844494  0.0645695443  0.0064798119  0.0276866770 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0653939284  0.0648485263  0.1065939994 -0.0567333672  0.0253283524 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  0.1503847606  0.0780723784 -0.0757880939  0.1092131872 -0.0569123553 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1060574995 -0.0716796245 -0.0679502886 -0.0015663726 -0.0119550872 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0662170070  0.0044311689  0.0394980931  0.0792263643  0.0337399590 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0075149380 -0.0677548009  0.0270051510 -0.0189271681  0.1046699459 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0930292553  0.0185990879 -0.0287343346  0.1823209057 -0.1328336606 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0999123382  0.0067815406  0.0344835428  0.0553890394  0.0521299683 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0037674770  0.0375511879  0.0393311770  0.1315997325  0.1009080572 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -0.1067990260 -0.0441313520 -0.0968810659 -0.0557593668  0.1780842360 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0214134167  0.0029093220  0.0007679367 -0.0340637581 -0.0117608947 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0407292131  0.0893921825  0.0264823453 -0.1428741195 -0.0458414755 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0583725541  0.0548752060 -0.0387283568 -0.0551658062 -0.1824230542 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0723702897  0.1217402716  0.0209035567 -0.0362770129 -0.1603407329 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0999347146 -0.0304714136 -0.0847339448  0.1955157943 -0.0117318434 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0508314779 -0.0663280593 -0.0513874388  0.0366558247  0.0471128050 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0162331131  0.0446770435 -0.0627783632  0.0064383038  0.0335611572 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0592392174  0.0080913498 -0.1239192090 -0.0170679444  0.1014091450 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  0.0611633250  0.0082799912  0.0574230048  0.0270070059  0.0836379560 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.0755897486 -0.0630063635  0.0979194034 -0.0948395717  0.0227880354 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.0064845434  0.0426177806  0.0575884129  0.0634619966 -0.1256178762 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0008571810 -0.0907579704 -0.1710647235  0.0461387542 -0.0502281881 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.1668813217  0.0119391705 -0.0501028928  0.0216537892  0.0204755406 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.1668046803 -0.0357251937 -0.0979807495  0.0413998288  0.0154119642 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0558263036 -0.0772419480 -0.0042253712  0.0381123805  0.0334948223 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0595422897  0.0257541284 -0.0492861155 -0.1163221969  0.1711559106 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0770605887  0.0179561006  0.0030744666 -0.1858188258 -0.2146717715 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0237391778  0.0574536349 -0.0243304762  0.0724149111 -0.1095198682 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.0887788849  0.0134617753  0.0497217555  0.0991814036 -0.1160077924 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.1209738836  0.0635901281 -0.1050133219  0.0227059624  0.0962806535 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0820434916 -0.0702929215  0.0291662343