CTD Descriptors - Composition

This function calculates the Composition descriptor of the CTD descriptors (dim: 21).

Usage

extractCTDC(x)

Arguments

x

A character vector, as the input protein sequence.

Value

A length 21 named vector

Note

For this descriptor type, users need to intelligently evaluate the underlying details of the descriptors provided, instead of using this function with their data blindly. It would be wise to use some negative and positive control comparisons where relevant to help guide interpretation of the results.

References

Inna Dubchak, Ilya Muchink, Stephen R. Holbrook and Sung-Hou Kim. Prediction of protein folding class using global description of amino acid sequence. Proceedings of the National Academy of Sciences. USA, 1995, 92, 8700-8704.

Inna Dubchak, Ilya Muchink, Christopher Mayor, Igor Dralyuk and Sung-Hou Kim. Recognition of a Protein Fold in the Context of the SCOP classification. Proteins: Structure, Function and Genetics, 1999, 35, 401-407.

See also

Author

Nan Xiao <https://nanx.me>

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractCTDC(x)
#>  hydrophobicity.Group1  hydrophobicity.Group2  hydrophobicity.Group3 
#>             0.29715302             0.40569395             0.29715302 
#> normwaalsvolume.Group1 normwaalsvolume.Group2 normwaalsvolume.Group3 
#>             0.45195730             0.29715302             0.25088968 
#>        polarity.Group1        polarity.Group2        polarity.Group3 
#>             0.33985765             0.33274021             0.32740214 
#>  polarizability.Group1  polarizability.Group2  polarizability.Group3 
#>             0.33096085             0.41814947             0.25088968 
#>          charge.Group1          charge.Group2          charge.Group3 
#>             0.11032028             0.79003559             0.09964413 
#> secondarystruct.Group1 secondarystruct.Group2 secondarystruct.Group3 
#>             0.38967972             0.29537367             0.31494662 
#>   solventaccess.Group1   solventaccess.Group2   solventaccess.Group3 
#>             0.43060498             0.29715302             0.27224199