Conjoint Triad Descriptor (with customized amino acid classification support)

This function calculates the Conjoint Triad descriptor, with customized amino acid classification support.

Usage

extractCTriadClass(x, aaclass)

Arguments

x

A character vector, as the input protein sequence.

aaclass

A list containing the customized amino acid classification. See example below.

Value

A length k^3 named vector, where k is the number of customized classes of the amino acids.

Note

For this descriptor type, users need to intelligently evaluate the underlying details of the descriptors provided, instead of using this function with their data blindly. It would be wise to use some negative and positive control comparisons where relevant to help guide interpretation of the results.

References

J.W. Shen, J. Zhang, X.M. Luo, W.L. Zhu, K.Q. Yu, K.X. Chen, Y.X. Li, H.L. Jiang. Predicting Protein-protein Interactions Based Only on Sequences Information. Proceedings of the National Academy of Sciences. 007, 104, 4337--4341.

Author

Nan Xiao <https://nanx.me>

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]

# use customized amino acid classification (normalized van der Waals volume)
newclass <- list(
  c("G", "A", "S", "T", "P", "D", "C"),
  c("N", "V", "E", "Q", "I", "L"),
  c("M", "H", "K", "F", "R", "Y", "W")
)

extractCTriadClass(x, aaclass = newclass)
#>      VS111      VS211      VS311      VS121      VS221      VS321      VS131 
#> 0.90384615 0.55769231 0.46153846 0.59615385 0.23076923 0.26923077 0.42307692 
#>      VS231      VS331      VS112      VS212      VS312      VS122      VS222 
#> 0.38461538 0.17307692 0.57692308 0.34615385 0.28846154 0.26923077 0.17307692 
#>      VS322      VS132      VS232      VS332      VS113      VS213      VS313 
#> 0.11538462 0.26923077 0.21153846 0.09615385 0.44230769 0.19230769 0.23076923 
#>      VS123      VS223      VS323      VS133      VS233      VS333 
#> 0.34615385 0.15384615 0.19230769 0.17307692 0.09615385 0.00000000