Scales-Based Descriptors derived by Multidimensional Scaling

This function calculates scales-based descriptors derived by Multidimensional Scaling (MDS). Users can provide customized amino acid property matrices.

Usage

extractMDSScales(x, propmat, k, lag, scale = TRUE, silent = TRUE)

Arguments

x

A character vector, as the input protein sequence.

propmat

A matrix containing the properties for the amino acids. Each row represent one amino acid type, each column represents one property. Note that the one-letter row names must be provided for we need them to seek the properties for each AA type.

k

Integer. The maximum dimension of the space which the data are to be represented in. Must be no greater than the number of AA properties provided.

lag

The lag parameter. Must be less than the amino acids.

scale

Logical. Should we auto-scale the property matrix (propmat) before doing MDS? Default is TRUE.

silent

Logical. Whether to print the k eigenvalues computed during the scaling process or not. Default is TRUE.

Value

A length lag * p^2 named vector, p is the number of scales (dimensionality) selected.

References

Venkatarajan, M. S., & Braun, W. (2001). New quantitative descriptors of amino acids based on multidimensional scaling of a large number of physical-chemical properties. Molecular modeling annual, 7(12), 445--453.

See also

Author

Nan Xiao <https://nanx.me>

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
data(AATopo)
tprops <- AATopo[, c(37:41, 43:47)] # select a set of topological descriptors
mds <- extractMDSScales(x, propmat = tprops, k = 5, lag = 7, silent = FALSE)
#> Eigenvalues computed during the scaling process: 
#>  [1]  1.266223e+02  5.846270e+01  4.058692e+00  6.994430e-01  1.529469e-01
#>  [6]  3.434787e-03  4.284842e-04  4.918500e-05  3.626185e-06  3.042632e-10
#> [11]  1.542441e-14  3.147970e-15  2.061313e-15  1.697196e-15  1.170661e-15
#> [16]  9.342587e-16  8.999803e-16  3.065404e-16  1.256159e-16 -4.025676e-15