Sequence-Order-Coupling Numbers

This function calculates the Sequence-Order-Coupling Numbers (dim: nlag * 2, default is 60).

extractSOCN(x, nlag = 30)

Arguments

x
A character vector, as the input protein sequence.

nlag
The maximum lag, defualt is 30.

Value

A length nlag * 2 named vector

References

Kuo-Chen Chou. Prediction of Protein Subcellar Locations by Incorporating Quasi-Sequence-Order Effect. Biochemical and Biophysical Research Communications, 2000, 278, 477-483.

Kuo-Chen Chou and Yu-Dong Cai. Prediction of Protein Sucellular Locations by GO-FunD-PseAA Predictor. Biochemical and Biophysical Research Communications, 2004, 320, 1236-1239.

Gisbert Schneider and Paul Wrede. The Rational Design of Amino Acid Sequences by Artifical Neural Networks and Simulated Molecular Evolution: Do Novo Design of an Idealized Leader Cleavge Site. Biophys Journal, 1994, 66, 335-344.

See also

Author

Nan Xiao <https://nanx.me>

Examples

x <- readFASTA(system.file("protseq/P00750.fasta", package = "protr"))[[1]]
extractSOCN(x)

#>  Schneider.lag1  Schneider.lag2  Schneider.lag3  Schneider.lag4  Schneider.lag5 
#>        204.2036        199.8708        206.8102        197.4828        193.3366 
#>  Schneider.lag6  Schneider.lag7  Schneider.lag8  Schneider.lag9 Schneider.lag10 
#>        208.1936        195.5476        200.9789        196.7110        193.9931 
#> Schneider.lag11 Schneider.lag12 Schneider.lag13 Schneider.lag14 Schneider.lag15 
#>        199.7031        204.9389        187.0140        198.4702        205.4526 
#> Schneider.lag16 Schneider.lag17 Schneider.lag18 Schneider.lag19 Schneider.lag20 
#>        193.1274        187.3529        190.4949        202.8853        198.5299 
#> Schneider.lag21 Schneider.lag22 Schneider.lag23 Schneider.lag24 Schneider.lag25 
#>        191.1013        185.0074        189.9857        202.7113        201.6267 
#> Schneider.lag26 Schneider.lag27 Schneider.lag28 Schneider.lag29 Schneider.lag30 
#>        194.5770        185.9939        204.1297        191.1629        183.9073 
#>   Grantham.lag1   Grantham.lag2   Grantham.lag3   Grantham.lag4   Grantham.lag5 
#>    6674686.0000    6761609.0000    7138892.0000    6748261.0000    6291229.0000 
#>   Grantham.lag6   Grantham.lag7   Grantham.lag8   Grantham.lag9  Grantham.lag10 
#>    6839853.0000    6594164.0000    6556148.0000    6620183.0000    6770614.0000 
#>  Grantham.lag11  Grantham.lag12  Grantham.lag13  Grantham.lag14  Grantham.lag15 
#>    6495689.0000    6865537.0000    6297267.0000    6498247.0000    6615566.0000 
#>  Grantham.lag16  Grantham.lag17  Grantham.lag18  Grantham.lag19  Grantham.lag20 
#>    6572680.0000    6569081.0000    6173947.0000    6570829.0000    6471308.0000 
#>  Grantham.lag21  Grantham.lag22  Grantham.lag23  Grantham.lag24  Grantham.lag25 
#>    6461649.0000    5939432.0000    6532121.0000    6652472.0000    6480660.0000 
#>  Grantham.lag26  Grantham.lag27  Grantham.lag28  Grantham.lag29  Grantham.lag30 
#>    6382281.0000    6276521.0000    6537634.0000    6442991.0000    6350157.0000